\name{as.character_method__BoundInFlux}
\alias{as.character_method__BoundInFlux}
\title{convert BoundInFlux Objects to something printable.}
\description{returns a string describing the object}
\arguments{
  \item{x}{An object}
  \item{\dots}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





