\name{as.character_method__TimeMap}
\alias{as.character_method__TimeMap}
\title{convert TimeMap Objects to something printable.}
\description{This method is needed to print a TimeMap object.}
\arguments{
  \item{x}{An Object of class time map}
  \item{\dots}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





