\name{BoundFc_method__data.frame_missing_missing_missing_character_missing}
\alias{BoundFc_method__data.frame_missing_missing_missing_character_missing}
\title{constructor }
\description{wrapper for \code{\link{BoundFc_method__data.frame_missing_missing_numeric_character_missing}} with the assumption lag=0 interpolation = splinefun}
\arguments{
  \item{map}{
}
  \item{format}{
}
  \item{interpolation}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





