\name{getFunctionDefinition_method__ConstLinDecompOp}
\alias{getFunctionDefinition_method__ConstLinDecompOp}
\title{creates a constant timedependent function and returns it}
\description{The method creates a timedependent function from the existing matrix describing the operator }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





