\name{getFunctionDefinition_method__TimeMap}
\alias{getFunctionDefinition_method__TimeMap}
\title{getFunctionDefinition method  TimeMap}
\description{extract the function definition (the R-function) }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





