\name{getAccumulatedRelease_method__Model}
\alias{getAccumulatedRelease_method__Model}
\title{time integrals of release fluxes per pool}
\description{The method integrates the release flux of every pool for all times in the  interval specified by the model definition.}
\arguments{
  \item{object}{
}
}

\value{a matrix}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





