\name{BoundFc_method__function_numeric_numeric_missing_character_missing}
\alias{BoundFc_method__function_numeric_numeric_missing_character_missing}
\title{constructor}
\description{wrapper for \code{\link{BoundFc_method__function_numeric_numeric_numeric_character_missing}} with the assumption lag=0}
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{format}{
}
}

\value{An object  that contains the interpolation function and the limits of the time range where the function is valid. Note that the limits change according to the time lag}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





