\name{getC14_method__Model_14}
\alias{getC14_method__Model_14}
\title{getC14 method  Model 14}
\description{This function computes the value for \eqn{^{14}C}{14C} (mass or concentration ) as function of time}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





