\name{fromDataFrame_method__TimeMap_data.frame}
\alias{fromDataFrame_method__TimeMap_data.frame}
\title{assemble the objects components}
\description{The method is used internally to convert TimeMap objects to BoundInFlux objects, since the use of TimeMap objects is now deprecated.}
\arguments{
  \item{.Object}{
}
  \item{map}{
}
  \item{lag}{
}
  \item{interpolation}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





