\name{getF14C}
\alias{getF14C}
\title{Calculates the average radiocarbon fraction weighted by the mass of carbon}
\description{This function calculates the average radiocarbon fraction weighted by the mass of carbon at each time step 
Have a look at the methods for details.}
\usage{getF14C(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





