\name{getF14C_method__Model_14}
\alias{getF14C_method__Model_14}
\title{read access to the models F14C variable }
\description{The model was created with a F14C object to describe the atmospheric 14C content. This method serves to investigate those settings from the model.}
\arguments{
  \item{object}{a Model_14 object}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





