\name{getF14R}
\alias{getF14R}
\title{Calculates the average radiocarbon fraction weighted by the amount of carbon release}
\description{This function calculates the average radiocarbon fraction weighted by the amount of carbon release at each time step.
Have a look at the methods for details.}
\usage{getF14R(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





