\name{getF14_method__Model_14}
\alias{getF14_method__Model_14}
\title{radiocarbon}
\description{Calculates the radiocarbon fraction for each pool at each time step.
}
\arguments{
  \item{object}{
}
}

\value{A matrix of dimension n x m; i.e. n time steps as rows and m pools as columns.}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





