\name{getFormat_method__BoundFc}
\alias{getFormat_method__BoundFc}
\title{extract the format string}
\description{the function just yields the format as a string}
\arguments{
  \item{object}{object  containing imformation about the format that could be Delta14C or AFM (Absolute Fraction Modern) for instance}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





