\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\title{getFunctionDefinition}
\description{Extracts the function definition (the R-function) from the argument
Have a look at the methods for details.}
\usage{getFunctionDefinition(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





