\name{getFunctionDefinition_method__BoundFc}
\alias{getFunctionDefinition_method__BoundFc}
\title{function definition }
\description{extract the function definition (the R-function) }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





