\name{getFunctionDefinition_method__BoundInFlux}
\alias{getFunctionDefinition_method__BoundInFlux}
\title{getFunctionDefinition method  BoundInFlux}
\description{extract the function definition (the R-function) from the BoundInFlux }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





