\name{getFunctionDefinition_method__BoundLinDecompOp}
\alias{getFunctionDefinition_method__BoundLinDecompOp}
\title{getFunctionDefinition method  BoundLinDecompOp}
\description{extract the function definition (the R-function) }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





