\name{getFunctionDefinition_method__DecompositionOperator}
\alias{getFunctionDefinition_method__DecompositionOperator}
\title{getFunctionDefinition method  DecompositionOperator}
\description{extract the function definition (the R-function) }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





