\name{getReleaseFlux14}
\alias{getReleaseFlux14}
\title{Calculates the mass of radiocarbon in the release flux (14C fraction times release flux)}
\description{This function computes the mass of radiocarbon in the release flux (14C fraction times release flux) as a function of time.
Have a look at the methods for details.}
\usage{getReleaseFlux14(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





