\name{BoundInFlux}
\alias{BoundInFlux}
\title{generic constructor}
\description{create a BoundInFlux object from different sources}
\usage{BoundInFlux(map, starttime, endtime, lag, interpolation)}\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
  \item{interpolation}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





