\name{getTimeRange_method__BoundInFlux}
\alias{getTimeRange_method__BoundInFlux}
\title{time domain of the function}
\description{The method returns a vector containing the start and end time where the intepolation is valid.}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





