\name{BoundInFlux_method__TimeMap_missing_missing_missing_missing}
\alias{BoundInFlux_method__TimeMap_missing_missing_missing_missing}
\title{convert to BoundInFlux}
\description{The method is used internally to convert TimeMap objects to BoundInFlux objects, since the use of TimeMap objects is now deprecated.}
\arguments{
  \item{map}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





