\name{getTimes_method__Model}
\alias{getTimes_method__Model}
\title{getTimes method  Model}
\description{This functions extracts the times argument from an argument of class Model}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





