\name{SoilR-package}
\alias{SoilR-package}
\alias{SoilR}
\docType{package}
\title{
\packageTitle{SoilR}
}
\description{
\packageDescription{SoilR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SoilR}
\packageIndices{SoilR}
}
\author{
\packageAuthor{SoilR}

Maintainer: \packageMaintainer{SoilR}
}
\keyword{ package }
