\name{initialize_method__BoundFc}
\alias{initialize_method__BoundFc}
\title{internal constructor (new)}
\description{The function is probably called internally only but used by all other constructors
It calls a sanity check on its arguments and initialized the object }
\arguments{
  \item{.Object}{
}
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
  \item{format}{
}
  \item{interpolation}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





