\name{initialize_method__ConstLinDecompOp}
\alias{initialize_method__ConstLinDecompOp}
\title{internal constructor }
\description{This mehtod is intendet to be used internally. It may change in the future.
For user code it is recommended to use one of the generic constructor \code{ConstLinDecompOp} instead.}
\arguments{
  \item{.Object}{
}
  \item{mat}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





