\name{initialize_method__DecompositionOperator}
\alias{initialize_method__DecompositionOperator}
\title{initialize called by (new)}
\description{This is the internal constructor for objects 
of the deprecated Class DecompositonOperator}
\arguments{
  \item{.Object}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{map}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





