\name{initialize_method__TimeMap}
\alias{initialize_method__TimeMap}
\title{initialize called by (new)}
\description{This is the standard constructor for objects of this class
It is called by statements of the form 
\code{new("TimeMap",start,end,f,lag)}}
\arguments{
  \item{.Object}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{map}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





