\name{BoundInFlux_method__function_numeric_numeric_missing_missing}
\alias{BoundInFlux_method__function_numeric_numeric_missing_missing}
\title{constructor }
\description{the method constructs an object from its basic ingredients}
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





