\name{[_method__Model_character}
\alias{[_method__Model_character}
\title{(experimenta) partially overload [ ] for models }
\description{This method overloads the [] operator for Model objects but is not yet finished so the full interface for [] is not yet implemented
and the behavior is very likely to change in future versions of SoilR}
\arguments{
  \item{x}{
}
  \item{i}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





