\name{BoundFc}
\alias{BoundFc}
\title{generic constructor}
\description{create a BoundFc object from different sources}
\usage{BoundFc(map, starttime, endtime, lag, format, interpolation)}\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
  \item{format}{
}
  \item{interpolation}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





