\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\title{Generic constructor}
\description{Creates a LinearDecompositonOperator from different sources.
Please look at the methods to see what kind of input is supported. }
\usage{BoundLinDecompOp(map, starttime, endtime, lag)}\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





