\name{BoundLinDecompOp_method__ConstLinDecompOp}
\alias{BoundLinDecompOp_method__ConstLinDecompOp}
\title{convert a ConstLinDecompOp to a BoundLinDecompOp}
\description{The method creates a BoundLinDecompOp consisting of a constant time dependent function 
and the limits of its domain (starttime and endtime) set to -Inf and Inf respectively}
\arguments{
  \item{map}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





