\name{BoundLinDecompOp_method__TimeMap_missing_missing_missing}
\alias{BoundLinDecompOp_method__TimeMap_missing_missing_missing}
\title{create a BoundLinDecompOp from a TimeMap}
\description{The method is used internally to convert TimeMap objects to BoundLinDecompOp where the use of TimeMap is now deprecated.}
\arguments{
  \item{map}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





