\name{BoundLinDecompOp_method__function_numeric_numeric_missing}
\alias{BoundLinDecompOp_method__function_numeric_numeric_missing}
\title{a constructor }
\description{This method creates a BoundLinDecompOp from a timedependent function and its domain }
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





