\name{BoundLinDecompOp_method__function_numeric_numeric_numeric}
\alias{BoundLinDecompOp_method__function_numeric_numeric_numeric}
\title{a constructor }
\description{This method creates a BoundLinDecompOp from a timedependent function and its domain }
\arguments{
  \item{map}{a function}
  \item{starttime}{the begin of the time domain}
  \item{endtime}{the end of the time domain}
  \item{lag}{lag time}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





