\name{BoundLinDecompOp-class}
\Rdversion{1.1}
\docType{class}
\alias{BoundLinDecompOp-class}
\title{Class \code{"BoundLinDecompOp"}}
\description{
NA
}
\section{Slots}{
  \describe{
    \item{\code{map}:}{Object of class \code{"function"} ~~ }
    \item{\code{starttime}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{endtime}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{lag}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{DecompOp}"}, directly.
Class \code{"\linkS4class{TimeMap}"}, directly.
}
\section{Methods}{
No methods defined with class "BoundLinDecompOp" in the signature.
}
\author{
 Carlos A. Sierra, Markus Mueller 
}
\examples{
showClass("BoundLinDecompOp")
}
\keyword{classes}
