\name{BoundLinDecompOp_method__data.frame_missing_missing}
\alias{BoundLinDecompOp_method__data.frame_missing_missing}
\title{a constructor }
\description{This method creates a BoundLinDecompOp from a dataframe}
\arguments{
  \item{map}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





