\name{ConstLinDecompOp_method__matrix}
\alias{ConstLinDecompOp_method__matrix}
\title{construct from matric}
\description{This method creates a ConstLinDecompOp from a matrix
The operator is assumed to act on the vector of carbon stocks
by multiplication of the (time invariant) matrix from the left.}
\arguments{
  \item{mat}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





