\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\title{Generic constructor}
\description{Creates a LinearDecompositonOperator from different sources.
Please look at the methods to see what kind of input is supported. }
\usage{BoundLinDecompOp(map, starttime, endtime)}\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





