\name{AbsoluteFractionModern_from_Delta14C_method__matrix}
\alias{AbsoluteFractionModern_from_Delta14C_method__matrix}
\title{Converts from Delta14C to Absolute Fraction Modern}
\description{This method produces a matrix of Delta14C values from a Matrix or number of  Absolute Fraction Modern}
\arguments{
  \item{delta14C}{An object of class matrix containing the values in Delta14C format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





