\name{DecompOp_method__TimeMap}
\alias{DecompOp_method__TimeMap}
\title{creates a BoundLinDecompOp from a TimeMap object}
\description{The resulting operator is creted by a call to the constructor of class
BoundLinDecompOp
The method is used to ensure backward compatibility with the now deprecated
TimeMap class}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





