\name{DecompOp_method__matrix}
\alias{DecompOp_method__matrix}
\title{creates a ConstanDecompOp from a matrix}
\description{The resulting operator is creted by a call to the constructor of class
ConstLinDecompOp}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





