\name{Delta14C_from_AbsoluteFractionModern_method__matrix}
\alias{Delta14C_from_AbsoluteFractionModern_method__matrix}
\title{Converts Absolute Fraction Modern values to Delta14C}
\description{This method produces a matrix of  Delta14C values from a matrix of values in Absolute Fraction Modern.}
\arguments{
  \item{AbsoluteFractionModern}{An object of class matrix containing the values in Absolute Fraction Modern format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





