\name{ConstLinDecompOp}
\alias{ConstLinDecompOp}
\title{Generic constructor}
\description{Creates a ConstantDecompositonOperator object from different sources.
Please look at the different methods to see what kind of input is supported. }
\usage{ConstLinDecompOp(mat)}\arguments{
  \item{mat}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





