\name{Delta14C_method__BoundFc}
\alias{Delta14C_method__BoundFc}
\title{convert to Absolute Fraction Normal values  }
\description{convert object containing values in any supported format to the appropriate Absolute Fraction Modern values.}
\arguments{
  \item{F}{object of containing the values in any format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





