\name{AbsoluteFractionModern_from_Delta14C_method__numeric}
\alias{AbsoluteFractionModern_from_Delta14C_method__numeric}
\title{Converts from Delta14C to Absolute Fraction Modern}
\description{Converts a number or vector containing Delta14C values to the appropriate Absolute Fraction Modern values.
Have a look at the methods for details.}
\arguments{
  \item{delta14C}{A numeric object containing the values in Delta14C format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





