\name{GeneralModel}
\alias{GeneralModel}
\title{A general constructor }
\description{Creates a Model object from different sources
Have a look at the methods for details.}
\usage{GeneralModel(t, A, ivList, inputFluxes, ...)}\arguments{
  \item{t}{
}
  \item{A}{
}
  \item{ivList}{
}
  \item{inputFluxes}{
}
  \item{\dots}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





