\name{GeneralModel_14_method__numeric_ANY_numeric_ANY_ANY_ANY_missing_numeric_missing_function_logical}
\alias{GeneralModel_14_method__numeric_ANY_numeric_ANY_ANY_ANY_missing_numeric_missing_function_logical}
\title{a constructor for class Model_14}
\description{This method tries to create a Model object from any combination of arguments 
that can be converted into  the required set of building blocks for a model
for n arbitrarily connected pools.}
\arguments{
  \item{t}{A vector containing the points in time where the solution is sought.}
  \item{A}{
}
  \item{ivList}{
}
  \item{initialValF}{
}
  \item{inputFluxes}{
}
  \item{inputFc}{
}
  \item{di}{
}
  \item{solverfunc}{
}
  \item{pass}{
}
}

\value{A model object that can be further queried. }

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}



\seealso{\code{\link{TwopParallelModel}}, \code{\link{TwopSeriesModel}}, \code{\link{TwopFeedbackModel}} }

