\name{GeneralModel_14_method__numeric_ANY_numeric_ANY_ANY_ANY_missing_numeric_missing_missing_logical}
\alias{GeneralModel_14_method__numeric_ANY_numeric_ANY_ANY_ANY_missing_numeric_missing_missing_logical}
\title{a constructor for class Model_ 14}
\description{A wrapper for \code{\link{Model_14}} with pass=FALSE}
\arguments{
  \item{t}{
}
  \item{A}{
}
  \item{ivList}{
}
  \item{initialValF}{
}
  \item{inputFluxes}{
}
  \item{inputFc}{
}
  \item{di}{
}
  \item{pass}{
}
}

\value{A model object that can be further queried. }

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}



\seealso{\code{\link{TwopParallelModel}}, \code{\link{TwopSeriesModel}}, \code{\link{TwopFeedbackModel}} }

