\name{AbsoluteFractionModern_method__BoundFc}
\alias{AbsoluteFractionModern_method__BoundFc}
\title{convert to Absolute Fraction Normal values  }
\description{convert a BoundFc object containing values in any supported format to the appropriate Absolute Fraction Modern values.}
\arguments{
  \item{F}{object containing the values in any format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





