\name{InFlux-class}
\Rdversion{1.1}
\docType{class}
\alias{InFlux-class}
\title{Class \code{"InFlux"}}
\description{
NA
}
\section{Methods}{
  \describe{
    \item{InFlux}{\code{signature(object = "InFlux")}: ... }
	 }
}
\author{
 Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de> 
}
\examples{
showClass("InFlux")
}
\keyword{classes}
