\name{InFlux_method__TimeMap}
\alias{InFlux_method__TimeMap}
\title{create a BoundInFlux from a TimeMap object}
\description{The method is used to ensure backward compatibility with the now deprecated
TimeMap class.
The resulting BoundInFlux is created by a call to
the constructor BoundInFlux(object) of that class.}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





