\name{Model_14}
\alias{Model_14}
\title{A general constructor }
\description{Creates a Model_14 object from different sources
Have a look at the methods for details.}
\usage{Model_14(t, A, ivList, initialValF, inputFluxes, inputFc, c14DecayRate, 
    solverfunc, pass)}\arguments{
  \item{t}{
}
  \item{A}{
}
  \item{ivList}{
}
  \item{initialValF}{
}
  \item{inputFluxes}{
}
  \item{inputFc}{
}
  \item{c14DecayRate}{
}
  \item{solverfunc}{
}
  \item{pass}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





