\name{SoilR-package}
\alias{SoilR-package}
\alias{SoilR}
\docType{package}
\title{Models of Soil Organic Matter Decomposition}
\description{This package contains functions for modeling Soil Organic
Matter decomposition in terrestrial ecosystems.}
\details{
\tabular{ll}{Package: \tab SoilR\cr
Title: \tab Models of Soil Organic Matter Decomposition\cr
Version: \tab 1.1-22\cr
Date: \tab 2014-04-26\cr
Author: \tab Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>\cr
Maintainer: \tab Markus Mueller <mamueller@bgc-jena.mpg.de>\cr
License: \tab GPL-3\cr
Depends: \tab deSolve,methods,parallel,RUnit\cr
Suggests: \tab FME,lattice,MASS\cr
LazyData: \tab TRUE\cr}
}
\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}

\keyword{ package }


