% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstFc.R
\name{AbsoluteFractionModern,ConstFc-method}
\alias{AbsoluteFractionModern,ConstFc-method}
\title{automatic title}
\usage{
\S4method{AbsoluteFractionModern}{ConstFc}(F)
}
\arguments{
\item{F}{object of class:\code{ConstFc}, : no manual documentation}
}
\description{
automatic title
}
