% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AbsoluteFractionModern_from_Delta14C,matrix-method}
\alias{AbsoluteFractionModern_from_Delta14C,matrix-method}
\title{Automatic description: AbsoluteFractionModern_from_Delta14C,matrix-method}
\usage{
\S4method{AbsoluteFractionModern_from_Delta14C}{matrix}(delta14C)
}
\arguments{
\item{delta14C}{object of class:\code{matrix}. : object of class:\code{matrix}, no manual documentation}
}
\description{
Automatic description: AbsoluteFractionModern_from_Delta14C,matrix-method
}
