\name{AbsoluteFractionModern_from_Delta14C_method__matrix}
\alias{AbsoluteFractionModern_from_Delta14C_method__matrix}
\title{convert from Delta14C to Absolute Fraction Normal values  }
\description{This method produces a matrix of Delta14C values from  a Matrix of  Absolute Fraction Normal values.}
\arguments{
  \item{delta14C}{a Matrix containing the values in Delta14C format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





