\name{AbsoluteFractionModern_from_Delta14C_method__numeric}
\alias{AbsoluteFractionModern_from_Delta14C_method__numeric}
\title{convert from Delta14C to Absolute Fraction Normal values  }
\description{convert a number or vector containing Delta14C values to the appropriate Absolute Fraction Modern values .}
\arguments{
  \item{delta14C}{numeric containing the values in Delta14C format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





