\docType{methods}
\name{add_plot,TimeMap-method}
\title{add_plot,TimeMap-method 
 overview plot}
\alias{add_plot,TimeMap-method}
\usage{\S4method{add_plot}{TimeMap}(x,
...)}
\arguments{
\item{x}{: of class TimeMap, the TimeMap Object to be plotted}
\item{...}{:  other plot parameters not implemented yet}
}
\description{The method adds a simple overview plot of a the scalar,
vector, matrix or arrayvalued function plotting all time dependent component.}
