% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{add_plot,TimeMap-method}
\alias{add_plot,TimeMap-method}
\title{add_plot,TimeMap-method overview plot}
\arguments{
\item{x}{: of class TimeMap, the TimeMap Object to be plotted}

\item{...}{: other plot parameters not implemented yet}
}
\description{
The method adds a simple overview plot of a the scalar, vector, matrix or
arrayvalued function plotting all time dependent component.
}
