% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{add_plot}
\alias{add_plot}
\title{add_plot}
\usage{
add_plot(x, ...)
}
\arguments{
\item{x}{see method arguments}

\item{...}{see method arguments}
}
\description{
add_plot
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{add_plot,TimeMap-method}}
  \item \code{\link{add_plot,TimeMap-method}}
}
}

