% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_1_phosphate.R
\name{agr_1_phosphate}
\alias{agr_1_phosphate}
\title{Recommendation of phosphate application in KY, USA.}
\usage{
agr_1_phosphate(p, crop)
}
\arguments{
\item{p}{Phosphorous soil test result.}

\item{crop}{\code{corn} or \code{soybean}}
}
\value{
Amount of P2O5 (lb/ac) to apply.
}
\description{
Based on the reference values of AGR1 guidelines issued by the University
of Kentucky, phosphate fertilizer rates can be calculated for corn and
soybean.
}
\examples{
agr_1_phosphate(p = 6, crop = "corn")

agr_1_phosphate(p = 6, crop = "soybean")
}
