% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr_1_potash.R
\name{agr_1_potash}
\alias{agr_1_potash}
\title{Recommendation of potash application in KY, USA.}
\usage{
agr_1_potash(k, crop)
}
\arguments{
\item{k}{Potash soil test result.}

\item{crop}{\code{corn} or \code{soybean}}
}
\value{
Amount of K2O (lb/ac) to apply.
}
\description{
Based on the reference values of AGR1 guidelines issued by the University
of Kentucky, potash fertilizer rates can be calculated for corn and
soybean.
}
\examples{
agr_1_potash(k = 20, crop = "corn")

agr_1_potash(k = 20, crop = "soybean")
}
