\docType{methods}
\name{as.character,TimeMap-method}
\title{as.character,TimeMap-method}
\alias{as.character,TimeMap-method}
\usage{\S4method{as.character}{TimeMap}(x,
...)}
\arguments{
\item{x}{: of class TimeMap, An Object of class time map}
\item{...}{:  will be ignored}
}
\description{This method is needed to print a TimeMap object.}
