\name{as.character}
\alias{as.character}
\title{creates a character representation of the object in question}
\description{This function computes the carbon content of the pools as function of time}
\usage{as.character(object)}
\arguments{
  \item{object}{The object to be printed.}
}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}



