% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFlux_by_PoolName.R
\name{as.numeric,InternalFlux_by_PoolName-method}
\alias{as.numeric,InternalFlux_by_PoolName-method}
\title{Convert to a numeric value with name of the form 'a->b'}
\usage{
\S4method{as.numeric}{InternalFlux_by_PoolName}(x, y, t, time_symbol, ...)
}
\arguments{
\item{x}{object of class:\code{InternalFlux_by_PoolName}, The list of fluxes. Every element 
contains a function that depends on a combination of
of state variables and time.}

\item{y}{A vector indexed by the names of the state 
variables}

\item{t}{a number representing the current point in time}

\item{time_symbol}{The name of the time argument used 
in the definition of the flux functions}

\item{...}{no manual documentation}
}
\description{
Convert to a numeric value with name of the form 'a->b'
}
