% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{availableParticleProperties}
\alias{availableParticleProperties}
\title{S4 generic: availableParticleProperties}
\usage{
availableParticleProperties(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: availableParticleProperties
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{availableParticleProperties,MCSim-method}}
}
}

