% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{availableResidentSets,MCSim-method}
\alias{availableResidentSets,MCSim-method}
\title{Automatic description: availableResidentSets,MCSim-method}
\usage{
\S4method{availableResidentSets}{MCSim}(object)
}
\arguments{
\item{object}{object of class:\code{MCSim}. : object of class:\code{MCSim}, no manual documentation}
}
\description{
Automatic description: availableResidentSets,MCSim-method
}
