% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{BoundFc,character-method}
\alias{BoundFc,character-method}
\title{BoundFc,character-method A constructor}
\arguments{
\item{format}{: of class character, a string that specifies the format used
to represent the atmospheric fraction. Possible values are "Delta14C" which
is the default or "afn" the Absolute Fraction Normal representation}

\item{...}{: passed on to TimeMap}
}
\description{
The method constructs an object from a format string and other parameters
which are passed on to \code{\link{TimeMap}}
}
