% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{BoundFc-class}
\alias{BoundFc-class}
\title{S4-class to represent atmospheric 14C concentration as
scalar function of time.}
\description{
As time dependent scalar function which remembers its
domain ( see \code{\linkS4class{ScalarTimeMap}}) and its
format.
}
\section{S4-superclasses (in the package)}{
      \itemize{
        \item \code{\linkS4class{ScalarTimeMap}}
        \item \code{\linkS4class{Fc}}
        \item \code{\linkS4class{TimeMap}}
      }
}
