\docType{methods}
\name{BoundFc,data.frame,missing,missing,missing,character,missing-method}
\title{BoundFc,data.frame,missing,missing,missing,character,missing-method}
\alias{BoundFc,data.frame,missing,missing,missing,character,missing-method}
\usage{\S4method{BoundFc}{data.frame,missing,missing,missing,character,missing}(map,
format,
interpolation)}
\arguments{
\item{map}{: of class data.frame}
\item{format}{: of class character}
\item{interpolation}{: of class missing}
}
\description{wrapper with the assumption lag=0 interpolation = splinefun
Start- and end time are determined from the data.frame }
