\docType{methods}
\name{BoundFc,data.frame,missing,missing,missing,character,function-method}
\title{BoundFc,data.frame,missing,missing,missing,character,function-method}
\alias{BoundFc,data.frame,missing,missing,missing,character,function-method}
\usage{\S4method{BoundFc}{data.frame,missing,missing,missing,character,function}(map,
format,
interpolation)}
\arguments{
\item{map}{: of class data.frame}
\item{format}{: of class character}
\item{interpolation}{: of class function}
}
\description{wrapper  with the assumption lag=0}
