\docType{methods}
\name{BoundFc,data.frame,missing,missing,numeric,character,missing-method}
\title{BoundFc,data.frame,missing,missing,numeric,character,missing-method}
\alias{BoundFc,data.frame,missing,missing,numeric,character,missing-method}
\usage{\S4method{BoundFc}{data.frame,missing,missing,numeric,character,missing}(map,
lag,
format)}
\arguments{
\item{map}{: of class data.frame}
\item{lag}{: of class numeric}
\item{format}{: of class character}
}
\description{wrapper  with the assumption interpolation =splinefun}
