\docType{methods}
\name{BoundFc,missing-method}
\title{BoundFc,missing-method 
 A constructor}
\alias{BoundFc,missing-method}
\usage{\S4method{BoundFc}{missing}(format,
...)}
\arguments{
\item{format}{: of class missing, If this method is called the format argument
was not present.}
\item{...}{:  passed on to TimeMap}
}
\description{The method constructs an object from a list
which must contain a }
