\description{no Description}
\name{BoundFc}
\alias{BoundFc}
\usage{BoundFc(map,
starttime,
endtime,
lag,
format,
interpolation)}
\title{BoundFc S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{format}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundFc,ANY,ANY,ANY-method}}\cr
	\code{\link{BoundFc,ANY,missing,missing-method}}\cr}
