\title{BoundInFlux S4 class}
\description{defines a time dependent inputrate as function of time and including the domain where the function is well defined. This can be used to avoid interpolations out of range when mixing different time dependent data sets
defines a time dependent inputrate as function of time and including the domain where the function is well defined. This can be used to avoid interpolations out of range when mixing different time dependent data sets}
\name{BoundInFlux-class}
\alias{BoundInFlux-class}
\docType{class}
\section{Methods}{No methods defined with class BoundInFlux in the signature.}
