\docType{methods}
\name{BoundInFlux,ANY,missing,missing-method}
\title{BoundInFlux,ANY,missing,missing-method}
\alias{BoundInFlux,ANY,missing,missing-method}
\usage{\S4method{BoundInFlux}{ANY,missing,missing}(map,
lag=0,
interpolation)}
\arguments{
\item{map}{: of class ANY}
\item{lag}{}
\item{interpolation}{}
}
\description{The method internally calls \code{\link{TimeMap}} and expects
the same kind of map argument}
