\docType{methods}
\name{BoundInFlux,TimeMap,missing,missing,missing,missing-method}
\title{BoundInFlux,TimeMap,missing,missing,missing,missing-method}
\alias{BoundInFlux,TimeMap,missing,missing,missing,missing-method}
\usage{\S4method{BoundInFlux}{TimeMap,missing,missing,missing,missing}(map)}
\arguments{
\item{map}{: of class TimeMap}
}
\description{The method is used internally to convert TimeMap objects to BoundInFlux objects, since the use of TimeMap objects is now deprecated.}
