\description{no Description}
\name{BoundInFlux}
\alias{BoundInFlux}
\usage{BoundInFlux(map,
starttime,
endtime,
lag,
interpolation)}
\title{BoundInFlux S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundInFlux,ANY,ANY,ANY-method}}\cr
	\code{\link{BoundInFlux,ANY,missing,missing-method}}\cr}
