\docType{methods}
\name{BoundLinDecompOp,UnBoundLinDecompOp,ANY,ANY-method}
\title{BoundLinDecompOp,UnBoundLinDecompOp,ANY,ANY-method}
\alias{BoundLinDecompOp,UnBoundLinDecompOp,ANY,ANY-method}
\usage{\S4method{BoundLinDecompOp}{UnBoundLinDecompOp,ANY,ANY}(map,
starttime=-Inf,
endtime=Inf)}
\arguments{
\item{map}{: of class UnBoundLinDecompOp}
\item{starttime}{: of class ANY}
\item{endtime}{: of class ANY}
}
\description{The method creates a BoundLinDecompOp consisting of a constant time dependent function 
and the limits of its domain (starttime and endtime) set to -Inf and Inf respectively}
