\docType{methods}
\name{BoundLinDecompOp,data.frame,missing,missing-method}
\title{BoundLinDecompOp,data.frame,missing,missing-method}
\alias{BoundLinDecompOp,data.frame,missing,missing-method}
\usage{\S4method{BoundLinDecompOp}{data.frame,missing,missing}(map)}
\arguments{
\item{map}{: of class data.frame}
}
\description{This method creates a BoundLinDecompOp from a dataframe}
