% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundLinDecompOp.R
\name{BoundLinDecompOp,ANY-method}
\alias{BoundLinDecompOp,ANY-method}
\title{automatic title}
\usage{
\S4method{BoundLinDecompOp}{ANY}(map, ...)
}
\arguments{
\item{map}{: no manual documentation}

\item{...}{: no manual documentation}
}
\description{
automatic title
}
