\docType{methods}
\name{BoundLinDecompOp,function,numeric,numeric-method}
\title{BoundLinDecompOp,function,numeric,numeric-method}
\alias{BoundLinDecompOp,function,numeric,numeric-method}
\usage{\S4method{BoundLinDecompOp}{function,numeric,numeric}(map,
starttime,
endtime)}
\arguments{
\item{map}{: of class function, a function}
\item{starttime}{: of class numeric, the begin of the time domain}
\item{endtime}{: of class numeric, the end of the time domain}
}
\description{This method creates a BoundLinDecompOp from a timedependent function and its domain }
