\docType{methods}
\name{BoundLinDecompOp,TimeMap,missing,missing-method}
\title{BoundLinDecompOp,TimeMap,missing,missing-method}
\alias{BoundLinDecompOp,TimeMap,missing,missing-method}
\usage{\S4method{BoundLinDecompOp}{TimeMap,missing,missing}(map)}
\arguments{
\item{map}{: of class TimeMap}
}
\description{The method is used internally to convert TimeMap objects to BoundLinDecompOp where the use of TimeMap is now deprecated.}
