\docType{methods}
\name{BoundLinDecompOp,ConstLinDecompOp,missing,missing-method}
\title{BoundLinDecompOp,ConstLinDecompOp,missing,missing-method}
\alias{BoundLinDecompOp,ConstLinDecompOp,missing,missing-method}
\usage{\S4method{BoundLinDecompOp}{ConstLinDecompOp,missing,missing}(map)}
\arguments{
\item{map}{: of class ConstLinDecompOp}
}
\description{The method creates a BoundLinDecompOp consisting of a constant time dependent function 
and the limits of its domain (starttime and endtime) set to -Inf and Inf respectively}
