\docType{methods}
\name{BoundLinDecompOp,UnBoundLinDecompOp,missing,missing-method}
\title{BoundLinDecompOp,UnBoundLinDecompOp,missing,missing-method}
\alias{BoundLinDecompOp,UnBoundLinDecompOp,missing,missing-method}
\usage{\S4method{BoundLinDecompOp}{UnBoundLinDecompOp,missing,missing}(map)}
\arguments{
\item{map}{: of class UnBoundLinDecompOp}
}
\description{The method creates a BoundLinDecompOp consisting of a constant time dependent function 
and the limits of its domain (starttime and endtime) set to -Inf and Inf respectively}
