\docType{methods}
\name{BoundLinDecompOp,ANY-method}
\title{BoundLinDecompOp,ANY-method 
 a constructor }
\alias{BoundLinDecompOp,ANY-method}
\usage{\S4method{BoundLinDecompOp}{ANY}(map,
...)}
\arguments{
\item{map}{: of class ANY, anything that can be used as a single \code{map} parameter}
\item{...}{:  passed on to TimeMpa}
}
\description{Creates a BoundLinDecompOp Object. }
