\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\usage{BoundLinDecompOp(map,
starttime,
endtime)}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
}
\title{Generic constructor}
\description{Creates a LinearDecompositonOperator from different sources.
Please look at the methods to see what kind of input is supported. }
