% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantOutFluxRate_by_PoolName.R
\docType{methods}
\name{by_PoolIndex,ConstantOutFluxRate_by_PoolName,ANY,ANY-method}
\alias{by_PoolIndex,ConstantOutFluxRate_by_PoolName,ANY,ANY-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolIndex}{ConstantOutFluxRate_by_PoolName,ANY,ANY}(obj,
  poolNames)
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
