% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFlux_by_PoolName.R
\name{by_PoolIndex,InternalFlux_by_PoolName,character,character-method}
\alias{by_PoolIndex,InternalFlux_by_PoolName,character,character-method}
\title{automatic title}
\usage{
\S4method{by_PoolIndex}{InternalFlux_by_PoolName,character,character}(obj, poolNames, timeSymbol)
}
\arguments{
\item{obj}{no manual documentation}

\item{poolNames}{no manual documentation}

\item{timeSymbol}{no manual documentation}
}
\description{
automatic title
}
