% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFluxList_by_PoolName.R
\name{by_PoolIndex,InternalFluxList_by_PoolName,character,character-method}
\alias{by_PoolIndex,InternalFluxList_by_PoolName,character,character-method}
\title{by_PoolIndex,InternalFluxList_by_PoolName,character,character-method}
\usage{
\S4method{by_PoolIndex}{InternalFluxList_by_PoolName,character,character}(obj, poolNames, timeSymbol)
}
\arguments{
\item{obj}{object of class:\code{InternalFluxList_by_PoolName}. }

\item{poolNames}{object of class:\code{character}. }

\item{timeSymbol}{object of class:\code{character}. }
}
\description{
by_PoolIndex,InternalFluxList_by_PoolName,character,character-method
}
