% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection.R
\docType{methods}
\name{by_PoolIndex,PoolConnection,ANY,ANY-method}
\alias{by_PoolIndex,PoolConnection,ANY,ANY-method}
\title{new object with the source pool id and the destination pool id guranteed to be of class PoolIndex}
\usage{
\S4method{by_PoolIndex}{PoolConnection,ANY,ANY}(obj, poolNames)
}
\description{
converts the ids if necessary otherwise returns an 
identical object
}
