% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolSource.R
\name{by_PoolIndex,PoolSource,ANY,ANY-method}
\alias{by_PoolIndex,PoolSource,ANY,ANY-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolIndex}{PoolSource,ANY,ANY}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{PoolSource}. }

\item{poolNames}{object of class:\code{ANY}. }
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
