% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolTarget.R
\docType{methods}
\name{by_PoolIndex,PoolTarget,ANY,ANY-method}
\alias{by_PoolIndex,PoolTarget,ANY,ANY-method}
\title{convert the source pool id to a number if necessary}
\usage{
\S4method{by_PoolIndex}{PoolTarget,ANY,ANY}(obj, poolNames)
}
\description{
convert the source pool id to a number if necessary
}
