% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFlux_by_PoolIndex.R
\name{by_PoolName,ConstantInFlux_by_PoolIndex-method}
\alias{by_PoolName,ConstantInFlux_by_PoolIndex-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolName}{ConstantInFlux_by_PoolIndex}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{ConstantInFlux_by_PoolIndex}, no manual documentation}

\item{poolNames}{no manual documentation}
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
