% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFluxList_by_PoolIndex.R
\name{by_PoolName,ConstantInFluxList_by_PoolIndex-method}
\alias{by_PoolName,ConstantInFluxList_by_PoolIndex-method}
\title{convert to a list indexed by pool names}
\usage{
\S4method{by_PoolName}{ConstantInFluxList_by_PoolIndex}(obj, poolNames)
}
\arguments{
\item{obj}{object of class ConstantInFluxList_by_PoolIndex}

\item{poolNames}{names of pools}
}
\description{
convert to a list indexed by pool names
}
