% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolIndex.R
\name{by_PoolName,ConstantInternalFluxRateList_by_PoolIndex-method}
\alias{by_PoolName,ConstantInternalFluxRateList_by_PoolIndex-method}
\title{convert to a list indexed by pool names}
\usage{
\S4method{by_PoolName}{ConstantInternalFluxRateList_by_PoolIndex}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{ConstantInternalFluxRateList_by_PoolIndex}. }

\item{poolNames}{found automatically by inspection, not documented yet}
}
\description{
convert to a list indexed by pool names
}
