% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantOutFluxRate_by_PoolIndex.R
\docType{methods}
\name{by_PoolName,ConstantOutFluxRate_by_PoolIndex-method}
\alias{by_PoolName,ConstantOutFluxRate_by_PoolIndex-method}
\title{new object with the source pool id converted to a PoolName if necessary}
\usage{
\S4method{by_PoolName}{ConstantOutFluxRate_by_PoolIndex}(obj, poolNames)
}
\description{
This method exists only for classes that do not contain functions of 
the state_variables since we cannot automatically translate functions 
with a state vector arguments to functions of the respective state variables
which would require symbolic computations.
The reverse direction is always possible and is therefore 
the preferred way to input 
rate functions that depend on state variables.
}
