% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection.R
\name{by_PoolName,PoolConnection-method}
\alias{by_PoolName,PoolConnection-method}
\title{new object with the source pool id and the destination pool id guaranteed to be of class PoolName}
\usage{
\S4method{by_PoolName}{PoolConnection}(obj, poolNames)
}
\description{
converts the ids if necessary otherwise returns an 
identical object
}
