% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolTarget.R
\docType{methods}
\name{by_PoolName,PoolTarget-method}
\alias{by_PoolName,PoolTarget-method}
\title{convert the source pool id to a name if necessary}
\usage{
\S4method{by_PoolName}{PoolTarget}(obj, poolNames)
}
\description{
convert the source pool id to a name if necessary
}
