% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{by_PoolName}
\alias{by_PoolName}
\title{automatic title}
\usage{
by_PoolName(obj, poolNames)
}
\arguments{
\item{obj}{see method arguments}

\item{poolNames}{see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{by_PoolName,ConstantInFluxRate_by_PoolIndex-method}}
  \item \code{\link{by_PoolName,ConstantInternalFluxRate_by_PoolIndex-method}}
  \item \code{\link{by_PoolName,ConstantInternalFluxRateList_by_PoolIndex-method}}
  \item \code{\link{by_PoolName,ConstantOutFluxRate_by_PoolIndex-method}}
  \item \code{\link{by_PoolName,ConstantOutFluxRateList_by_PoolIndex-method}}
}
}

