% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolId.R
\name{check_id_length}
\alias{check_id_length}
\title{helper function to check that the length of the argument is exactly 1}
\usage{
check_id_length(id)
}
\arguments{
\item{id}{Either a string or a number}
}
\description{
helper function to check that the length of the argument is exactly 1
}
