% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection.R
\name{check_pool_ids,PoolConnection,integer-method}
\alias{check_pool_ids,PoolConnection,integer-method}
\title{check_pool_ids,PoolConnection,integer-method}
\usage{
\S4method{check_pool_ids}{PoolConnection,integer}(obj, pools)
}
\arguments{
\item{obj}{object of class:\code{PoolConnection}. }

\item{pools}{object of class:\code{integer}. }
}
\description{
check_pool_ids,PoolConnection,integer-method
}
