% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection_by_PoolIndex.R
\name{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}
\alias{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}
\title{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}
\usage{
\S4method{check_pool_ids}{PoolConnection_by_PoolIndex,integer}(obj, pools)
}
\arguments{
\item{obj}{object of class:\code{PoolConnection_by_PoolIndex}. }

\item{pools}{object of class:\code{integer}. }
}
\description{
check_pool_ids,PoolConnection_by_PoolIndex,integer-method
}
