% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{computeResults,MCSim-method}
\alias{computeResults,MCSim-method}
\title{computeResults,MCSim-method}
\usage{
\S4method{computeResults}{MCSim}(object)
}
\arguments{
\item{object}{object of class:\code{MCSim}. }
}
\description{
computeResults,MCSim-method
}
