% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{computeResults,NlModel-method}
\alias{computeResults,NlModel-method}
\title{computeResults,NlModel-method}
\usage{
\S4method{computeResults}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}. }
}
\description{
computeResults,NlModel-method
}
