\name{ConstantDecompositionOperator}
\alias{ConstantDecompositionOperator}
\title{generic constructor}
\description{create a ConstantDecompositonOperator from different sources
Have a look at the methods to see what kind of input is supported. }
\usage{ConstantDecompositionOperator(mat)}\arguments{
  \item{mat}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





