% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantInFlux_by_PoolIndex-class}
\alias{ConstantInFlux_by_PoolIndex-class}
\title{class for a constan influx to a single pool identified by index}
\description{
class for a constan influx to a single pool identified by index
}
