% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantInFlux_by_PoolName-class}
\alias{ConstantInFlux_by_PoolName-class}
\title{class for a constant influx to a single pool identified by pool name}
\description{
class for a constant influx to a single pool identified by pool name
}
