% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantInFluxList_by_PoolIndex-class}
\alias{ConstantInFluxList_by_PoolIndex-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInFlux_by_PoolIndex}}
\description{
Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInFlux_by_PoolIndex}
}
