% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFluxList_by_PoolIndex.R
\docType{methods}
\name{ConstantInFluxList_by_PoolIndex,ConstInFluxes-method}
\alias{ConstantInFluxList_by_PoolIndex,ConstInFluxes-method}
\title{constructor from ConstInFluxes}
\usage{
\S4method{ConstantInFluxList_by_PoolIndex}{ConstInFluxes}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{ConstInFluxes}}
}
\value{
An object of class \linkS4class{ConstantInFluxList_by_PoolIndex}
}
\description{
constructor from ConstInFluxes
}
