% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFluxList_by_PoolIndex.R
\name{ConstantInFluxList_by_PoolIndex,numeric-method}
\alias{ConstantInFluxList_by_PoolIndex,numeric-method}
\title{ConstantInFluxList_by_PoolIndex,numeric-method}
\usage{
\S4method{ConstantInFluxList_by_PoolIndex}{numeric}(object)
}
\arguments{
\item{object}{object of class:\code{numeric}. }
}
\description{
ConstantInFluxList_by_PoolIndex,numeric-method
}
