% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFluxRate_by_PoolName.R
\name{ConstantInFluxRate_by_PoolName}
\alias{ConstantInFluxRate_by_PoolName}
\title{Constructor for the class with the same name}
\usage{
ConstantInFluxRate_by_PoolName(destinationName, rate_constant)
}
\arguments{
\item{destinationName}{Index of the receiving pool (positive integer)}

\item{rate_constant}{Rate (Flux/content) positive real number}
}
\description{
Constructor for the class with the same name
}
