% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantInternalFluxRate_by_PoolIndex-class}
\alias{ConstantInternalFluxRate_by_PoolIndex-class}
\title{S4 class representing a constant internal flux rate}
\description{
The class is used to dispatch specific methods for the creation of the compartmental matrix which is simplified in case of constant rates.
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{ConstantInternalFluxRate_by_PoolIndex} in their signature:}{
      \itemize{
        \item \code{\link{by_PoolName,ConstantInternalFluxRate_by_PoolIndex-method}}
      }
  }
}
