% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRate_by_PoolIndex.R
\docType{methods}
\name{ConstantInternalFluxRate_by_PoolIndex,missing,missing,character,numeric-method}
\alias{ConstantInternalFluxRate_by_PoolIndex,missing,missing,character,numeric-method}
\title{constructor from strings of the form '1_to_2'}
\usage{

  \S4method{ConstantInternalFluxRate_by_PoolIndex}{missing,missing,character,numeric}(src_to_dest,
  rate_constant)
}
\description{
constructor from strings of the form '1_to_2'
}
