% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRate_by_PoolIndex.R
\name{ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method}
\alias{ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method}
\title{automatic title}
\usage{
\S4method{ConstantInternalFluxRate_by_PoolIndex}{numeric,numeric,missing,numeric}(
  sourceIndex,
  destinationIndex,
  rate_constant
)
}
\arguments{
\item{sourceIndex}{object of class:\code{numeric}, no manual documentation}

\item{destinationIndex}{object of class:\code{numeric}, no manual documentation}

\item{rate_constant}{object of class:\code{numeric}, no manual documentation}
}
\description{
automatic title
}
