% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRate_by_PoolIndex.R
\name{ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method}
\alias{ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method}
\title{ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method}
\usage{
\S4method{ConstantInternalFluxRate_by_PoolIndex}{numeric,numeric,missing,numeric}(
  sourceIndex,
  destinationIndex,
  rate_constant
)
}
\arguments{
\item{sourceIndex}{object of class:\code{numeric}. }

\item{destinationIndex}{object of class:\code{numeric}. }

\item{rate_constant}{object of class:\code{numeric}. }
}
\description{
ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method
}
