% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantInternalFluxRate_by_PoolName-class}
\alias{ConstantInternalFluxRate_by_PoolName-class}
\title{S4-class to represent a constant internal flux rate with source and target indexed by name}
\description{
S4-class to represent a constant internal flux rate with source and target indexed by name
}
