% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstantInternalFluxRate_by_PoolName}
\alias{ConstantInternalFluxRate_by_PoolName}
\title{Generic constructor for the class with the same name}
\usage{
ConstantInternalFluxRate_by_PoolName(
  sourceName,
  destinationName,
  src_to_dest,
  rate_constant
)
}
\description{
Generic constructor for the class with the same name
}
