% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolIndex.R
\docType{class}
\name{ConstantInternalFluxRateList_by_PoolIndex-class}
\alias{ConstantInternalFluxRateList_by_PoolIndex-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInternalFluxRate_by_PoolIndex}}
\description{
Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInternalFluxRate_by_PoolIndex}
}
