% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstantInternalFluxRateList_by_PoolIndex}
\alias{ConstantInternalFluxRateList_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
ConstantInternalFluxRateList_by_PoolIndex(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{ConstantInternalFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolIndex,numeric-method}}
      }
}
