% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolName.R
\docType{class}
\name{ConstantInternalFluxRateList_by_PoolName-class}
\alias{ConstantInternalFluxRateList_by_PoolName-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInternalFluxRate_by_PoolName}}
\description{
Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInternalFluxRate_by_PoolName}
}
