% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantOutFluxRate_by_PoolIndex-class}
\alias{ConstantOutFluxRate_by_PoolIndex-class}
\title{S4 Class to represent a single constant out-flux rate with the 
source pool specified by an index}
\description{
S4 Class to represent a single constant out-flux rate with the 
source pool specified by an index
}
