% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstantOutFluxRate_by_PoolIndex}
\alias{ConstantOutFluxRate_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
ConstantOutFluxRate_by_PoolIndex(sourceIndex, rate_constant)
}
\arguments{
\item{sourceIndex}{see method arguments}

\item{rate_constant}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
