% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantOutFluxRateList_by_PoolIndex.R
\docType{class}
\name{ConstantOutFluxRateList_by_PoolIndex-class}
\alias{ConstantOutFluxRateList_by_PoolIndex-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantOutFluxRate_by_PoolIndex}}
\description{
The main purpose of the class is to be used in method signatures 
which would otherwise only indicate an object of class \code{'list'}
in their signature an then check that the list contains the right
kind of elements inside the function.
Using this class the method signature becomes much more informative.
}
