% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantOutFluxRateList_by_PoolIndex.R
\name{ConstantOutFluxRateList_by_PoolIndex,numeric-method}
\alias{ConstantOutFluxRateList_by_PoolIndex,numeric-method}
\title{ConstantOutFluxRateList_by_PoolIndex,numeric-method}
\usage{
\S4method{ConstantOutFluxRateList_by_PoolIndex}{numeric}(object)
}
\arguments{
\item{object}{object of class:\code{numeric}. }
}
\description{
ConstantOutFluxRateList_by_PoolIndex,numeric-method
}
