\title{ConstFc S4 class}
\description{no Description}
\name{ConstFc-class}
\alias{ConstFc-class}
\docType{class}
\section{Methods}{  \describe{
    \item{Model_14}{\code{signature(t = "numeric", A = "ANY", ivList = "numeric", initialValF = "ConstFc", inputFluxes = "ANY", inputFc = "ANY", c14DecayRate = "numeric", solverfunc = "function", pass = "logical")}: ... } \code{\link{Model_14,numeric,ANY,numeric,ConstFc,ANY,ANY,numeric,function,logical-method}}
    \item{Model_14}{\code{signature(t = "numeric", A = "ANY", ivList = "numeric", initialValF = "ConstFc", inputFluxes = "ANY", inputFc = "ANY", c14DecayRate = "numeric", solverfunc = "missing", pass = "logical")}: ... } \code{\link{Model_14,numeric,ANY,numeric,ConstFc,ANY,ANY,numeric,missing,logical-method}}
	 }}
\section{Constructors}{	\code{\link{ConstFc}}\cr
 Please also look at constructors of non virtual subclasses }
