\description{defines a constant inputrate }
\name{ConstInFlux-class}
\alias{ConstInFlux-class}
\docType{class}
\section{Methods}{Exported methods directly defined for class ConstInFlux:

  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "ConstInFlux")}: ... } \code{\link{getFunctionDefinition,ConstInFlux-method}}
    \item{getTimeRange}{\code{signature(object = "ConstInFlux")}: ... } \code{\link{getTimeRange,ConstInFlux-method}}
	 }

Methods inherited from superclasses:

from class InFlux:

  \describe{
    \item{GeneralInFlux}{\code{signature(object = "InFlux")}: ... } \code{\link{GeneralInFlux,InFlux-method}}
	 }}
\title{ConstInFlux S4 class}
\section{Superclasses}{\code{\link{InFlux-class}}\cr}
\section{Constructors}{	\code{\link{ConstInFlux}}\cr}
