\docType{methods}
\name{ConstInFlux,numeric-method}
\title{ConstInFlux,numeric-method}
\alias{ConstInFlux,numeric-method}
\usage{\S4method{ConstInFlux}{numeric}(map)}
\arguments{
\item{map}{: of class numeric}
}
\description{the method converts a vector of constant input rates into an object of Class
\code{\link{ConstInFlux}}.}
