% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{ConstInFlux,numeric-method}
\alias{ConstInFlux,numeric-method}
\title{ConstInFlux,numeric-method constructor}
\arguments{
\item{map}{: of class numeric}
}
\description{
the method converts a vector of constant input rates into an object of Class
\code{\link{ConstInFlux}}.
}
