% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstInFluxes.R
\name{ConstInFluxes,numeric,ANY-method}
\alias{ConstInFluxes,numeric,ANY-method}
\title{automatic title}
\usage{
\S4method{ConstInFluxes}{numeric,ANY}(map)
}
\arguments{
\item{map}{object of class:\code{numeric}, no manual documentation}
}
\description{
automatic title
}
