% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstInFluxes}
\alias{ConstInFluxes}
\title{S4 generic: ConstInFluxes}
\usage{
ConstInFluxes(map, numberOfPools)
}
\arguments{
\item{map}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
S4 generic: ConstInFluxes
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method}}
  \item \code{\link{ConstInFluxes,numeric,ANY-method}}
}
}

