% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{ConstLinDecompOp,matrix-method}
\alias{ConstLinDecompOp,matrix-method}
\title{ConstLinDecompOp,matrix-method construct from matric}
\arguments{
\item{mat}{: of class matrix}
}
\description{
This method creates a ConstLinDecompOp from a matrix The operator is assumed
to act on the vector of carbon stocks by multiplication of the (time
invariant) matrix from the left.
}
