% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{ConstLinDecompOp,matrix,missing,missing,missing,missing-method}
\alias{ConstLinDecompOp,matrix,missing,missing,missing,missing-method}
\title{Constructor}
\usage{
\S4method{ConstLinDecompOp}{matrix,missing,missing,missing,missing}(mat)
}
\arguments{
\item{mat}{object of class:\code{matrix}. }
}
\description{
Constructor
}
