% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolIndex,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method}
\alias{ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolIndex,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method}
\title{Constructor}
\usage{
\S4method{ConstLinDecompOp}{missing,ConstantInternalFluxRateList_by_PoolIndex,ConstantOutFluxRateList_by_PoolIndex,numeric,missing}(internal_flux_rates, out_flux_rates, numberOfPools)
}
\arguments{
\item{internal_flux_rates}{object of class:\code{ConstantInternalFluxRateList_by_PoolIndex}, no manual documentation}

\item{out_flux_rates}{object of class:\code{ConstantOutFluxRateList_by_PoolIndex}, no manual documentation}

\item{numberOfPools}{object of class:\code{numeric}, no manual documentation}
}
\description{
Constructor
}
