\docType{methods}
\name{ConstLinDecompOp,matrix-method}
\title{ConstLinDecompOp,matrix-method 
 construct from matric}
\alias{ConstLinDecompOp,matrix-method}
\usage{\S4method{ConstLinDecompOp}{matrix}(mat)}
\arguments{
\item{mat}{: of class matrix}
}
\description{This method creates a ConstLinDecompOp from a matrix
The operator is assumed to act on the vector of carbon stocks
by multiplication of the (time invariant) matrix from the left.}
