% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolName,ConstantOutFluxRateList_by_PoolName,missing,character-method}
\alias{ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolName,ConstantOutFluxRateList_by_PoolName,missing,character-method}
\title{alternative Constructor with pool names}
\usage{
\S4method{ConstLinDecompOp}{missing,ConstantInternalFluxRateList_by_PoolName,ConstantOutFluxRateList_by_PoolName,missing,character}(internal_flux_rates, out_flux_rates, poolNames)
}
\arguments{
\item{internal_flux_rates}{object of class:\code{ConstantInternalFluxRateList_by_PoolName}. }

\item{out_flux_rates}{object of class:\code{ConstantOutFluxRateList_by_PoolName}. }

\item{poolNames}{object of class:\code{character}. }
}
\description{
alternative Constructor with pool names
}
