% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{ConstLinDecompOp,missing,missing,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method}
\alias{ConstLinDecompOp,missing,missing,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method}
\title{Constructor}
\usage{
\S4method{ConstLinDecompOp}{missing,missing,ConstantOutFluxRateList_by_PoolIndex,numeric,missing}(out_flux_rates, numberOfPools)
}
\arguments{
\item{out_flux_rates}{object of class:\code{ConstantOutFluxRateList_by_PoolIndex}. }

\item{numberOfPools}{object of class:\code{numeric}. }
}
\description{
Constructor
}
