% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{ConstLinDecompOp}
\alias{ConstLinDecompOp}
\title{Constructor}
\usage{
ConstLinDecompOp(mat, internal_flux_rates, out_flux_rates, numberOfPools)
}
\arguments{
\item{mat}{A quadratic compartmental matrix. 
If present nome of \code{internal_flux_rates}, 
code{out_flux_rates}, code{numberOfPools} 
may be present.}

\item{internal_flux_rates}{A list that can be supplied 
in one of the following ways:
\enumerate{
\item A \code{list} of numbers with names of the form "2->4"  
\item A \code{numeric} vector of numbers with names of the form "2->4"  
\item A \code{list} of objects of class 
\linkS4class{ConstantInternalFluxRate_by_PoolIndex}
\item An object of class  
}}

\item{out_flux_rates}{A list that can be supplied in one of the 
following ways:
\enumerate{
\item A \code{list} of numbers with names that can be converted 
to integers ("2" for pool 2  
\item A \code{numeric} vector of numbers with names that can be converted 
to integers.
\item A \code{list} of objects of class 
\linkS4class{ConstantOutFluxRate_by_PoolIndex}
\item An object of class
\linkS4class{ConstantOutFluxRateList_by_PoolIndex}
}}

\item{numberOfPools}{The number of pools, should not be present if the 
\code{mat} argument is supplied but must be present otherwise.}
}
\description{
Constructor
}
