% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstLinDecompOp_by_PoolName}
\alias{ConstLinDecompOp_by_PoolName}
\title{Generic constructor for the class with the same name}
\usage{
ConstLinDecompOp_by_PoolName(internal_flux_rates, out_flux_rates, poolNames)
}
\arguments{
\item{internal_flux_rates}{see method arguments}

\item{out_flux_rates}{see method arguments}

\item{poolNames}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
