% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\docType{class}
\name{ConstLinDecompOpWithLinearScalarFactor-class}
\alias{ConstLinDecompOpWithLinearScalarFactor-class}
\title{convert names of  vectors or lists to class ConstantOutFluxRate 
convert names of  vectors or lists to class ConstantInternalFluxRate 
A class to represent a constant (=nonautonomuous,linear) compartmental matrix}
\description{
convert names of  vectors or lists to class ConstantOutFluxRate 
convert names of  vectors or lists to class ConstantInternalFluxRate 
A class to represent a constant (=nonautonomuous,linear) compartmental matrix
}
