\name{correctnessOfFcAtm}
\alias{correctnessOfFcAtm}
\title{check for unreasonable parameters or unsupported formats}
\description{The atmospheric C14 data can be represented in more than one format 
The function checks if the user required format is supported at the moment}
\usage{correctnessOfFcAtm(object)}
\arguments{
  \item{object}{the object to be tested}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





