\name{correctnessOfModel14}
\alias{correctnessOfModel14}
\title{check for unreasonable input parameters to Model constructor}
\description{The parameters used by the function \code{\link{GeneralModel_14}} in SoilR have a biological meaning, and therefore cannot be arbitrary.
This functions tests some of the obvious constraints of the general model. 
by calling \code{\link{correctnessOfModel}} and additionally takes care of the
following }
\usage{correctnessOfModel14(object)}
\arguments{
  \item{object}{the object to be tested}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





