\name{correctnessOfSoilR.F0}
\alias{correctnessOfSoilR.F0}
\title{check for unreasonable parameters or unsupported formats}
\description{14C fraction data can be represented in more than one format 
The function checks if the user required format is supported at the moment}
\usage{correctnessOfSoilR.F0(object)}
\arguments{
  \item{object}{the object to be tested}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





