% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling.R
\name{cycling}
\alias{cycling}
\title{Cycling analysis of compartmental matrices}
\usage{
cycling(A)
}
\arguments{
\item{A}{A compartmental linear square matrix with cycling rates in the
diagonal and transfer rates in the off-diagonal.}
}
\value{
A list with 2 objects: the fundamental matrix N, and the expected
number of steps Et.
}
\description{
Computes the fundamental matrix N, and the expected number of steps from a
compartmental matrix A
}
\seealso{
\code{\link{systemAge}}
}
