\docType{methods}
\name{DecompOpSubClassInstance,DecompOp-method}
\title{DecompOpSubClassInstance,DecompOp-method}
\alias{DecompOpSubClassInstance,DecompOp-method}
\usage{\S4method{DecompOpSubClassInstance}{DecompOp}(object)}
\arguments{
\item{object}{: of class DecompOp}
}
\description{This method handles the case that no actual construction is necessary since
the argument is already of a subclass of DecompOp }
