\docType{methods}
\name{DecompOpSubClassInstance,matrix-method}
\title{DecompOpSubClassInstance,matrix-method}
\alias{DecompOpSubClassInstance,matrix-method}
\usage{\S4method{DecompOpSubClassInstance}{matrix}(object)}
\arguments{
\item{object}{: of class matrix}
}
\description{The resulting operator is creted by a call to the constructor of class
ConstLinDecompOp}
