\description{no Description}
\name{DecompositionOperator-class}
\alias{DecompositionOperator-class}
\docType{class}
\section{Methods}{Exported methods directly defined for class DecompositionOperator:

  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "DecompositionOperator")}: ... } \code{\link{getFunctionDefinition,DecompositionOperator-method}}
    \item{getTimeRange}{\code{signature(object = "DecompositionOperator")}: ... } \code{\link{getTimeRange,DecompositionOperator-method}}
	 }

Methods inherited from superclasses:

from class DecompOp:

  \describe{
    \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ... } \code{\link{GeneralDecompOp,DecompOp-method}}
	 }}
\title{deprecated decomposition operator class }
\section{Superclasses}{\code{\link{DecompOp-class}}\cr}
