\name{DecompositionOperator}
\alias{DecompositionOperator}
\title{generic constructor and converter }
\description{If the argument is already of a subclass of class DecompositionOperator 
the function returns the unchanged object.
Ohterwise it creates an object of a subclass of DecompositionOperator.
Note that the actual class of the output depends on the argument.
For examples please look at the methods of this function.}
\usage{DecompositionOperator(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





