\name{DecompositionOperator_method__DecompositionOperator}
\alias{DecompositionOperator_method__DecompositionOperator}
\title{pass through constructor}
\description{This method handles the case that no actual construction is necessary since
the argument is already of a subclass of DecompositionOperator }
\arguments{
  \item{object}{
}
}

\value{the unchanged argument}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





