\name{DecompositionOperator_method__matrix}
\alias{DecompositionOperator_method__matrix}
\title{creates a ConstanDecompositionOperator from a matrix}
\description{The resulting operator is creted by a call to the constructor of class
ConstantDecompositionOperator}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





