% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{Delta14C}
\alias{Delta14C}
\title{S4 generic: Delta14C}
\usage{
Delta14C(F)
}
\arguments{
\item{F}{see method arguments}
}
\description{
S4 generic: Delta14C
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{Delta14C,BoundFc-method}}
  \item \code{\link{Delta14C,ConstFc-method}}
}
}

