% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{Delta14C_from_AbsoluteFractionModern,matrix-method}
\alias{Delta14C_from_AbsoluteFractionModern,matrix-method}
\title{Delta14C_from_AbsoluteFractionModern,matrix-method}
\usage{
\S4method{Delta14C_from_AbsoluteFractionModern}{matrix}(AbsoluteFractionModern)
}
\arguments{
\item{AbsoluteFractionModern}{object of class:\code{matrix}. }
}
\description{
Delta14C_from_AbsoluteFractionModern,matrix-method
}
