\docType{methods}
\name{Delta14C_from_AbsoluteFractionModern,matrix-method}
\title{Delta14C_from_AbsoluteFractionModern,matrix-method}
\alias{Delta14C_from_AbsoluteFractionModern,matrix-method}
\usage{\S4method{Delta14C_from_AbsoluteFractionModern}{matrix}(AbsoluteFractionModern)}
\arguments{
\item{AbsoluteFractionModern}{: of class matrix, An object of class matrix containing the values in Absolute Fraction Modern format}
}
\description{This method produces a matrix of  Delta14C values from a matrix of values in Absolute Fraction Modern.}
