% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{Delta14C_from_AbsoluteFractionModern,numeric-method}
\alias{Delta14C_from_AbsoluteFractionModern,numeric-method}
\title{Delta14C_from_AbsoluteFractionModern,numeric-method Converts to Delta14C
format}
\arguments{
\item{AbsoluteFractionModern}{: of class numeric, A numeric object
containing the values in Absolute Fraction Modern format}
}
\description{
This method produces Delta14C values from Absolute Fraction Modern Have a
look at the methods for details.
}
