\name{Delta14C_from_AbsoluteFractionModern}
\alias{Delta14C_from_AbsoluteFractionModern}
\usage{Delta14C_from_AbsoluteFractionModern(AbsoluteFractionModern)}
\arguments{
\item{AbsoluteFractionModern}{A numeric object}
}
\section{Methods}{	\code{\link{Delta14C_from_AbsoluteFractionModern,matrix-method}}\cr
	\code{\link{Delta14C_from_AbsoluteFractionModern,numeric-method}}\cr}
\title{Converts its argument from an Absolute Fraction Modern to a Delta14C representation}
\description{The function returns an object of the same type as its input,
which can be of different type.
Have a look at the methods for details.}
