\name{Delta14C_from_AbsoluteFractionModern_method__matrix}
\alias{Delta14C_from_AbsoluteFractionModern_method__matrix}
\title{convert Absolute Fraction Normal values to Delta14C values }
\description{This method produces a matrix of  Delta14C values from a matrix of Absolute Fraction Normal values.}
\arguments{
  \item{AbsoluteFractionModern}{numeric containing the values in Absolute Fraction Modern format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





