\name{Delta14C_from_AbsoluteFractionModern_method__numeric}
\alias{Delta14C_from_AbsoluteFractionModern_method__numeric}
\title{convert to Delta14C format}
\description{This method produces Delta14C values from   Absolute Fraction Normal values}
\arguments{
  \item{AbsoluteFractionModern}{numeric containing the values in Absolute Fraction Modern format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





