% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection.R
\name{elements_are_PoolConnections}
\alias{elements_are_PoolConnections}
\title{for lists and vectors check if the elements are PoolConnections}
\usage{
elements_are_PoolConnections(vec_or_list)
}
\description{
for lists and vectors check if the elements are PoolConnections
}
