% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DBoundInFluxesFromFunction.R
\name{example.2DBoundInFluxesFromFunction}
\alias{example.2DBoundInFluxesFromFunction}
\title{example.2DBoundInFluxesFromFunction}
\usage{
example.2DBoundInFluxesFromFunction()
}
\value{
The returned object represents a time dependent Influx into a two
pool model.
}
\description{
Create a 2-dimensionsonal example of a BoundInFluxes object
}
