\name{example.2DBoundInFluxFromFunction}
\alias{example.2DBoundInFluxFromFunction}
\usage{example.2DBoundInFluxFromFunction()}
\description{Create a 2-dimensionsonal example of a BoundInFlux object}
\title{example.2DBoundInFluxFromFunction}
\value{The returned object represents a time dependent Influx into 
a two pool model.}
