% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DBoundLinDecompOpFromFunction.R
\name{example.2DBoundLinDecompOpFromFunction}
\alias{example.2DBoundLinDecompOpFromFunction}
\title{example.2DBoundLinDecompOpFromFunction}
\usage{
example.2DBoundLinDecompOpFromFunction()
}
\description{
An example used in tests and other examples.
}
