% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DConstInFluxFromVector.R
\name{example.2DConstInFluxFromVector}
\alias{example.2DConstInFluxFromVector}
\title{2D example for constant Influx}
\usage{
example.2DConstInFluxFromVector()
}
\value{
The returned object represents a time invariant constant influx into
a two pool model.
}
\description{
An example used in tests and other examples.
}
