% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DInFlux.Args.R
\name{example.2DInFlux.Args}
\alias{example.2DInFlux.Args}
\title{example.2DInFlux.Args}
\usage{
example.2DInFlux.Args()
}
\description{
Create a 2-dimensionsonal examples of a Influx objects from different
arguments
}
