% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DInFluxes.Args.R
\name{example.2DInFluxes.Args}
\alias{example.2DInFluxes.Args}
\title{example.2DInFluxes.Args}
\usage{
example.2DInFluxes.Args()
}
\description{
Create a 2-dimensional examples of a Influx objects from different
arguments
}
