% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DUnBoundLinDecompOp.R
\name{example.2DUnBoundLinDecompOpFromFunction}
\alias{example.2DUnBoundLinDecompOpFromFunction}
\title{example.2DUnBoundLinDecompOpFromFunction}
\usage{
example.2DUnBoundLinDecompOpFromFunction()
}
\description{
An example used in tests and other examples.
}
