% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.nestedTime2DMatrixList.R
\name{example.nestedTime2DMatrixList}
\alias{example.nestedTime2DMatrixList}
\title{create an example nested list that can be}
\usage{
example.nestedTime2DMatrixList()
}
\description{
An example used in tests and other examples.
}
