\description{The \eqn{^{14}C}{14C} fraction is a necessary ingredient of any \code{\linkS4class{Model_14}} object.
In the most general case it is a real valued function of time, accompanied     
by a string describing the unit or format (i.e. "Delta14C" or "afn" for Absolute Fraction Modern) .
In the most simple case it is constant real number plus format.}
\name{Fc-class}
\alias{Fc-class}
\docType{class}
\section{Methods}{No exported methods directly defined for class Fc:
}
\title{Fc S4 class}
\section{Subclasses}{\describe{
	\code{\linkS4class{BoundFc}}\cr
	\code{\linkS4class{ConstFc}}\cr
}}
\section{Constructors}{The class is abstract ( \code{contains "VIRTUAL"}).
           It can therefore not be instanciated directly.
           Look at non virtual subclasses and their constructors!
}
