\name{fT.Daycent2}
\alias{fT.Daycent2}
\usage{fT.Daycent2(Temp= )}
\arguments{
\item{Temp}{A scalar or vector containing values of soil temperature for which the effects on decomposition rates are calculated.}
}
\description{Calculates the effects of temperature on decomposition rates according to the Daycent/Century models.}
\references{Del Grosso, S. J., W. J. Parton, A. R. Mosier, E. A. Holland, E. Pendall, D. S. Schimel, and D. S. Ojima (2005), 
Modeling soil CO2 emissions from ecosystems, Biogeochemistry, 73(1), 71-91.}
\title{Effects of temperature on decomposition rates according to the DAYCENT model}
\value{A scalar or a vector containing the effects of temperature on decomposition rates (unitless).}
