% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fT.Demeter.R
\name{fT.Demeter}
\alias{fT.Demeter}
\title{Effects of temperature on decomposition rates according to the DEMETER model}
\usage{
fT.Demeter(Temp, Q10 = 2)
}
\arguments{
\item{Temp}{A scalar or vector containing values of temperature for which
the effects on decomposition rates are calculated}

\item{Q10}{A scalar. Temperature coefficient Q10}
}
\value{
A scalar or a vector containing the effects of temperature on
decomposition rates (unitless).
}
\description{
Calculates the effects of temperature on decomposition rates according to
the DEMETER model.
}
\references{
Foley, J. A. (1995), An equilibrium model of the terrestrial
carbon budget, Tellus B, 47(3), 310-319.
}
