\name{fW.Candy}
\alias{fW.Candy}
\usage{fW.Candy(theta,
PV)}
\arguments{
\item{theta}{A scalar or vector containing values of volumetric soil water content.}
\item{PV}{A scalar or vector containing values of pore volume.}
}
\title{Effects of moisture on decomposition rates according to the Candy model}
\description{Calculates the effects of water content and pore volume on decomposition rates.}
\references{J. Bauer, M. Herbst, J.A. Huisman, L. Weiherm\"uller, H. Vereecken. 2008.
Sensitivity of simulated soil heterotrophic respiration to temperature and moisture reduction functions.
Geoderma, Volume 145, Issues 1-2, 15 May 2008, Pages 17-27.}
