% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fW.Century.R
\name{fW.Century}
\alias{fW.Century}
\title{Effects of moisture on decomposition rates according to the CENTURY model}
\usage{
fW.Century(PPT, PET)
}
\arguments{
\item{PPT}{A scalar or vector containing values of monthly precipitation.}

\item{PET}{A scalar or vector containing values of potential
evapotranspiration.}
}
\value{
A scalar or a vector containing the effects of precipitation and
potential evapotranspiration on decomposition rates (unitless).
}
\description{
Calculates the effects of precipitation and potential evapotranspiration on
decomposition rates.
}
