\name{fW.Demeter}
\alias{fW.Demeter}
\usage{fW.Demeter(M= ,
Msat=100)}
\arguments{
\item{M}{A scalar or vector containing values of soil moisture for which the effects on decomposition rates are calculated.}
\item{Msat}{A scalar representing saturated soil moisture.}
}
\description{Calculates the effects of soil moisture on decomposition rates according to the DEMETER model.}
\references{Foley, J. A. (1995), An equilibrium model of the terrestrial carbon budget, Tellus B, 47(3), 310-319.}
\title{Effects of moisture on decomposition rates according to the DEMETER model}
\value{A scalar or a vector containing the effects of moisture on decomposition rates (unitless).}
