\name{fW.Moyano}
\alias{fW.Moyano}
\usage{fW.Moyano(theta,
a=3.11,
b=2.42)}
\arguments{
\item{theta}{A scalar or vector containing values of volumetric soil water content.}
\item{a}{Empirical parameter}
\item{b}{Empirical parameter}
}
\description{Calculates the effects of water content on decomposition rates.}
\references{F. E. Moyano, S. Manzoni, C. Chenu. 2013
Responses of soil heterotrophic respiration to moisture availability: An exploration of processes and models. 
Soil Biology and Biochemistry, Volume 59, April 2013, Pages 72-85}
\title{Effects of moisture on decomposition rates according to the function proposed by Moyano et al. (2013)}
