\name{fW.Skopp}
\alias{fW.Skopp}
\usage{fW.Skopp(rwc= ,
alpha=2,
beta=2,
f=1.3,
g=0.8)}
\arguments{
\item{rwc}{relative water content}
\item{alpha}{Empirical parameter}
\item{beta}{Empirical parameter}
\item{f}{Empirical parameter}
\item{g}{Empirical parameter}
}
\description{Calculates the effects of relative soil water content on decomposition rates.}
\references{J. Skopp, M. D. Jawson, and J. W. Doran. 1990.
Steady-state aerobic microbial activity as a function of soil water content. Soil Sci. Soc. Am. J., 54(6):1619-1625}
\title{Effects of moisture on decomposition rates according to the function proposed by Skopp et al. 1990}
