% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{GeneralDecompOp,function-method}
\alias{GeneralDecompOp,function-method}
\title{GeneralDecompOp,function-method}
\usage{
\S4method{GeneralDecompOp}{`function`}(object)
}
\arguments{
\item{object}{object of class:\code{function}. }
}
\description{
GeneralDecompOp,function-method
}
