% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralDecompOp,matrix-method}
\alias{GeneralDecompOp,matrix-method}
\title{GeneralDecompOp,matrix-method creates a ConstanDecompOp from a matrix}
\arguments{
\item{object}{: of class matrix}
}
\description{
The resulting operator is creted by a call to the constructor of class
ConstLinDecompOp
}
