\docType{methods}
\name{GeneralDecompOp,list-method}
\title{GeneralDecompOp,list-method 
 creates a BoundLinDecompOp from a nested list of a times vector  and a  list of matrices (a matrix for each time step)}
\alias{GeneralDecompOp,list-method}
\usage{\S4method{GeneralDecompOp}{list}(object)}
\arguments{
\item{object}{: of class list}
}
\description{The resulting operator is creted by a call to the constructor of class
BoundLinDecompOp}
