% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompOp.R
\name{GeneralDecompOp,DecompOp-method}
\alias{GeneralDecompOp,DecompOp-method}
\title{automatic title}
\usage{
\S4method{GeneralDecompOp}{DecompOp}(object)
}
\arguments{
\item{object}{object of class:\code{DecompOp}, : no manual documentation}
}
\description{
automatic title
}
