\docType{methods}
\name{GeneralDecompOp,function-method}
\title{GeneralDecompOp,function-method 
 creates a UnBoundLinDecompOp from a matrix valued function  }
\alias{GeneralDecompOp,function-method}
\usage{\S4method{GeneralDecompOp}{function}(object= )}
\arguments{
\item{object}{: of class function}
}
\description{The resulting operator is created by a call to the constructor of class
UnBoundLinDecompOp}
