\docType{methods}
\name{GeneralDecompOp,matrix-method}
\title{GeneralDecompOp,matrix-method}
\alias{GeneralDecompOp,matrix-method}
\usage{\S4method{GeneralDecompOp}{matrix}(object)}
\arguments{
\item{object}{: of class matrix}
}
\description{The resulting operator is creted by a call to the constructor of class
ConstLinDecompOp}
