% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralDecompOp,TimeMap-method}
\alias{GeneralDecompOp,TimeMap-method}
\title{GeneralDecompOp,TimeMap-method creates a BoundLinDecompOp from a TimeMap
object}
\arguments{
\item{object}{: of class TimeMap}
}
\description{
The resulting operator is creted by a call to the constructor of class
BoundLinDecompOp
}
