% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralDecompOp}
\alias{GeneralDecompOp}
\title{A generic factory for subclasses of GeneralDecompOp}
\usage{
GeneralDecompOp(object)
}
\description{
A generic factory for subclasses of GeneralDecompOp
}
