% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralInFlux,function-method}
\alias{GeneralInFlux,function-method}
\title{GeneralInFlux,function-method creates a UnBoundInFlux from a vector valued
function}
\arguments{
\item{object}{: of class function}
}
\description{
The resulting operator is created by a call to the constructor of class
UnBoundInFlux. You should only use this if the domain of your function is
the complete time axis (-Inf,+Inf). If your function has a finite domain
create an object of class \code{\link{BoundInFlux-class}} ### by calling
\code{\link{BoundInFlux}}. This will activeate checks on that avoid
unintended extrapolation.
}
