% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralInFlux,InFlux-method}
\alias{GeneralInFlux,InFlux-method}
\title{GeneralInFlux,InFlux-method pass through conversion}
\arguments{
\item{object}{: of class InFlux}
}
\description{
This method handles the case that no actual conversion is necessary since
the argument is already of a subclass of \link{InFlux-class}
}
