% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralInFlux,list-method}
\alias{GeneralInFlux,list-method}
\title{GeneralInFlux,list-method creates a BoundInFlux from a nested list of a
times vector and a list of vectors (a vector for each time step)}
\arguments{
\item{object}{: of class list}
}
\description{
The resulting object is created by a call to the constructor of class
BoundInFlux
}
