\docType{methods}
\name{GeneralInFlux,InFlux-method}
\title{GeneralInFlux,InFlux-method}
\alias{GeneralInFlux,InFlux-method}
\usage{\S4method{GeneralInFlux}{InFlux}(object)}
\arguments{
\item{object}{: of class InFlux}
}
\description{This method handles the case that no actual conversion is necessary since
the argument is already of a subclass of \link{InFlux-class} }
