\docType{methods}
\name{GeneralInFlux,list-method}
\title{GeneralInFlux,list-method}
\alias{GeneralInFlux,list-method}
\usage{\S4method{GeneralInFlux}{list}(object)}
\arguments{
\item{object}{: of class list}
}
\description{an a list of vectors(a vector for each time step)
The resulting object is created by a call to the constructor of class
BoundInFlux}
