\docType{methods}
\name{GeneralInFlux,numeric-method}
\title{GeneralInFlux,numeric-method}
\alias{GeneralInFlux,numeric-method}
\usage{\S4method{GeneralInFlux}{numeric}(object)}
\arguments{
\item{object}{: of class numeric}
}
\description{This method enables the model creating functions to handle constant input streams }
