\docType{methods}
\name{GeneralInFlux,TimeMap-method}
\title{GeneralInFlux,TimeMap-method}
\alias{GeneralInFlux,TimeMap-method}
\usage{\S4method{GeneralInFlux}{TimeMap}(object)}
\arguments{
\item{object}{: of class TimeMap}
}
\description{The method is used to ensure backward compatibility with the now deprecated
TimeMap class.
The resulting BoundInFlux is created by a call to
the constructor BoundInFlux(object) of that class.}
