% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralInFlux,numeric-method}
\alias{GeneralInFlux,numeric-method}
\title{GeneralInFlux,numeric-method conversion of a vector to an object of class
\code{\link{ConstInFlux}}}
\arguments{
\item{object}{: of class numeric}
}
\description{
This method enables the model creating functions to handle constant input
streams
}
