% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralInFlux,TimeMap-method}
\alias{GeneralInFlux,TimeMap-method}
\title{GeneralInFlux,TimeMap-method create a BoundInFlux from a TimeMap object}
\arguments{
\item{object}{: of class TimeMap}
}
\description{
The method is used to ensure compatibility TimeMap class. The resulting
BoundInFlux is created by a call to the constructor BoundInFlux(object) of
that class.
}
