\description{no Description}
\name{GeneralInFlux}
\alias{GeneralInFlux}
\usage{GeneralInFlux(object= )}
\title{GeneralInFlux S4 generic}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{GeneralInFlux,InFlux-method}}\cr
	\code{\link{GeneralInFlux,TimeMap-method}}\cr
	\code{\link{GeneralInFlux,function-method}}\cr
	\code{\link{GeneralInFlux,list-method}}\cr
	\code{\link{GeneralInFlux,numeric-method}}\cr}
