\docType{methods}
\name{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,missing-method}
\title{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,missing-method 
 a constructor for class Model <- 14}
\alias{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,missing-method}
\usage{\S4method{GeneralModel_14}{numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,missing}(t,
A,
ivList,
initialValF,
inputFluxes,
inputFc,
di=-0.0001209681,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE)}
\arguments{
\item{t}{: of class numeric}
\item{A}{: of class ANY}
\item{ivList}{: of class numeric}
\item{initialValF}{: of class ANY}
\item{inputFluxes}{: of class ANY}
\item{inputFc}{: of class ANY}
\item{di}{: of class numeric}
\item{solverfunc}{: of class function}
\item{pass}{: of class missing}
}
\description{A wrapper for \code{\link{Model_14}} with pass=FALSE}
