\docType{methods}
\name{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,missing,missing,missing,logical-method}
\title{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,missing,missing,missing,logical-method}
\alias{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,missing,missing,missing,logical-method}
\usage{\S4method{GeneralModel_14}{numeric,ANY,numeric,ANY,ANY,ANY,missing,missing,missing,missing,logical}(t,
A,
ivList,
initialValF,
inputFluxes,
inputFc,
pass)}
\arguments{
\item{t}{: of class numeric, A vector containing the points in time where the solution is sought.}
\item{A}{: of class ANY}
\item{ivList}{: of class numeric}
\item{initialValF}{: of class ANY}
\item{inputFluxes}{: of class ANY}
\item{inputFc}{: of class ANY}
\item{pass}{: of class logical}
}
\description{A wrapper for \code{\link{Model_14}} with pass=FALSE}
