% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolId.R
\name{GeneralPoolId,character-method}
\alias{GeneralPoolId,character-method}
\title{automatic title}
\usage{
\S4method{GeneralPoolId}{character}(id)
}
\arguments{
\item{id}{object of class:\code{character}, : no manual documentation}
}
\description{
automatic title
}
