% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolId.R
\name{GeneralPoolId,numeric-method}
\alias{GeneralPoolId,numeric-method}
\title{generic factory for this virtual class}
\usage{
\S4method{GeneralPoolId}{numeric}(id)
}
\arguments{
\item{id}{object of class:\code{numeric}, no manual documentation}
}
\description{
the class returned depends on the method dispached depending on the supplied arguments
}
