% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralPoolId}
\alias{GeneralPoolId}
\title{General pool Id}
\usage{
GeneralPoolId(id)

GeneralPoolId(id)
}
\arguments{
\item{id}{see method arguments}
}
\description{
General pool Id

General pool Id
}
