\docType{methods}
\name{getAccumulatedRelease,Model-method}
\title{getAccumulatedRelease,Model-method 
 time integrals of release fluxes per pool}
\alias{getAccumulatedRelease,Model-method}
\usage{\S4method{getAccumulatedRelease}{Model}(object= )}
\arguments{
\item{object}{: of class Model}
}
\description{The method integrates the release flux of every pool for all times in the  interval specified by the model definition.}
