\name{getAccumulatedRelease14}
\alias{getAccumulatedRelease14}
\title{access to the C content of the pools }

\usage{getAccumulatedRelease14(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





