\docType{methods}
\name{getC,NlModel-method}
\title{getC,NlModel-method}
\alias{getC,NlModel-method}
\usage{\S4method{getC}{NlModel}(object,
as.closures=FALSE)}
\arguments{
\item{object}{: of class NlModel, the model}
\item{as.closures}{:  a flag that in TRUE will return an approximating function instead of the values}
}
\description{This function computes the value for C for each time and pool. }
