% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getC,NlModel-method}
\alias{getC,NlModel-method}
\title{getC,NlModel-method compute the solution for the content of the pools}
\arguments{
\item{object}{: of class NlModel, the model}

\item{as.closures}{: a flag that in TRUE will return an approximating
function instead of the values}
}
\description{
This function computes the value for C for each time and pool.
}
