\docType{methods}
\name{getC14,Model_14-method}
\title{getC14,Model_14-method}
\alias{getC14,Model_14-method}
\usage{\S4method{getC14}{Model_14}(object= )}
\arguments{
\item{object}{: of class Model_14}
}
\description{This function computes the value for \eqn{^{14}C}{14C} (mass or concentration ) as function of time}
