% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getC14,Model_14-method}
\alias{getC14,Model_14-method}
\title{getC14,Model_14-method}
\arguments{
\item{object}{: of class Model_14}
}
\description{
This function computes the value for \eqn{^{14}C}{14C} (mass or
concentration ) as function of time
}
