\name{getC_method__Model}
\alias{getC_method__Model}
\title{getC method  Model}
\description{This function computes the value for C for each time and pool. }
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





