% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\alias{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\title{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\usage{
\S4method{getCompartmentalMatrixFunc}{ConstLinDecompOp,ANY,ANY}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. }
}
\description{
getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method
}
