% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp.R
\name{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp,ANY,ANY-method}
\alias{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{getCompartmentalMatrixFunc}{UnBoundNonLinDecompOp,ANY,ANY}(object)

\S4method{getCompartmentalMatrixFunc}{UnBoundNonLinDecompOp,ANY,ANY}(object)
}
\arguments{
\item{object}{object of class:\code{UnBoundNonLinDecompOp}, : no manual documentation}
}
\description{
automatic title

automatic title
}
