% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getCompartmentalMatrixFunc,TransportDecompositionOperator,ANY,ANY-method}
\alias{getCompartmentalMatrixFunc,TransportDecompositionOperator,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{getCompartmentalMatrixFunc}{TransportDecompositionOperator,ANY,ANY}(object)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}, : no manual documentation}
}
\description{
automatic title
}
