% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getCompartmentalMatrixFunc,TransportDecompositionOperator-method}
\alias{getCompartmentalMatrixFunc,TransportDecompositionOperator-method}
\title{automatic title}
\usage{
\S4method{getCompartmentalMatrixFunc}{TransportDecompositionOperator}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
