% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\name{getConstantCompartmentalMatrix,ConstLinDecompOpWithLinearScalarFactor-method}
\alias{getConstantCompartmentalMatrix,ConstLinDecompOpWithLinearScalarFactor-method}
\title{automatic title}
\usage{
\S4method{getConstantCompartmentalMatrix}{ConstLinDecompOpWithLinearScalarFactor}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOpWithLinearScalarFactor}, : no manual documentation}
}
\description{
automatic title
}
