% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getConstantCompartmentalMatrix}
\alias{getConstantCompartmentalMatrix}
\title{getConstantCompartmentalMatrix}
\usage{
getConstantCompartmentalMatrix(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
getConstantCompartmentalMatrix
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getConstantCompartmentalMatrix,ConstLinDecompOp-method}}
  \item \code{\link{getConstantCompartmentalMatrix,ConstLinDecompOpWithLinearScalarFactor-method}}
  \item \code{\link{getConstantCompartmentalMatrix,ConstLinDecompOp-method}}
  \item \code{\link{getConstantCompartmentalMatrix,ConstLinDecompOpWithLinearScalarFactor-method}}
}
}

