% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstInFluxes.R
\name{getConstantInFluxVector,ConstInFluxes-method}
\alias{getConstantInFluxVector,ConstInFluxes-method}
\title{getConstantInFluxVector,ConstInFluxes-method}
\usage{
\S4method{getConstantInFluxVector}{ConstInFluxes}(object)
}
\arguments{
\item{object}{object of class:\code{ConstInFluxes}. }
}
\description{
getConstantInFluxVector,ConstInFluxes-method
}
