% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getConstantInFluxVector}
\alias{getConstantInFluxVector}
\title{getConstantInFluxVector}
\usage{
getConstantInFluxVector(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
getConstantInFluxVector
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getConstantInFluxVector,ConstInFluxes-method}}
  \item \code{\link{getConstantInFluxVector,ConstInFluxes-method}}
}
}

