% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getConstantInternalFluxRateList_by_PoolIndex,ConstLinDecompOp-method}
\alias{getConstantInternalFluxRateList_by_PoolIndex,ConstLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{getConstantInternalFluxRateList_by_PoolIndex}{ConstLinDecompOp}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}, no manual documentation}
}
\description{
automatic title
}
