% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getConstantOutFluxRateList_by_PoolIndex,ConstLinDecompOp-method}
\alias{getConstantOutFluxRateList_by_PoolIndex,ConstLinDecompOp-method}
\title{Constant out flux rate list by pool index}
\usage{
\S4method{getConstantOutFluxRateList_by_PoolIndex}{ConstLinDecompOp}(object)
}
\arguments{
\item{object}{a ConstLinDecompOp object}
}
\description{
Constant out flux rate list by pool index
}
