% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getCumulativeC}
\alias{getCumulativeC}
\title{S4 generic: getCumulativeC}
\usage{
getCumulativeC(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getCumulativeC
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getCumulativeC,NlModel-method}}
}
}

