% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getDecompOp,NlModel-method}
\alias{getDecompOp,NlModel-method}
\title{getDecompOp,NlModel-method}
\usage{
\S4method{getDecompOp}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}. }
}
\description{
getDecompOp,NlModel-method
}
