\name{getDecompositionOperator}
\alias{getDecompositionOperator}
\title{getDecompositionOperator}
\description{extract the InputFluxes from a model object}
\usage{getDecompositionOperator(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





