% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getDotOut,TransportDecompositionOperator-method}
\alias{getDotOut,TransportDecompositionOperator-method}
\title{automatic title}
\usage{
\S4method{getDotOut}{TransportDecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}, : no manual documentation}
}
\description{
automatic title
}
