% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getDotOut}
\alias{getDotOut}
\title{S4 generic: getDotOut}
\usage{
getDotOut(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getDotOut
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getDotOut,TransportDecompositionOperator-method}}
}
}

