\docType{methods}
\name{getF14,Model_14-method}
\title{getF14,Model_14-method}
\alias{getF14,Model_14-method}
\usage{\S4method{getF14}{Model_14}(object)}
\arguments{
\item{object}{: of class Model_14, The model}
}
\description{Calculates the radiocarbon fraction for each pool at each time step.
}
\value{A matrix of dimension n x m; i.e. n time steps as rows and m pools as columns.}
