% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getF14,Model_14-method}
\alias{getF14,Model_14-method}
\title{getF14,Model_14-method radiocarbon}
\arguments{
\item{object}{: of class Model_14, The model}
}
\description{
Calculates the radiocarbon fraction for each pool at each time step.
}
