% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getF14}
\alias{getF14}
\title{Generic that yields the 14C fraction for the content all pools and all times}
\usage{
getF14(object)
}
\arguments{
\item{object}{A SoilR object of class Model14}
}
\description{
Generic that yields the 14C fraction for the content all pools and all times
}
