% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getF14C,Model_14-method}
\alias{getF14C,Model_14-method}
\title{getF14C,Model_14-method read access to the models F14C variable}
\arguments{
\item{object}{: of class Model_14}
}
\description{
The model was created with a F14C object to describe the atmospheric 14C
content. This method serves to investigate those settings from the model.
}
