% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_14.R
\name{getF14R,Model_14-method}
\alias{getF14R,Model_14-method}
\title{getF14R,Model_14-method}
\usage{
\S4method{getF14R}{Model_14}(object)
}
\arguments{
\item{object}{object of class:\code{Model_14}. }
}
\description{
getF14R,Model_14-method
}
