\name{getF14R}
\alias{getF14R}
\usage{getF14R(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getF14R,Model_14-method}}\cr}
\title{Calculates the average radiocarbon fraction weighted by the amount of carbon release}
\description{This function calculates the average radiocarbon fraction weighted by the amount of carbon release at each time step.
Have a look at the methods for details.}
