% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fc.R
\name{getFormat,Fc-method}
\alias{getFormat,Fc-method}
\title{Automatic description: getFormat,Fc-method}
\usage{
\S4method{getFormat}{Fc}(object)
}
\arguments{
\item{object}{object of class:\code{Fc}. : object of class:\code{Fc}, no manual documentation}
}
\description{
Automatic description: getFormat,Fc-method
}
