% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getFormat}
\alias{getFormat}
\title{S4 generic: getFormat}
\usage{
getFormat(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getFormat
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getFormat,Fc-method}}
}
}

