% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getFunctionDefinition,ConstInFlux-method}
\alias{getFunctionDefinition,ConstInFlux-method}
\title{getFunctionDefinition,ConstInFlux-method}
\arguments{
\item{object}{: of class ConstInFlux}
}
\value{
A constant function of time that is used by the Models to represent
the input fluxes.
}
\description{
create the (constant) function of time that is required by the models
}
