% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getFunctionDefinition,ConstLinDecompOp-method}
\alias{getFunctionDefinition,ConstLinDecompOp-method}
\title{getFunctionDefinition,ConstLinDecompOp-method}
\usage{
\S4method{getFunctionDefinition}{ConstLinDecompOp}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. }
}
\description{
getFunctionDefinition,ConstLinDecompOp-method
}
