% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\name{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}
\alias{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}
\title{helper function}
\usage{
\S4method{getFunctionDefinition}{ConstLinDecompOpWithLinearScalarFactor}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOpWithLinearScalarFactor}, no manual documentation}
}
\description{
helper function
}
