% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getFunctionDefinition,DecompositionOperator-method}
\alias{getFunctionDefinition,DecompositionOperator-method}
\title{getFunctionDefinition,DecompositionOperator-method}
\arguments{
\item{object}{: of class DecompositionOperator}
}
\description{
extract the function definition (the R-function)
}
