% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxList_by_PoolIndex.R
\name{getFunctionDefinition,InFluxList_by_PoolIndex-method}
\alias{getFunctionDefinition,InFluxList_by_PoolIndex-method}
\title{getFunctionDefinition,InFluxList_by_PoolIndex-method}
\usage{
\S4method{getFunctionDefinition}{InFluxList_by_PoolIndex}(object, numberOfPools)
}
\arguments{
\item{object}{object of class:\code{InFluxList_by_PoolIndex}. }

\item{numberOfPools}{}
}
\description{
getFunctionDefinition,InFluxList_by_PoolIndex-method
}
