\docType{methods}
\name{getFunctionDefinition,DecompositionOperator-method}
\title{getFunctionDefinition,DecompositionOperator-method}
\alias{getFunctionDefinition,DecompositionOperator-method}
\usage{\S4method{getFunctionDefinition}{DecompositionOperator}(object)}
\arguments{
\item{object}{: of class DecompositionOperator}
}
\description{extract the function definition (the R-function) }
