% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{getFunctionDefinition,TimeMap-method}
\alias{getFunctionDefinition,TimeMap-method}
\title{automatic title}
\usage{
\S4method{getFunctionDefinition}{TimeMap}(object)
}
\arguments{
\item{object}{object of class:\code{TimeMap}, : no manual documentation}
}
\description{
automatic title
}
