\docType{methods}
\name{getFunctionDefinition,TransportDecompositionOperator-method}
\title{getFunctionDefinition,TransportDecompositionOperator-method}
\alias{getFunctionDefinition,TransportDecompositionOperator-method}
\usage{\S4method{getFunctionDefinition}{TransportDecompositionOperator}(object)}
\arguments{
\item{object}{: of class TransportDecompositionOperator}
}
\description{extract the function definition (the R-function) from the object}
