\docType{methods}
\name{getFunctionDefinition,UnBoundInFlux-method}
\title{getFunctionDefinition,UnBoundInFlux-method}
\alias{getFunctionDefinition,UnBoundInFlux-method}
\usage{\S4method{getFunctionDefinition}{UnBoundInFlux}(object)}
\arguments{
\item{object}{: of class UnBoundInFlux}
}
\description{The method creates a timedependent function from the existing matrix describing the operator }
