\docType{methods}
\name{getFunctionDefinition,TimeMap-method}
\title{getFunctionDefinition,TimeMap-method}
\alias{getFunctionDefinition,TimeMap-method}
\usage{\S4method{getFunctionDefinition}{TimeMap}(object= )}
\arguments{
\item{object}{: of class TimeMap}
}
\description{extract the function definition (the R-function) }
