\docType{methods}
\name{getFunctionDefinition,ConstInFlux-method}
\title{getFunctionDefinition,ConstInFlux-method}
\alias{getFunctionDefinition,ConstInFlux-method}
\usage{\S4method{getFunctionDefinition}{ConstInFlux}(object= )}
\arguments{
\item{object}{: of class ConstInFlux}
}
\description{create the (constant) function of time that is required by the models }
\value{A constant function of time that is used by the Models to represent the input fluxes.}
