\docType{methods}
\name{getFunctionDefinition,UnBoundLinDecompOp-method}
\title{getFunctionDefinition,UnBoundLinDecompOp-method}
\alias{getFunctionDefinition,UnBoundLinDecompOp-method}
\usage{\S4method{getFunctionDefinition}{UnBoundLinDecompOp}(object)}
\arguments{
\item{object}{: of class UnBoundLinDecompOp}
}
\description{The method creates a timedependent function from the existing matrix describing the operator }
