% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{getFunctionDefinition,UnBoundLinDecompOp-method}
\alias{getFunctionDefinition,UnBoundLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{getFunctionDefinition}{UnBoundLinDecompOp}(object)
}
\arguments{
\item{object}{object of class:\code{UnBoundLinDecompOp}, : no manual documentation}
}
\description{
automatic title
}
