\docType{methods}
\name{getFunctionDefinition,ConstLinDecompOp-method}
\title{getFunctionDefinition,ConstLinDecompOp-method}
\alias{getFunctionDefinition,ConstLinDecompOp-method}
\usage{\S4method{getFunctionDefinition}{ConstLinDecompOp}(object)}
\arguments{
\item{object}{: of class ConstLinDecompOp}
}
\description{The method creates a timedependent function from the existing matrix describing the operator }
