% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateDependentInFluxVector.R
\name{getFunctionDefinition,StateDependentInFluxVector-method}
\alias{getFunctionDefinition,StateDependentInFluxVector-method}
\title{automatic title}
\usage{
\S4method{getFunctionDefinition}{StateDependentInFluxVector}(object)
}
\arguments{
\item{object}{object of class:\code{StateDependentInFluxVector}, no manual documentation}
}
\description{
automatic title
}
