% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getFunctionDefinition,TransportDecompositionOperator-method}
\alias{getFunctionDefinition,TransportDecompositionOperator-method}
\title{getFunctionDefinition,TransportDecompositionOperator-method}
\arguments{
\item{object}{: of class TransportDecompositionOperator}
}
\description{
extract the function definition (the R-function) from the object
}
