% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getFunctionDefinition,UnBoundInFlux-method}
\alias{getFunctionDefinition,UnBoundInFlux-method}
\title{getFunctionDefinition,UnBoundInFlux-method creates a constant timedependent
function and returns it}
\arguments{
\item{object}{: of class UnBoundInFlux}
}
\description{
The method creates a timedependent function from the existing matrix
describing the operator
}
