% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getFunctionDefinition,UnBoundLinDecompOp-method}
\alias{getFunctionDefinition,UnBoundLinDecompOp-method}
\title{getFunctionDefinition,UnBoundLinDecompOp-method creates a constant
timedependent function and returns it}
\arguments{
\item{object}{: of class UnBoundLinDecompOp}
}
\description{
The method creates a timedependent function from the existing matrix
describing the operator
}
