\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\usage{getFunctionDefinition(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\title{getFunctionDefinition generic function}
\description{Extracts the function definition (the R-function) from the argument
Have a look at the methods for details.}
