% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getInFluxes,Model-method}
\alias{getInFluxes,Model-method}
\title{getInFluxes,Model-method}
\usage{
\S4method{getInFluxes}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. }
}
\description{
getInFluxes,Model-method
}
