% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getInFluxes}
\alias{getInFluxes}
\title{Extract the influxes}
\usage{
getInFluxes(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Extract the influxes
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getInFluxes,Model-method}}
        \item \code{\link{getInFluxes,NlModel-method}}
      }
}
