% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getInitialValues}
\alias{getInitialValues}
\title{S4 generic: getInitialValues}
\usage{
getInitialValues(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getInitialValues
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getInitialValues,NlModel-method}}
}
}

