% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getMeanTransitTime,ConstLinDecompOp-method}
\alias{getMeanTransitTime,ConstLinDecompOp-method}
\title{Automatic description: getMeanTransitTime,ConstLinDecompOp-method}
\usage{
\S4method{getMeanTransitTime}{ConstLinDecompOp}(object, inputDistribution)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. : object of class:\code{ConstLinDecompOp}, no manual
documentation}

\item{inputDistribution}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: getMeanTransitTime,ConstLinDecompOp-method
}
