\name{getMeanTransitTime_method__ConstantDecompositionOperator}
\alias{getMeanTransitTime_method__ConstantDecompositionOperator}
\title{compute the mean transit time }
\description{the description is incomplete}
\arguments{
  \item{object}{
}
  \item{inputDistribution}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





