% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{getNumberOfPools,MCSim-method}
\alias{getNumberOfPools,MCSim-method}
\title{Automatic description: getNumberOfPools,MCSim-method}
\usage{
\S4method{getNumberOfPools}{MCSim}(object)
}
\arguments{
\item{object}{object of class:\code{MCSim}. : object of class:\code{MCSim}, no manual documentation}
}
\description{
Automatic description: getNumberOfPools,MCSim-method
}
