% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getNumberOfPools,NlModel-method}
\alias{getNumberOfPools,NlModel-method}
\title{automatic title}
\usage{
\S4method{getNumberOfPools}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}, : no manual documentation}
}
\description{
automatic title
}
