% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getNumberOfPools}
\alias{getNumberOfPools}
\title{S4 generic: getNumberOfPools}
\usage{
getNumberOfPools(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getNumberOfPools
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getNumberOfPools,MCSim-method}}
  \item \code{\link{getNumberOfPools,NlModel-method}}
  \item \code{\link{getNumberOfPools,TransportDecompositionOperator-method}}
}
}

