% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getOutputFluxes,NlModel-method}
\alias{getOutputFluxes,NlModel-method}
\title{automatic title}
\usage{
\S4method{getOutputFluxes}{NlModel}(object, as.closures = F)
}
\arguments{
\item{object}{object of class:\code{NlModel}, : no manual documentation}

\item{as.closures}{: no manual documentation}
}
\description{
automatic title
}
