% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getOutputFluxes}
\alias{getOutputFluxes}
\title{Generic Function to obtain the fluxes out of of the pools}
\usage{
getOutputFluxes(object, as.closures = F)
}
\arguments{
\item{object}{a SoilR object}

\item{as.closures}{a logical variable. Default to FALSE}
}
\description{
Generic Function to obtain the fluxes out of of the pools
}
