% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\alias{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\title{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\usage{
\S4method{getOutputReceivers}{TransportDecompositionOperator,numeric}(object, i)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}. }

\item{i}{object of class:\code{numeric}. }
}
\description{
getOutputReceivers,TransportDecompositionOperator,numeric-method
}
