% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getParticleMonteCarloSimulator}
\alias{getParticleMonteCarloSimulator}
\title{getParticleMonteCarloSimulator}
\usage{
getParticleMonteCarloSimulator(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
getParticleMonteCarloSimulator
}
