\name{getRelease14}
\alias{getRelease14}
\title{access to the C content of the pools }
\description{This function computes the overall  14C  release of the given model as funtion of time }
\usage{getRelease14(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





