\docType{methods}
\name{getReleaseFlux,NlModel-method}
\title{getReleaseFlux,NlModel-method 
 get the release rate for all pools }
\alias{getReleaseFlux,NlModel-method}
\usage{\S4method{getReleaseFlux}{NlModel}(object= )}
\arguments{
\item{object}{: of class NlModel, an object of class NlModel}
}
\description{The method computes the release of carbon per time for all points in time 
specified in the objects time slot.}
