% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getReleaseFlux,Model-method}
\alias{getReleaseFlux,Model-method}
\title{automatic title}
\usage{
\S4method{getReleaseFlux}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}, : no manual documentation}
}
\description{
automatic title
}
