% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getReleaseFlux,Model-method}
\alias{getReleaseFlux,Model-method}
\title{Automatic description: getReleaseFlux,Model-method}
\usage{
\S4method{getReleaseFlux}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. : object of class:\code{Model}, no manual documentation}
}
\description{
Automatic description: getReleaseFlux,Model-method
}
