% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{getReleaseFlux,Model_by_PoolNames-method}
\alias{getReleaseFlux,Model_by_PoolNames-method}
\title{getReleaseFlux,Model_by_PoolNames-method}
\usage{
\S4method{getReleaseFlux}{Model_by_PoolNames}(object)
}
\arguments{
\item{object}{object of class:\code{Model_by_PoolNames}. }
}
\description{
getReleaseFlux,Model_by_PoolNames-method
}
