% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getReleaseFlux,NlModel-method}
\alias{getReleaseFlux,NlModel-method}
\title{getReleaseFlux,NlModel-method get the release rate for all pools}
\arguments{
\item{object}{: of class NlModel, an object of class NlModel}
}
\description{
The method computes the release of carbon per time for all points in time
specified in the objects time slot.
}
