% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getReleaseFlux14}
\alias{getReleaseFlux14}
\title{S4 generic: getReleaseFlux14}
\usage{
getReleaseFlux14(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getReleaseFlux14
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getReleaseFlux14,Model_14-method}}
}
}

