% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getRightHandSideOfODE,Model-method}
\alias{getRightHandSideOfODE,Model-method}
\title{Automatic description: getRightHandSideOfODE,Model-method}
\usage{
\S4method{getRightHandSideOfODE}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. : object of class:\code{Model}, no manual documentation}
}
\description{
Automatic description: getRightHandSideOfODE,Model-method
}
