% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{getRightHandSideOfODE,Model_by_PoolNames-method}
\alias{getRightHandSideOfODE,Model_by_PoolNames-method}
\title{Provide the (vector valued) derivative of the stocks with respect to time}
\usage{
\S4method{getRightHandSideOfODE}{Model_by_PoolNames}(object)
}
\arguments{
\item{object}{The model}
}
\description{
This function is required by the ODE solvers.
}
