% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getRightHandSideOfODE}
\alias{getRightHandSideOfODE}
\title{getRightHandSideOfODE}
\usage{
getRightHandSideOfODE(object, timeSymbol, poolNames, numberOfPools)
}
\arguments{
\item{object}{see method arguments}

\item{timeSymbol}{see method arguments}

\item{poolNames}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
getRightHandSideOfODE
}
