\name{getSoilC14Fraction}
\alias{getSoilC14Fraction}
\title{access to the C content of the pools }
\description{This function computes the \eqn{\frac{^{14}C}{C}}{14C/C} ratio in the soil as funtion of time }
\usage{getSoilC14Fraction(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





