% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getSolution}
\alias{getSolution}
\title{Calculates all stocks all fluxes to ,in and out of  the compartment system and also their integrals over time}
\usage{
getSolution(object, params, as.closures = F)
}
\arguments{
\item{object}{A modelrun as produced by the constructors: 
 \code{\link{Model}},
 \code{\link{Model_by_PoolNames}},
 \code{\link{Model_14}}
the function \code{\link{GeneralModel}} or the functions listed in 
\code{\link{predefinedModels}}. \cr
A model represents the initial value problem (IVP)  
for the contents of the pool consisting of 
\itemize{
 \item{ The initial values of the pool content}
 \item{ The system of ordinary differential equations, as dictated by the fluxes}
 \item{ The times for which the solution of the IVP is evaluated.}
}}

\item{params}{parameter values}

\item{as.closures}{a logical variable. Default to FALSE}
}
\value{
A matrix with columns representing the name of the statevariable, flux and accumulated flux for every time

as specified by the \code{times} of the model.
}
\description{
Have a look at the methods for details.
}
