% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getSrcDim}
\alias{getSrcDim}
\title{automatic title}
\usage{
getSrcDim(obj)
}
\arguments{
\item{obj}{see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{}}
}
}

