% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstInFluxes.R
\name{getTimeRange,ConstInFluxes-method}
\alias{getTimeRange,ConstInFluxes-method}
\title{getTimeRange,ConstInFluxes-method}
\usage{
\S4method{getTimeRange}{ConstInFluxes}(object)
}
\arguments{
\item{object}{object of class:\code{ConstInFluxes}. }
}
\description{
getTimeRange,ConstInFluxes-method
}
