% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{getTimeRange,UnBoundInFluxes-method}
\alias{getTimeRange,UnBoundInFluxes-method}
\title{Automatic description: getTimeRange,UnBoundInFluxes-method}
\usage{
\S4method{getTimeRange}{UnBoundInFluxes}(object)
}
\arguments{
\item{object}{object of class:\code{UnBoundInFluxes}. : object of class:\code{UnBoundInFluxes}, no manual
documentation}
}
\description{
Automatic description: getTimeRange,UnBoundInFluxes-method
}
