% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getTimeRange,UnBoundLinDecompOp-method}
\alias{getTimeRange,UnBoundLinDecompOp-method}
\title{getTimeRange,UnBoundLinDecompOp-method return an (infinite) time range since
the function is assumed to be valid for all times}
\arguments{
\item{object}{: of class UnBoundLinDecompOp}
}
\description{
some functions dealing with DecompOps in general rely on this so we have to
implement it even though the timerange is always the same: (-inf,inf)
}
