\name{getTimeRange}
\alias{getTimeRange}
\usage{getTimeRange(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getTimeRange,ConstInFlux-method}}\cr
	\code{\link{getTimeRange,ConstLinDecompOp-method}}\cr
	\code{\link{getTimeRange,DecompositionOperator-method}}\cr
	\code{\link{getTimeRange,TimeMap-method}}\cr}
\title{getTimeRange generic function}
\description{This function returns the time range of the given object. 
Have a look at the methods for details.}
