\name{getTimeRange_method__ConstantDecompositionOperator}
\alias{getTimeRange_method__ConstantDecompositionOperator}
\title{return an (infinite) time range since the operator is constant}
\description{some functions dealing with DecompositionOperators in general rely on this
so we have to implement it even though the timerange is always the same: (-inf,inf)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





