% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getTimes,Model-method}
\alias{getTimes,Model-method}
\title{getTimes,Model-method}
\arguments{
\item{object}{: of class Model}
}
\description{
This functions extracts the times argument from an argument of class Model
}
