% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getTimes,NlModel-method}
\alias{getTimes,NlModel-method}
\title{getTimes,NlModel-method}
\arguments{
\item{object}{: of class NlModel}
}
\description{
This functions extracts the times argument from an object of class NlModel
}
