\name{getTimes}
\alias{getTimes}
\usage{getTimes(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getTimes,Model-method}}\cr
	\code{\link{getTimes,NlModel-method}}\cr}
\title{Extracts the times argument}
\description{This functions extracts the times argument from an argument of class NlModel}
