\name{getTotalReleaseFluxC14CRatio}
\alias{getTotalReleaseFluxC14CRatio}
\title{access to the C release flux from the pools }
\description{This function computes the \eqn{\frac{^{14}C}{C}}{14C/C} ratio of the released C as funtion of time }
\usage{getTotalReleaseFluxC14CRatio(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





