% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransferMatrix.R
\name{getTransferMatrix}
\alias{getTransferMatrix}
\title{deprecated, use getTransferMatrixFunc instead}
\usage{
getTransferMatrix(object)
}
\arguments{
\item{object}{A compartmental operator}
}
\description{
deprecated, use getTransferMatrixFunc instead
}
