% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getTransferMatrixFunc,TransportDecompositionOperator-method}
\alias{getTransferMatrixFunc,TransportDecompositionOperator-method}
\title{getTransferMatrixFunc,TransportDecompositionOperator-method}
\usage{
\S4method{getTransferMatrixFunc}{TransportDecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}. }
}
\description{
getTransferMatrixFunc,TransportDecompositionOperator-method
}
