\docType{methods}
\name{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\title{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\alias{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\usage{\S4method{getTransitTimeDistributionDensity}{ConstLinDecompOp}(object,
inputDistribution,
times)}
\arguments{
\item{object}{: of class ConstLinDecompOp}
\item{inputDistribution}{:  distribution of the inputs}
\item{times}{:  the points in time where the solution is sought}
}
\description{This mehtod computes the probability density of the transit time of the linear time invariant system
that can be constructed from the given operator and input distribution.}
\references{Manzoni, S., Katul, G.~G., and Porporato, A.: Analysis of soil carbon transit
times and age distributions using network theories, J. Geophys. Res., 114,
}
