% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTransitTimeDistributionDensity}
\alias{getTransitTimeDistributionDensity}
\title{getTransitTimeDistributionDensity}
\usage{
getTransitTimeDistributionDensity(object, inputDistribution, times)
}
\arguments{
\item{object}{see method arguments}

\item{inputDistribution}{see method arguments}

\item{times}{see method arguments}
}
\description{
getTransitTimeDistributionDensity
}
