\name{getTransitTimeDistributionDensity_method__ConstantDecompositionOperator}
\alias{getTransitTimeDistributionDensity_method__ConstantDecompositionOperator}
\title{compute the TransitTimeDistributionDensity}
\description{the description is incomplete}
\arguments{
  \item{object}{
}
  \item{inputDistribution}{
}
  \item{times}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





