% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstFc.R
\name{getValues,ConstFc-method}
\alias{getValues,ConstFc-method}
\title{getValues,ConstFc-method}
\usage{
\S4method{getValues}{ConstFc}(object)
}
\arguments{
\item{object}{object of class:\code{ConstFc}. }
}
\description{
getValues,ConstFc-method
}
