\name{HIV_R}
\alias{HIV_R}
\title{HIV R}
\description{a test function to be put somewhere else soon}
\usage{HIV_R(pars, V_0 = 50000, dV_0 = -200750, T_0 = 100)}
\arguments{
  \item{pars}{
}
  \item{V_0}{
}
  \item{dV_0}{
}
  \item{T_0}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





