\title{InFlux S4 class}
\description{no Description}
\name{InFlux-class}
\alias{InFlux-class}
\docType{class}
\section{Methods}{  \describe{
    \item{GeneralInFlux}{\code{signature(object = "InFlux")}: ... } \code{\link{GeneralInFlux,InFlux-method}}
	 }}
\section{Subclasses}{\describe{
	\code{\link{BoundInFlux-class}}\cr
	\code{\link{ConstInFlux-class}}\cr
}}
