\docType{methods}
\name{InFlux,InFlux-method}
\title{InFlux,InFlux-method}
\alias{InFlux,InFlux-method}
\usage{\S4method{InFlux}{InFlux}(object)}
\arguments{
\item{object}{: of class InFlux}
}
\description{This method handles the case that no actual conversion is necessary since
the argument is already of a subclass of InFlux }
\value{the unchanged argument}
