% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InFlux_by_PoolIndex-class}
\alias{InFlux_by_PoolIndex-class}
\title{S4 class for the influx to a single pool identified by the index}
\description{
S4 class for the influx to a single pool identified by the index
}
