% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InFlux_by_PoolIndex}
\alias{InFlux_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
InFlux_by_PoolIndex(func, destinationIndex)
}
\arguments{
\item{func}{see method arguments}

\item{destinationIndex}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
