% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InFlux_by_PoolName-class}
\alias{InFlux_by_PoolName-class}
\title{S4 class for the influx to a single pool identified by the name}
\description{
S4 class for the influx to a single pool identified by the name
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{InFlux_by_PoolName} in their signature:}{
      \itemize{
        \item \code{\link{by_PoolIndex,InFlux_by_PoolName,character,character-method}}
      }
  }
}
