% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFlux_by_PoolName.R
\docType{methods}
\name{InFlux_by_PoolName,function,character-method}
\alias{InFlux_by_PoolName,function,character-method}
\title{constructor from a PoolName (integer like) object and a function}
\usage{
\S4method{InFlux_by_PoolName}{`function`,character}(func, destinationName)
}
\arguments{
\item{func}{A function. The names of the formal aguemnts have to be a subset of the state variable names and the time symbol 
This allows subsequent automatic reordering of the state variables.
In the presence of a vector of stave variabl names the formulation can 
automatically be transformed to a function of a s tate VECTOR argument and #' time
constructor from an ordered pair of PoolName (integer like) objects}
}
\description{
constructor from a PoolName (integer like) object and a function
}
