% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFlux_by_PoolName.R
\name{InFlux_by_PoolName,function,character-method}
\alias{InFlux_by_PoolName,function,character-method}
\title{Constructor from an ordered pair of PoolName (string like) and function  objects}
\usage{
\S4method{InFlux_by_PoolName}{`function`,character}(func, destinationName)
}
\arguments{
\item{func}{object of class:\code{function}, A function. The names of the formal arguments have to be a subset of the state variable names and the time symbol 
This allows subsequent automatic reordering of the state variables.
In the presence of a vector of state-variable-names the formulation can 
automatically be transformed to a function of a s tate VECTOR argument and 
time}

\item{destinationName}{object of class:\code{character}, PoolName (string like) object and a function}
}
\description{
Constructor from an ordered pair of PoolName (string like) and function  objects
}
