% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InFluxes-class}
\alias{InFluxes-class}
\title{A virtual S4-class representing (different subclasses) of in-fluxes to the system}
\description{
A virtual S4-class representing (different subclasses) of in-fluxes to the system
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{InFluxes} in their signature:}{
      \itemize{
        \item \code{\link{InFluxes,InFluxes-method}}
      }
  }
}
