% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{InFluxes,function-method}
\alias{InFluxes,function-method}
\title{InFluxes,function-method}
\usage{
\S4method{InFluxes}{`function`}(object)
}
\arguments{
\item{object}{object of class:\code{function}. }
}
\description{
InFluxes,function-method
}
