% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,numeric-method}
\alias{InFluxes,numeric-method}
\title{automatic title}
\usage{
\S4method{InFluxes}{numeric}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
