% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,StateIndependentInFluxList_by_PoolIndex-method}
\alias{InFluxes,StateIndependentInFluxList_by_PoolIndex-method}
\title{automatic title}
\usage{
\S4method{InFluxes}{StateIndependentInFluxList_by_PoolIndex}(object, numberOfPools)
}
\arguments{
\item{object}{no manual documentation}

\item{numberOfPools}{no manual documentation
basically produces a vector valued function from a list of scalar functions}
}
\description{
automatic title
}
