% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,TimeMap-method}
\alias{InFluxes,TimeMap-method}
\title{automatic title}
\usage{
\S4method{InFluxes}{TimeMap}(object)
}
\arguments{
\item{object}{object of class:\code{TimeMap}, no manual documentation}
}
\description{
automatic title
}
