% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InFluxes}
\alias{InFluxes}
\title{A generic factory for subclasses of \linkS4class{InFluxes}}
\usage{
InFluxes(object, numberOfPools)
}
\arguments{
\item{object}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
The actual class of the returned object depends on the arguments 
provided
}
