% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InFluxList_by_PoolName-class}
\alias{InFluxList_by_PoolName-class}
\title{Class for a list of influxes indexed by the names of the target pools}
\description{
Class for a list of influxes indexed by the names of the target pools
}
