\docType{methods}
\name{InFluxSubClassInstance,InFlux-method}
\title{InFluxSubClassInstance,InFlux-method}
\alias{InFluxSubClassInstance,InFlux-method}
\usage{\S4method{InFluxSubClassInstance}{InFlux}(object)}
\arguments{
\item{object}{: of class InFlux}
}
\description{This method handles the case that no actual conversion is necessary since
the argument is already of a subclass of \link{InFlux-class} }
