\docType{methods}
\name{InFluxSubClassInstance,numeric-method}
\title{InFluxSubClassInstance,numeric-method}
\alias{InFluxSubClassInstance,numeric-method}
\usage{\S4method{InFluxSubClassInstance}{numeric}(object)}
\arguments{
\item{object}{: of class numeric}
}
\description{This method enables the model creating functions to handle constant input streams }
