\docType{methods}
\name{InFluxSubClassInstance,TimeMap-method}
\title{InFluxSubClassInstance,TimeMap-method}
\alias{InFluxSubClassInstance,TimeMap-method}
\usage{\S4method{InFluxSubClassInstance}{TimeMap}(object)}
\arguments{
\item{object}{: of class TimeMap}
}
\description{The method is used to ensure backward compatibility with the now deprecated
TimeMap class.
The resulting BoundInFlux is created by a call to
the constructor BoundInFlux(object) of that class.}
