\name{InFluxSubClassInstance}
\alias{InFluxSubClassInstance}
\usage{InFluxSubClassInstance(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{InFluxSubClassInstance,InFlux-method}}\cr
	\code{\link{InFluxSubClassInstance,TimeMap-method}}\cr
	\code{\link{InFluxSubClassInstance,numeric-method}}\cr}
\title{Generic constructor and converter }
\description{If the argument is already of a subclass of class InFlux
the function returns the unchanged object.
Otherwise, it creates an object of a subclass of InFlux.
Note that the actual class of the output depends on the argument.
For examples, please look at the methods of this function.}
