% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{initialize,ConstLinDecompOp-method}
\alias{initialize,ConstLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{initialize}{ConstLinDecompOp}(.Object, mat = matrix())
}
\arguments{
\item{.Object}{no manual documentation}

\item{mat}{no manual documentation}
}
\description{
automatic title
}
