% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompositionOperator_deprecated_in_1.2.R
\name{initialize,DecompositionOperator-method}
\alias{initialize,DecompositionOperator-method}
\title{automatic title}
\usage{
\S4method{initialize}{DecompositionOperator}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  map = function(t) {
     t
 },
  lag = 0
)
}
\arguments{
\item{.Object}{no manual documentation}

\item{starttime}{no manual documentation}

\item{endtime}{no manual documentation}

\item{map}{no manual documentation}

\item{lag}{no manual documentation}
}
\description{
automatic title
}
