% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{initialize,MCSim-method}
\alias{initialize,MCSim-method}
\title{Automatic description: initialize,MCSim-method}
\usage{
\S4method{initialize}{MCSim}(.Object, model = new(Class = "NlModel"), tasklist = list())
}
\arguments{
\item{.Object}{object of class:\code{MCSim}. : object of class:\code{MCSim}, no manual documentation}

\item{model}{found automatically by inspection, not documented yet}

\item{tasklist}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,MCSim-method
}
