% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{initialize,Model-method}
\alias{initialize,Model-method}
\title{automatic title}
\usage{
\S4method{initialize}{Model}(
  .Object,
  times = c(0, 1),
  mat = ConstLinDecompOp(matrix(nrow = 1, ncol = 1, 0)),
  initialValues = numeric(),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{Model}, : no manual documentation}

\item{times}{: no manual documentation}

\item{mat}{: no manual documentation}

\item{initialValues}{: no manual documentation}

\item{inputFluxes}{: no manual documentation}

\item{solverfunc}{: no manual documentation}

\item{pass}{: no manual documentation}
}
\description{
automatic title
}
