% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{initialize,Model-method}
\alias{initialize,Model-method}
\title{Internal method to supervise creation of objects of this class}
\usage{
\S4method{initialize}{Model}(
  .Object,
  times = c(0, 1),
  mat = ConstLinDecompOp(matrix(nrow = 1, ncol = 1, 0)),
  initialValues = numeric(),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{Model}, no manual documentation}

\item{times}{no manual documentation}

\item{mat}{no manual documentation}

\item{initialValues}{no manual documentation}

\item{inputFluxes}{no manual documentation}

\item{solverfunc}{no manual documentation}

\item{pass}{no manual documentation}
}
\description{
It is usually not necessary for user code to call this method.
It's purpose is to enforce some sanity checks since it gets automatically
called by \code{new} whenever an object of this class is created
}
