% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_14.R
\name{initialize,Model_14-method}
\alias{initialize,Model_14-method}
\title{Internal method to supervise creation of objects of this class}
\usage{
\S4method{initialize}{Model_14}(
  .Object,
  times = c(0, 1),
  mat = ConstLinDecompOp(matrix(nrow = 1, ncol = 1, 0)),
  initialValues = numeric(),
  initialValF = ConstFc(values = c(0), format = "Delta14C"),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1))
    }, 0, 1),
  c14Fraction = BoundFc(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) }, 0,
    1),
  c14DecayRate = 0,
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{Model_14}, no manual documentation}

\item{times}{no manual documentation}

\item{mat}{no manual documentation}

\item{initialValues}{no manual documentation}

\item{initialValF}{no manual documentation}

\item{inputFluxes}{no manual documentation}

\item{c14Fraction}{no manual documentation}

\item{c14DecayRate}{no manual documentation}

\item{solverfunc}{no manual documentation}

\item{pass}{no manual documentation}
}
\description{
It is usually not necessary for user code to call this method.
It's purpose is to enforce some sanity checks since it gets automatically
called by \code{new} whenever an object of this class is created
}
