% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\name{initialize,Model_by_PoolNames-method}
\alias{initialize,Model_by_PoolNames-method}
\title{initialize,Model_by_PoolNames-method}
\usage{
\S4method{initialize}{Model_by_PoolNames}(
  .Object,
  times,
  mat,
  initialValues,
  inputFluxes,
  solverfunc,
  pass,
  timeSymbol
)
}
\arguments{
\item{.Object}{object of class:\code{Model_by_PoolNames}. }

\item{times}{found automatically by inspection, not documented yet}

\item{mat}{found automatically by inspection, not documented yet}

\item{initialValues}{found automatically by inspection, not documented yet}

\item{inputFluxes}{found automatically by inspection, not documented yet}

\item{solverfunc}{found automatically by inspection, not documented yet}

\item{pass}{found automatically by inspection, not documented yet}

\item{timeSymbol}{found automatically by inspection, not documented yet}
}
\description{
initialize,Model_by_PoolNames-method
}
