% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{initialize,NlModel-method}
\alias{initialize,NlModel-method}
\title{Automatic description: initialize,NlModel-method}
\usage{
\S4method{initialize}{NlModel}(
  .Object,
  times = c(0, 1),
  DepComp = new(Class = "TransportDecompositionOperator", 0, 1, function(t) {    
    return(matrix(nrow = 1, ncol = 1, 0)) }, function(t) {     return(matrix(nrow = 1,
    ncol = 1, 0)) }),
  initialValues = numeric(),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{NlModel}. : object of class:\code{NlModel}, no manual documentation}

\item{times}{found automatically by inspection, not documented yet}

\item{DepComp}{found automatically by inspection, not documented yet}

\item{initialValues}{found automatically by inspection, not documented yet}

\item{inputFluxes}{found automatically by inspection, not documented yet}

\item{solverfunc}{found automatically by inspection, not documented yet}

\item{pass}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,NlModel-method
}
