% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{initialize,TransportDecompositionOperator-method}
\alias{initialize,TransportDecompositionOperator-method}
\title{Automatic description: initialize,TransportDecompositionOperator-method}
\usage{
\S4method{initialize}{TransportDecompositionOperator}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  numberOfPools = 1,
  alpha = list(),
  f = function(t, O) {     t },
  lag = 0
)
}
\arguments{
\item{.Object}{object of class:\code{TransportDecompositionOperator}. : object of class:\code{TransportDecompositionOperator}, no
manual documentation}

\item{starttime}{found automatically by inspection, not documented yet}

\item{endtime}{found automatically by inspection, not documented yet}

\item{numberOfPools}{found automatically by inspection, not documented yet}

\item{alpha}{found automatically by inspection, not documented yet}

\item{f}{found automatically by inspection, not documented yet}

\item{lag}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,TransportDecompositionOperator-method
}
