% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{initialize,UnBoundLinDecompOp-method}
\alias{initialize,UnBoundLinDecompOp-method}
\title{initialize,UnBoundLinDecompOp-method}
\usage{
\S4method{initialize}{UnBoundLinDecompOp}(.Object, matFunc = function() {
})
}
\arguments{
\item{.Object}{object of class:\code{UnBoundLinDecompOp}. }

\item{matFunc}{found automatically by inspection, not documented yet}
}
\description{
initialize,UnBoundLinDecompOp-method
}
