\name{initialize_method__LinearDecompositionOperator}
\alias{initialize_method__LinearDecompositionOperator}
\title{initialize called by (new)}
\description{This is the internal constructor for objects of this class
It is called by statements of the form new("..) 
which in turn is called by all other constructors 
which may have more convienient interfaces}
\arguments{
  \item{.Object}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{map}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





