\name{InputFlux-class}
\Rdversion{1.1}
\docType{class}
\alias{InputFlux-class}
\alias{InputFlux}
\alias{initialize,InputFlux-method}

\title{decomposition operator }
\description{#######################################################}
\section{Objects from the Class}{Objects can be created by calls of the form \code{new(InputFlux ...)}}
\section{Slots}{
  \describe{
    \item{\code{starttime}:}{Object of class \code{"numeric"}}
    \item{\code{endtime}:}{Object of class \code{"numeric"}}
    \item{\code{map}:}{Object of class \code{"function"}}
    \item{\code{lag}:}{Object of class \code{"numeric"}}
  }
}
\section{Extends}{
Class \code{"\linkS4class{TimeMap}"}, directly.
}
\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "InputFlux")}: ... }
	 }
}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





