\name{InputFlux_method__TimeMap}
\alias{InputFlux_method__TimeMap}
\title{create a TemporaryInputFlux from a TimeMap object}
\description{The method is used to ensure backward compatibility with the now deprecated
TimeMap class.
The resulting TemporaryInputFlux is created by a call to
the constructor TemporaryInputFlux(object) of that class.}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





