% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InternalFlux_by_PoolIndex}
\alias{InternalFlux_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
InternalFlux_by_PoolIndex(func, sourceIndex, destinationIndex, src_to_dest)
}
\arguments{
\item{func}{see method arguments}

\item{sourceIndex}{see method arguments}

\item{destinationIndex}{see method arguments}

\item{src_to_dest}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{InternalFlux_by_PoolIndex,function,numeric,numeric,missing-method}}
      }
}
