% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InternalFlux_by_PoolName-class}
\alias{InternalFlux_by_PoolName-class}
\title{S4-class for a single internal flux wiht source and destination pools specified by name}
\description{
S4-class for a single internal flux wiht source and destination pools specified by name
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{InternalFlux_by_PoolName} in their signature:}{
      \itemize{
        \item \code{\link{as.numeric,InternalFlux_by_PoolName-method}}
        \item \code{\link{by_PoolIndex,InternalFlux_by_PoolName,character,character-method}}
      }
  }
}
