% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InternalFlux_by_PoolName}
\alias{InternalFlux_by_PoolName}
\title{Generic constructor for the class with the same name}
\usage{
InternalFlux_by_PoolName(func, sourceName, destinationName, src_to_dest)
}
\arguments{
\item{func}{see method arguments}

\item{sourceName}{see method arguments}

\item{destinationName}{see method arguments}

\item{src_to_dest}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{InternalFlux_by_PoolName,function,character,character,missing-method}}
        \item \code{\link{InternalFlux_by_PoolName,function,missing,missing,character-method}}
      }
}
