% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InternalFluxList_by_PoolIndex-class}
\alias{InternalFluxList_by_PoolIndex-class}
\title{S4-class for a list of internal fluxes with source and destination pool inidices}
\description{
S4-class for a list of internal fluxes with source and destination pool inidices
}
\section{S4-methods}{
  \subsection{S4-methods with superclasses of class \code{InternalFluxList_by_PoolIndex} in their signature:}{
    \subsection{superclass list}{
      \itemize{
        \item \code{\link{ConstantInFluxList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolName,list-method}}
        \item \code{\link{ConstantOutFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantOutFluxRateList_by_PoolName,list-method}}
        \item \code{\link{GeneralDecompOp,list-method}}
        \item \code{\link{InFluxes,list-method}}
        \item \code{\link{InFluxList_by_PoolIndex,list-method}}
        \item \code{\link{InFluxList_by_PoolName,list-method}}
        \item \code{\link{InternalFluxList_by_PoolIndex,list-method}}
        \item \code{\link{InternalFluxList_by_PoolName,list-method}}
        \item \code{\link{OutFluxList_by_PoolIndex,list-method}}
        \item \code{\link{OutFluxList_by_PoolName,list-method}}
        \item \code{\link{StateIndependentInFluxList_by_PoolIndex,list-method}}
        \item \code{\link{TimeMap,list,missing,missing,missing,missing-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,list-method}}
      }
    }
    \subsection{superclass vector}{
      \itemize{
        \item \code{\link{UnBoundNonLinDecompOp,missing,vector,vector,numeric,ANY,ANY,ANY-method}}
      }
    }
  }
}
