% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFluxList_by_PoolIndex.R
\name{InternalFluxList_by_PoolIndex,list-method}
\alias{InternalFluxList_by_PoolIndex,list-method}
\title{constructor from a normal list}
\usage{
\S4method{InternalFluxList_by_PoolIndex}{list}(object)
}
\description{
after checking the elememts
}
