% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InternalFluxList_by_PoolIndex}
\alias{InternalFluxList_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
InternalFluxList_by_PoolIndex(object)
}
\description{
Generic constructor for the class with the same name
}
