% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFluxList_by_PoolName.R
\docType{methods}
\name{InternalFluxList_by_PoolName,list-method}
\alias{InternalFluxList_by_PoolName,list-method}
\title{constructor from a normal list}
\usage{
\S4method{InternalFluxList_by_PoolName}{list}(l)
}
\description{
after checking the elememts
}
