% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InternalFluxList_by_PoolName}
\alias{InternalFluxList_by_PoolName}
\title{Generic constructor for the class with the same name}
\usage{
InternalFluxList_by_PoolName(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
