% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.compartmental.R
\name{is.compartmental}
\alias{is.compartmental}
\title{Test whether a square matrix is compartmental}
\usage{
is.compartmental(B)
}
\arguments{
\item{B}{a square matrix}
}
\value{
Logical. TRUE if B is a strictly compartmental matrix
}
\description{
`is.compartmental` test three conditions on a square matrix. Condition 1: all diagonal elements are negative.
Condition 2: all off-diagonal elements are non-negative. Condition 3: Column sums are non-positive.
Notice that this is a test on strictly compartmental matrices, i.e. diagonal elements cannot be equal to zero.
}
\examples{
B<-matrix(c(-3,2,1,1,-2,1,0,0,-1),3,3)
is.compartmental(B)
}
