\name{is.negative}
\alias{is.negative}
\title{is negative}
\description{the function returns True if the argumente is negative}
\usage{is.negative(number)}
\arguments{
  \item{number}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





