\name{lambda}
\alias{lambda}
\docType{data}
\title{
The assumed decay constant for $^{14}C$
}
\description{
The assumed decay constant for $^{14}C$ in $year^{-1}$
}
\usage{data(lambda)}
\format{
  The format is:
 num 0.000124
}
\examples{
data(lambda)
## maybe str(lambda) ; plot(lambda) ...
}
\keyword{datasets}
