\name{LinearDecompositionOperator}
\alias{LinearDecompositionOperator}
\title{generic constructor}
\description{create a LinearDecompositonOperator from different sources.
Have a look at the methods to see what kind of input is supported. }
\usage{LinearDecompositionOperator(map, starttime, endtime, lag = 0)}\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





