\name{LinearDecompositionOperator_method__ConstantDecompositionOperator}
\alias{LinearDecompositionOperator_method__ConstantDecompositionOperator}
\title{convert a ConstantDecompositionOperator to a LinearDecompositionOperator}
\description{The method creates a LinearDecompositionOperator consisting of a constant time dependent function 
and the limits of its domain (starttime and endtime) set to -Inf and Inf respectively}
\arguments{
  \item{map}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





