\name{LinearDecompositionOperator_method__function_missing_missing_numeric}
\alias{LinearDecompositionOperator_method__function_missing_missing_numeric}
\title{a constructor for a single function without limits  }
\description{This method creates a LinearDecompositionOperator from a timedependent function and a lag only}
\arguments{
  \item{map}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





