% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{mat_from_integer_flux_lists}
\alias{mat_from_integer_flux_lists}
\title{helper function}
\usage{
mat_from_integer_flux_lists(internal_flux_rates, out_flux_rates,
  numberOfPools)
}
\description{
helper function
}
\keyword{internal}
