\name{MeanAge3}
\alias{MeanAge3}
\title{MeanAge3}

\usage{MeanAge3(IdotT, OdotLin, sol, times)}
\arguments{
  \item{IdotT}{
}
  \item{OdotLin}{
}
  \item{sol}{
}
  \item{times}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





