% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing-method}
\alias{Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing-method}
\title{Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing-method}
\usage{
\S4method{Model_by_PoolNames}{missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing}(times, mat, initialValues, inputFluxes)
}
\arguments{
\item{times}{object of class:\code{numeric}, no manual documentation}

\item{mat}{object of class:\code{UnBoundNonLinDecompOp_by_PoolNames}, no manual documentation}

\item{initialValues}{object of class:\code{numeric}, no manual documentation}

\item{inputFluxes}{object of class:\code{InFluxList_by_PoolName}, no manual documentation}
}
\description{
Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing-method
}
