% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{Model_by_PoolNames,SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing,missing-method}
\alias{Model_by_PoolNames,SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing,missing-method}
\title{Model_by_PoolNames,SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing,missing-method}
\usage{
\S4method{Model_by_PoolNames}{SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing,missing}(smod, times, initialValues)
}
\arguments{
\item{smod}{object of class:\code{SymbolicModel_by_PoolNames}, no manual documentation}

\item{times}{object of class:\code{numeric}, no manual documentation}

\item{initialValues}{object of class:\code{numeric}, no manual documentation}
}
\description{
Model_by_PoolNames,SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing,missing-method
}
