% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{Model_by_PoolNames}
\alias{Model_by_PoolNames}
\title{Constructor for \code{\linkS4class{Model_by_PoolNames}}}
\usage{
Model_by_PoolNames(
  smod,
  times,
  mat,
  initialValues,
  inputFluxes,
  internal_fluxes,
  out_fluxes,
  timeSymbol,
  pass,
  solverfunc
)
}
\arguments{
\item{smod}{see method arguments}

\item{times}{see method arguments}

\item{mat}{see method arguments}

\item{initialValues}{see method arguments}

\item{inputFluxes}{see method arguments}

\item{internal_fluxes}{see method arguments}

\item{out_fluxes}{see method arguments}

\item{timeSymbol}{see method arguments}

\item{pass}{see method arguments}

\item{solverfunc}{see method arguments}
}
\description{
Constructor for \code{\linkS4class{Model_by_PoolNames}}
}
