\name{Models}
\alias{Models}
\alias{creating Models}
\title{An overview about available shortcuts for model construction}
\description{
    \enumerate{
    \item \code{\link{GeneralModel}}
    \item \code{\link{GeneralModel_14}}
    \item \code{\link{ParallelModel}} 
    \item \code{\link{ParallelModel}}
    \item \code{\link{ICBMModel}}
    \item \code{\link{OnepModel}}
    \item \code{\link{OnepModel14}}
    \item \code{\link{ParallelModel}}
    \item \code{\link{RothCModel}}
    \item \code{\link{ThreepFeedbackModel}}
    \item \code{\link{ThreepFeedbackModel14}}
    \item \code{\link{ThreepParallelModel}}
    \item \code{\link{ThreepParallelModel14}}
    \item \code{\link{ThreepSeriesModel}}
    \item \code{\link{ThreepSeriesModel14}}
    \item \code{\link{TwopFeedbackModel}}
    \item \code{\link{TwopFeedbackModel14}}
    \item \code{\link{TwopParallelModel}}
    \item \code{\link{TwopParallelModel14}}
    \item \code{\link{TwopSeriesModel}}
    \item \code{\link{TwopSeriesModel14}}
    }
}
\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}
