% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{NlModel-class}
\alias{NlModel-class}
\title{NlModel}
\description{
serves as a fence to the interface of SoilR functions. So that later
implementations can differ
}
\section{Methods}{
 Exported methods directly defined for class NlModel:

\describe{ \item{$}{\code{signature(x = "NlModel")}: ... }
\code{\link{$,NlModel-method}} \item{[}{\code{signature(x = "NlModel", i =
"character", j = "ANY", drop = "ANY")}: ... }
\code{\link{[,NlModel,character,ANY,ANY-method}}
\item{getC}{\code{signature(object = "NlModel")}: ... }
\code{\link{getC,NlModel-method}} \item{getDecompOp}{\code{signature(object
= "NlModel")}: ... } \code{\link{getDecompOp,NlModel-method}}
\item{getReleaseFlux}{\code{signature(object = "NlModel")}: ... }
\code{\link{getReleaseFlux,NlModel-method}}
\item{getTimes}{\code{signature(object = "NlModel")}: ... }
\code{\link{getTimes,NlModel-method}} }
}

