% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutFlux.R
\name{OutFlux,function-method}
\alias{OutFlux,function-method}
\title{OutFlux,function-method}
\usage{
\S4method{OutFlux}{`function`}(map, ...)
}
\arguments{
\item{map}{object of class:\code{function}. }

\item{...}{}
}
\description{
OutFlux,function-method
}
