% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{OutFlux}
\alias{OutFlux}
\title{Generic constructor for the class with the same name}
\usage{
OutFlux(map, ...)
}
\arguments{
\item{map}{see method arguments}

\item{...}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
