% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{OutFlux_by_PoolIndex-class}
\alias{OutFlux_by_PoolIndex-class}
\title{S4 class for a single out-flux with source pool index}
\description{
S4 class for a single out-flux with source pool index
}
