% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutFlux_by_PoolIndex.R
\name{OutFlux_by_PoolIndex,function,numeric-method}
\alias{OutFlux_by_PoolIndex,function,numeric-method}
\title{constructor from a PoolIndex (integer like) objects and a function with vector argument}
\usage{
\S4method{OutFlux_by_PoolIndex}{`function`,numeric}(func, sourceIndex)
}
\arguments{
\item{func}{A function f(X,t) where X is a vector of the state varaibles. 
This form is required internally by the solvers and supported for backward compatibility with earlier versions of SoilR.
Note that the function func given in this form can not be tranformed to a different ordering of state variables, since the location of a state variable in the vector argument depends on a specific orderand will be 'hardcoded' into your function. 
See \code{\link{OutFlux_by_PoolName}} for the new, more powerful interface which allows subsequent reordering of the state variables by using the names of the state variables as formal arguments for \code{func}. In this case SoilR can infer (and later adapt) the
vector argument form needed for the solvers.
constructor from an ordered pair of PoolIndex (integer like) objects}
}
\description{
constructor from a PoolIndex (integer like) objects and a function with vector argument
}
