% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{OutFlux_by_PoolIndex}
\alias{OutFlux_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
OutFlux_by_PoolIndex(func, sourceIndex)
}
\arguments{
\item{func}{see method arguments}

\item{sourceIndex}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{OutFlux_by_PoolIndex,function,numeric-method}}
      }
}
