% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{OutFlux_by_PoolName-class}
\alias{OutFlux_by_PoolName-class}
\title{S4 class for a single out-flux with source pool name}
\description{
S4 class for a single out-flux with source pool name
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{OutFlux_by_PoolName} in their signature:}{
      \itemize{
        \item \code{\link{by_PoolIndex,OutFlux_by_PoolName,character,character-method}}
      }
  }
}
