% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutFlux_by_PoolName.R
\name{OutFlux_by_PoolName,function,character-method}
\alias{OutFlux_by_PoolName,function,character-method}
\title{constructor from a PoolName (integer like) object and a function}
\usage{
\S4method{OutFlux_by_PoolName}{function,character}(func, sourceName)
}
\arguments{
\item{func}{A function. The names of the formal arguments have to be a subset of the state variable names and the time symbol}

\item{sourceName}{the name of the source pool
This allows subsequent automatic reordering of the state variables.
In the presence of a vector of stave variable names the formulation can 
automatically be transformed to a function of a state VECTOR argument and #' time
constructor from an ordered pair of PoolName (integer like) objects}
}
\description{
constructor from a PoolName (integer like) object and a function
}
