% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{OutFlux_by_PoolName}
\alias{OutFlux_by_PoolName}
\title{Generic constructor for the class with the same name}
\usage{
OutFlux_by_PoolName(func, sourceName)
}
\arguments{
\item{func}{function with output flux}

\item{sourceName}{name of the source pool}
}
\description{
Generic constructor for the class with the same name
}
