% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_baumer_rice_1988.R
\name{p_baumer_rice_1988}
\alias{p_baumer_rice_1988}
\title{Calculation of the porosity.}
\usage{
p_baumer_rice_1988(bulk_density_g_cm3, organic_carbon_percentage)
}
\arguments{
\item{bulk_density_g_cm3}{Bulk density (g/cm3).}

\item{organic_carbon_percentage}{Organic carbon percentage.}
}
\value{
Porosity
}
\description{
The particle density is adjusted to take into account Soil
Organic Matter within this function.

The function coded here is: porosity = 1 - BD/rho with BD the bulk density
(g/cm3) and rho the particle density (g/cm3).
}
\examples{
p_baumer_rice_1988(bulk_density_g_cm3 = 1.5,
                   organic_carbon_percentage = 2)
}
