% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dalgliesh_foale_1998.R
\name{p_dalgliesh_foale_1998}
\alias{p_dalgliesh_foale_1998}
\title{Calculation of the saturation characteristic humidity.}
\usage{
p_dalgliesh_foale_1998(soil_texture_class_usda, porosity)
}
\arguments{
\item{soil_texture_class_usda}{Soil textural class as defined in the USDA
texture triangle.
\itemize{
\item C: clay
\item SIC: silty clay
\item SC: sandy clay
\item CL: clay loam
\item SICL: silty clay loam
\item SCL: sandy clay loam
\item L: loam
\item SIL: silt loam
\item SL: sandy loam
\item SI: silt
\item LS: loamy sand
\item S: sand
}}

\item{porosity}{Porosity.}
}
\value{
Humidity at saturation expressed in cm3/cm3.
}
\description{
Calculation of the saturation characteristic humidity.
}
\examples{
p_dalgliesh_foale_1998(soil_texture_class_usda = "S",
                       porosity = 0.1)
}
