% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_field_capacity.R
\name{p_field_capacity}
\alias{p_field_capacity}
\title{Calculation of the field capacity}
\usage{
p_field_capacity(sand, organic_carbon, permanent_wilting_point)
}
\arguments{
\item{sand}{Percentage sand}

\item{organic_carbon}{Percentage organic carbon}

\item{permanent_wilting_point}{Permanent wilting point (cm3/cm3)}
}
\value{
Field capacity (cm3/cm3)
}
\description{
This pedotransfert function comes from Rawls et al (1982),
Estimation of Soil Water Properties
}
\examples{
p_field_capacity(sand = 20,
                 organic_carbon = 2,
                 permanent_wilting_point = 0.25)
}
