% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_permanent_wilting_point.R
\name{p_permanent_wilting_point}
\alias{p_permanent_wilting_point}
\title{Calculation of the permanent wilting point}
\usage{
p_permanent_wilting_point(clay, organic_carbon)
}
\arguments{
\item{clay}{Percentage of clay}

\item{organic_carbon}{Percentage organic carbon}
}
\value{
Permanent wilting point (cm3/cm3)
}
\description{
This pedotransfert function comes from Rawls et al (1982),
Estimation of Soil Water Properties
}
\examples{
p_permanent_wilting_point(clay = 20,
                          organic_carbon = 2)
}
