% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rawls_1982.R
\name{p_rawls_1982}
\alias{p_rawls_1982}
\title{Calculation of the saturation, field capacity and permanent wilting point
characteristic humidities.}
\usage{
p_rawls_1982(
  clay_percentage,
  sand_percentage,
  organic_carbon_percentage,
  bulk_density_g_cm3
)
}
\arguments{
\item{clay_percentage}{Percentage clay.}

\item{sand_percentage}{Percentage sand.}

\item{organic_carbon_percentage}{Percentage organic carbon.}

\item{bulk_density_g_cm3}{Bulk density (g/cm3)}
}
\value{
List with humidity at permanent_wilting_point, field_capacity and
at saturation expressed in cm3/cm3.
}
\description{
This pedotransfert function comes from Rawls et al (1982),
Estimation of Soil Water Properties. Parameters used come from table 3 for
a metric potential (bars) of -0.04, -0.33 and -15.0 for humidity at
saturation, field capacity and permanent wilting point, respectively.
For each of these metric potential (bars), the equation with the highest
correlation coefficient were chosen.
}
\examples{
p_rawls_1982(clay_percentage = 20,
             sand_percentage = 20,
             organic_carbon_percentage = 2,
             bulk_density_g_cm3 = 1.5)
}
