% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_root_growth_factor.R
\name{p_root_growth_factor}
\alias{p_root_growth_factor}
\title{Calculation of root growth factor}
\usage{
p_root_growth_factor(center_of_layer)
}
\arguments{
\item{center_of_layer}{Center of the layer (cm)}
}
\value{
root growth factor (-)
}
\description{
The method to calculated root growth factor comes from the DSSAT
documentation (DSSAT45 Volume 2)
}
\examples{
p_root_growth_factor(center_of_layer = 5)
p_root_growth_factor(center_of_layer = 50)
}
