% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_saturation.R
\name{p_saturation}
\alias{p_saturation}
\title{Calculation of the saturation}
\usage{
p_saturation(sand, organic_carbon, bulk_density, field_capacity, wilting_point)
}
\arguments{
\item{sand}{Percentage sand}

\item{organic_carbon}{Percentage organic carbon}

\item{bulk_density}{Bulk density (g/cm3)}

\item{field_capacity}{Field capacity (cm3/cm3)}

\item{wilting_point}{Wilting point (cm3/cm3)}
}
\value{
Saturation (cm3/cm3)
}
\description{
This pedotransfert function comes from Rawls et al (1982),
Estimation of Soil Water Properties
}
\examples{
p_saturation(sand = 20,
             organic_carbon = 2,
             bulk_density = 1.5,
             field_capacity = 0.5,
             wilting_point = 0.25)
}
