% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_parameters.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{parameters}
\format{
A data frame with  rows and  variables:
\describe{
\item{ organic_carbon_to_organic_matter_conversion_factor }{  numeric }
}
}
\source{
Source
}
\usage{
parameters
}
\description{
List of the parameters used in the package..
}
\keyword{datasets}
