% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{plot,MCSim-method}
\alias{plot,MCSim-method}
\title{plot,MCSim-method}
\usage{
\S4method{plot}{MCSim}(x, y, ...)
}
\arguments{
\item{x}{object of class:\code{MCSim}. }

\item{y}{}

\item{...}{}
}
\description{
plot,MCSim-method
}
