% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{plot,NlModel-method}
\alias{plot,NlModel-method}
\title{automatic title}
\usage{
\S4method{plot}{NlModel}(x)
}
\arguments{
\item{x}{object of class:\code{NlModel}, : no manual documentation}
}
\description{
automatic title
}
