% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plot,Model-method}
\alias{plot,Model-method}
\title{automatic title}
\usage{
\S4method{plot}{Model}(x)
}
\arguments{
\item{x}{object of class:\code{Model}, no manual documentation}
}
\description{
automatic title
}
