% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{plot,Model_by_PoolNames-method}
\alias{plot,Model_by_PoolNames-method}
\title{Plot the graph of pool connections}
\usage{
\S4method{plot}{Model_by_PoolNames}(x)
}
\arguments{
\item{x}{The model (run) the results of which are plotted}
}
\description{
Plot the graph of pool connections
}
