% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{plot,NlModel-method}
\alias{plot,NlModel-method}
\title{plot,NlModel-method}
\usage{
\S4method{plot}{NlModel}(x)
}
\arguments{
\item{x}{object of class:\code{NlModel}. }
}
\description{
plot,NlModel-method
}
