% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plotPoolGraph,Model-method}
\alias{plotPoolGraph,Model-method}
\title{Plot the graph of pool connections}
\usage{
\S4method{plotPoolGraph}{Model}(x)
}
\arguments{
\item{x}{object of class:\code{Model}, The model (run) the of connection graph of which is plotted}
}
\description{
Plot the graph of pool connections
}
