% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{plotPoolGraph,Model_by_PoolNames-method}
\alias{plotPoolGraph,Model_by_PoolNames-method}
\title{Plot the graph of pool connections}
\usage{
\S4method{plotPoolGraph}{Model_by_PoolNames}(x)
}
\arguments{
\item{x}{object of class:\code{Model_by_PoolNames}, The modelrun the connection graph of which is plotted}
}
\description{
Plot the graph of pool connections
}
