% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{plotPoolGraph}
\alias{plotPoolGraph}
\title{Generic plotter}
\usage{
plotPoolGraph(x)
}
\arguments{
\item{x}{An argument containing sufficient information about the connections between the pools as well as from and to the exterior.}
}
\description{
Generic plotter
}
