% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection.R
\name{PoolConnection,missing,missing,character-method}
\alias{PoolConnection,missing,missing,character-method}
\title{constructor from strings of the form '1_to_2'}
\usage{
\S4method{PoolConnection}{missing,missing,character}(src_to_dest)
}
\arguments{
\item{src_to_dest}{object of class:\code{character}. }
}
\description{
constructor from strings of the form '1_to_2'
}
