% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection_by_PoolIndex.R
\name{PoolConnection_by_PoolIndex,ANY,ANY,missing-method}
\alias{PoolConnection_by_PoolIndex,ANY,ANY,missing-method}
\title{constructor from an ordered pair of PoolId objects}
\usage{
\S4method{PoolConnection_by_PoolIndex}{ANY,ANY,missing}(source, destination)
}
\arguments{
\item{source}{no manual documentation}

\item{destination}{no manual documentation}
}
\description{
constructor from an ordered pair of PoolId objects
}
