% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection_by_PoolIndex.R
\docType{class}
\name{PoolConnection_by_PoolIndex-class}
\alias{PoolConnection_by_PoolIndex-class}
\title{Objects that have a source and a destination described by integer like objects  ( of class PoolIndex)}
\description{
Examples are internal Fluxes and Fluxrates
Their 'topologic' part and many related sanity checks are implemented here rather than 
in every function that uses fluxes or rates
The methods are also essential for the translation from (internal) 
flux lists
to the respective parts of compartmental matrices and back
}
