% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection_by_PoolIndex.R
\name{PoolConnection_by_PoolIndex,missing,missing,character-method}
\alias{PoolConnection_by_PoolIndex,missing,missing,character-method}
\title{constructor from strings of the form '1_to_2'}
\usage{
\S4method{PoolConnection_by_PoolIndex}{missing,missing,character}(src_to_dest)
}
\arguments{
\item{src_to_dest}{object of class:\code{character}, no manual documentation}
}
\description{
constructor from strings of the form '1_to_2'
}
