% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{PoolConnection_by_PoolIndex}
\alias{PoolConnection_by_PoolIndex}
\title{PoolConnection_by_PoolIndex}
\usage{
PoolConnection_by_PoolIndex(source, destination, src_to_dest)
}
\arguments{
\item{source}{see method arguments}

\item{destination}{see method arguments}

\item{src_to_dest}{see method arguments}
}
\description{
PoolConnection_by_PoolIndex
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{PoolConnection_by_PoolIndex,ANY,ANY,missing-method}}
  \item \code{\link{PoolConnection_by_PoolIndex,missing,missing,character-method}}
  \item \code{\link{PoolConnection_by_PoolIndex,ANY,ANY,missing-method}}
  \item \code{\link{PoolConnection_by_PoolIndex,missing,missing,character-method}}
}
}

