% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection_by_PoolName.R
\docType{methods}
\name{PoolConnection_by_PoolName,ANY,ANY,missing-method}
\alias{PoolConnection_by_PoolName,ANY,ANY,missing-method}
\title{constructor from an ordered pair of PoolName objects}
\usage{
\S4method{PoolConnection_by_PoolName}{ANY,ANY,missing}(source, destination)
}
\description{
constructor from an ordered pair of PoolName objects
}
