% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{PoolConnection_by_PoolName}
\alias{PoolConnection_by_PoolName}
\title{S4 generic: PoolConnection_by_PoolName}
\usage{
PoolConnection_by_PoolName(source, destination, src_to_dest)
}
\arguments{
\item{source}{see method arguments}

\item{destination}{see method arguments}

\item{src_to_dest}{see method arguments}
}
\description{
S4 generic: PoolConnection_by_PoolName
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{PoolConnection_by_PoolName,ANY,ANY,missing-method}}
}
}

