% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{PoolId-class}
\alias{PoolId-class}
\title{common class for pool ids}
\description{
examples for ids are index or name
}
\section{S4-subclasses}{
      \itemize{
        \item \code{\linkS4class{PoolIndex}}
        \item \code{\linkS4class{PoolName}}
      }
}
