% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolIndex.R
\name{PoolIndex,character-method}
\alias{PoolIndex,character-method}
\title{construct from number string like '1' or '3'}
\usage{
\S4method{PoolIndex}{character}(id)
}
\arguments{
\item{id}{object of class:\code{character}, no manual documentation}
}
\description{
construct from number string like '1' or '3'
}
