% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolName.R
\name{PoolIndex,PoolName-method}
\alias{PoolIndex,PoolName-method}
\title{convert to number like object}
\usage{
\S4method{PoolIndex}{PoolName}(id, poolNames)
}
\arguments{
\item{id}{object of class:\code{PoolName}, no manual documentation}

\item{poolNames}{no manual documentation}
}
\description{
convert to number like object
}
