% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolIndex.R
\name{PoolIndex,numeric-method}
\alias{PoolIndex,numeric-method}
\title{construct from number}
\usage{
\S4method{PoolIndex}{numeric}(id)
}
\arguments{
\item{id}{object of class:\code{numeric}, no manual documentation}
}
\description{
construct from number
}
