% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolIndex.R
\name{PoolIndex,PoolIndex-method}
\alias{PoolIndex,PoolIndex-method}
\title{pass throug constructor fron an object of the same class}
\usage{
\S4method{PoolIndex}{PoolIndex}(id, poolNames)
}
\arguments{
\item{id}{object of class:\code{PoolIndex}. }

\item{poolNames}{found automatically by inspection, not documented yet}
}
\description{
This is here to be able to call PoolIndex on a PoolIndex ojbect without
having to chech before if it is necessary.
the unnecessary poolNames argument will be ignored
}
