% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolName.R
\docType{methods}
\name{PoolName,character-method}
\alias{PoolName,character-method}
\title{construct from string with checks}
\usage{
\S4method{PoolName}{character}(id)
}
\description{
construct from string with checks
}
