% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolIndex.R
\name{PoolName,PoolIndex-method}
\alias{PoolName,PoolIndex-method}
\title{convert to string like object}
\usage{
\S4method{PoolName}{PoolIndex}(id, poolNames)
}
\arguments{
\item{id}{pool id}

\item{poolNames}{name of pools}
}
\description{
convert to string like object
}
