% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolName.R
\name{PoolName,PoolName-method}
\alias{PoolName,PoolName-method}
\title{pass through constructor fron an object of the same class}
\usage{
\S4method{PoolName}{PoolName}(id, poolNames)
}
\arguments{
\item{id}{object of class:\code{PoolName}, no manual documentation}

\item{poolNames}{no manual documentation}
}
\description{
This is here to be able to call PoolName on a PoolName object without
having to test before if we have to.
}
