% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{PoolName}
\alias{PoolName}
\title{S4 generic: PoolName}
\usage{
PoolName(id, ...)
}
\arguments{
\item{id}{see method arguments}

\item{...}{see method arguments}
}
\description{
S4 generic: PoolName
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{PoolName,character-method}}
  \item \code{\link{PoolName,PoolIndex-method}}
  \item \code{\link{PoolName,PoolName-method}}
}
}

