\name{RespirationCoefficients}
\alias{RespirationCoefficients}
\usage{RespirationCoefficients(A)}
\arguments{
\item{A}{A matrix valued function representing the model.}
}
\title{helper function to compute respiration coefficients}
\description{This function computes the respiration coefficients as function of time for all pools
according to the given matrix A}
\value{A vector valued function of time containing the respiration coefficients for all pools.}
