% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ScalarTimeMap-class}
\alias{ScalarTimeMap-class}
\title{S4 class for a scalar time dependent function on a finite time interval}
\description{
S4 class for a scalar time dependent function on a finite time interval
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{ScalarTimeMap} in their signature:}{
      \itemize{
        \item \code{\link{ConstLinDecompOpWithLinearScalarFactor,matrix,missing,missing,missing,ScalarTimeMap-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{ScalarTimeMap} in their signature:}{
    \subsection{superclass TimeMap}{
      \itemize{
        \item \code{\link{add_plot,TimeMap-method}}
        \item \code{\link{as.character,TimeMap-method}}
        \item \code{\link{GeneralDecompOp,TimeMap-method}}
        \item \code{\link{getFunctionDefinition,TimeMap-method}}
        \item \code{\link{getTimeRange,TimeMap-method}}
        \item \code{\link{InFluxes,TimeMap-method}}
        \item \code{\link{initialize,TimeMap-method}}
        \item \code{\link{plot,TimeMap-method}}
        \item \code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
      }
    }
  }
}
