% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarTimeMap.R
\name{ScalarTimeMap,data.frame,missing,missing,missing,missing-method}
\alias{ScalarTimeMap,data.frame,missing,missing,missing,missing-method}
\title{constructor for data given as 2 column data.frame}
\usage{
\S4method{ScalarTimeMap}{data.frame,missing,missing,missing,missing}(map, lag = 0, interpolation = splinefun)
}
\arguments{
\item{map}{object of class:\code{data.frame}, In this case a data.frame. Only the first two columns will be
used}

\item{lag}{a (scalar) delay}

\item{interpolation}{the interpolation, usually splinefun or approxfun}
}
\description{
constructor for data given as 2 column data.frame
}
