% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarTimeMap.R
\name{ScalarTimeMap,function,missing,missing,missing,missing-method}
\alias{ScalarTimeMap,function,missing,missing,missing,missing-method}
\title{manual constructor for just a function}
\usage{
\S4method{ScalarTimeMap}{`function`,missing,missing,missing,missing}(map, lag = 0)
}
\arguments{
\item{map}{object of class:\code{function}, no manual documentation}

\item{lag}{no manual documentation}
}
\description{
The interval will be set to [-Inf,Inf]
}
