% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarTimeMap.R
\name{ScalarTimeMap,missing,missing,missing,numeric,numeric-method}
\alias{ScalarTimeMap,missing,missing,missing,numeric,numeric-method}
\title{constructor for data and times given as vectors}
\usage{
\S4method{ScalarTimeMap}{missing,missing,missing,numeric,numeric}(times, data, lag = 0, interpolation = splinefun)
}
\arguments{
\item{times}{object of class:\code{numeric}, (the times for the values in data)}

\item{data}{object of class:\code{numeric}, the values at times}

\item{lag}{a (scalar) delay}

\item{interpolation}{the interpolation, usually splinefun or approxfun}
}
\description{
constructor for data and times given as vectors
}
