% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{show,NlModel-method}
\alias{show,NlModel-method}
\title{Automatic description: show,NlModel-method}
\usage{
\S4method{show}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}. : object of class:\code{NlModel}, no manual documentation}
}
\description{
Automatic description: show,NlModel-method
}
