% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_soil_texture_triangle_aisne.R
\docType{data}
\name{soil_texture_triangle_aisne}
\alias{soil_texture_triangle_aisne}
\title{soil_texture_triangle_aisne}
\format{
A data frame with 69 rows and 6 variables:
\describe{
\item{ texture }{  character }
\item{ clay }{  numeric }
\item{ silt }{  numeric }
\item{ sand }{  numeric }
\item{ name }{  character }
\item{ categorie }{  character }
}
}
\source{
Source
}
\usage{
soil_texture_triangle_aisne
}
\description{
AISNE.
}
\keyword{datasets}
