% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_soil_texture_triangle_usda.R
\docType{data}
\name{soil_texture_triangle_usda}
\alias{soil_texture_triangle_usda}
\title{soil_texture_triangle_usda}
\format{
A data frame with 58 rows and 5 variables:
\describe{
\item{ texture }{  character }
\item{ name }{  character }
\item{ clay }{  numeric }
\item{ silt }{  numeric }
\item{ sand }{  numeric }
}
}
\source{
Source
}
\usage{
soil_texture_triangle_usda
}
\description{
USDA.
}
\keyword{datasets}
