\name{SoilR.F0-class}
\Rdversion{1.1}
\docType{class}
\alias{SoilR.F0-class}
%% \alias{AbsoluteFractionModern}
%% \alias{Delta14C}
%% \alias{getFormat}
%% \alias{getValues}
%% \alias{initialize}

\title{SoilR.F0 S4 class}
\description{NA}
\section{Objects from the Class}{Objects can be created by calls of the form \code{new(SoilR.F0 ...)}}
\section{Slots}{
  \describe{
    \item{\code{values}:}{Object of class \code{"numeric"}}
    \item{\code{format}:}{Object of class \code{"character"}}
  }
}
\section{Methods}{
  \describe{
    \item{AbsoluteFractionModern}{\code{signature(F = "SoilR.F0")}: ... }
    \item{Delta14C}{\code{signature(F = "SoilR.F0")}: ... }
    \item{getFormat}{\code{signature(object = "SoilR.F0")}: ... }
    \item{getValues}{\code{signature(object = "SoilR.F0")}: ... }
    \item{initialize}{\code{signature(.Object = "SoilR.F0")}: ... }
	 }
}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





