\name{SoilR.F0.new}
\alias{SoilR.F0.new}
\usage{SoilR.F0.new(values=c(0),
format=Delta14C)}
\arguments{
\item{values}{a numeric vector}
\item{format}{a character string describing the format e.g. "Delta14C"}
}
\description{The function internally calls the constructor of the replacement class \code{\link{ConstFc-class}}.}
\title{deprecated function that used to create an object of class SoilR.F0 }
\value{An object of class \code{\link{ConstFc-class}} that contains data and a format description that can later be used to convert the data into other formats if the conversion is implemented.}
