\name{spectralNorm}
\alias{spectralNorm}
\title{a utility function to compute the spectral norm of a matrix}
\description{This function computes the spectral norm of a matrix which is defined as the inverse of the minimum of absolute values of the eigenvalues.}
\usage{spectralNorm(m)}
\arguments{
  \item{m}{a (quadratic) matrix}
}

\value{The spectral norm of the matrix, a real number.}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





