% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{StateDependentInFluxVector-class}
\alias{StateDependentInFluxVector-class}
\alias{StateDependentInFluxVector}
\title{Input vector that is a function of the pool contenst and time}
\description{
Input vector that is a function of the pool contenst and time
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{StateDependentInFluxVector} in their signature:}{
      \itemize{
        \item \code{\link{getFunctionDefinition,StateDependentInFluxVector-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{StateDependentInFluxVector} in their signature:}{
    \subsection{superclass InFluxes}{
      \itemize{
        \item \code{\link{InFluxes,InFluxes-method}}
      }
    }
  }
}
\section{S4-superclasses (in the package)}{
      \itemize{
        \item \code{\linkS4class{InFluxes}}
      }
}
