% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{StateIndependentInFluxList_by_PoolIndex-class}
\alias{StateIndependentInFluxList_by_PoolIndex-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{StateIndependentInFlux_by_PoolIndex}}
\description{
Subclass of list that is guaranteed to contain only elements of type
\linkS4class{StateIndependentInFlux_by_PoolIndex}
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{StateIndependentInFluxList_by_PoolIndex} in their signature:}{
      \itemize{
        \item \code{\link{InFluxes,StateIndependentInFluxList_by_PoolIndex-method}}
      }
  }
}
