% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{[[,MCSim-method}
\alias{[[,MCSim-method}
\title{Automatic description: [[,MCSim-method}
\usage{
\S4method{[[}{MCSim}(x, i)
}
\arguments{
\item{x}{object of class:\code{MCSim}. : object of class:\code{MCSim}, no manual documentation}

\item{i}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: [[,MCSim-method
}
