% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{summary,Model-method}
\alias{summary,Model-method}
\title{summary,Model-method}
\usage{
\S4method{summary}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. }
}
\description{
summary,Model-method
}
