\name{supportedFcAtmFormats}
\alias{supportedFcAtmFormats}
\title{report the implemented formats for atmospheric 14C data}
\description{This information is needed at different places in the code and is presented as a function
to avoid duplication}
\usage{supportedFcAtmFormats()}

\value{a list of strings with the supported Format Strings}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





