% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_soil_texture_identifier.R
\name{t_soil_texture_identifier}
\alias{t_soil_texture_identifier}
\title{Identification of the soil class texture.}
\usage{
t_soil_texture_identifier(clay_percentage, sand_percentage)
}
\arguments{
\item{clay_percentage}{Measured clay content (\%)}

\item{sand_percentage}{Measured sand content (\%)}
}
\value{
USDA soil class texture.
}
\description{
Identification of the soil class texture.
}
\examples{
t_soil_texture_identifier(clay_percentage = 60,
                          sand_percentage = 20)
}
