\name{TemporaryInputFlux}
\alias{TemporaryInputFlux}
\title{generic constructor}
\description{create a TemporaryInputFlux object from different sources}
\usage{TemporaryInputFlux(map, starttime, endtime, lag)}\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





