\name{TemporaryInputFlux_method__function_numeric_numeric_numeric}
\alias{TemporaryInputFlux_method__function_numeric_numeric_numeric}
\title{costructor }
\description{the method constructs a Temporary Input Flux from its basic ingredients}
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





