\name{TemporaryInputFlux_method__TimeMap_missing_missing_missing}
\alias{TemporaryInputFlux_method__TimeMap_missing_missing_missing}
\title{convert to TemporaryInputFlux}
\description{The method is used internally to convert TimeMap objects to TemporaryInputFlux objects, since the use of TimeMap objects is now deprecated.}
\arguments{
  \item{map}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





