% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,data.frame,missing,missing,missing,missing-method}
\alias{TimeMap,data.frame,missing,missing,missing,missing-method}
\title{TimeMap,data.frame,missing,missing,missing,missing-method}
\usage{
\S4method{TimeMap}{data.frame,missing,missing,missing,missing}(map, lag = 0, interpolation = splinefun)
}
\arguments{
\item{map}{object of class:\code{data.frame}. }

\item{lag}{}

\item{interpolation}{}
}
\description{
TimeMap,data.frame,missing,missing,missing,missing-method
}
