% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\docType{methods}
\name{TimeMap,function,missing,missing,missing,missing-method}
\alias{TimeMap,function,missing,missing,missing,missing-method}
\title{manual constructor for just a function}
\usage{
\S4method{TimeMap}{`function`,missing,missing,missing,missing}(map,
  lag = 0)
}
\description{
The interval will be set to [-Inf,Inf]
}
