\docType{methods}
\name{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}
\title{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method 
 pass through constructor}
\alias{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}
\usage{\S4method{TimeMap}{TimeMap,ANY,ANY,ANY,ANY}(map)}
\arguments{
\item{map}{: of class TimeMap, The function just returns its argument and helps to avoid a lot of checks if a conversion is necessary.}
}
\description{}
