\docType{methods}
\name{TimeMap,list,ANY,ANY-method}
\title{TimeMap,list,ANY,ANY-method}
\alias{TimeMap,list,ANY,ANY-method}
\usage{\S4method{TimeMap}{list,ANY,ANY}(map)}
\arguments{
\item{map}{: of class list}
}
\description{an a list of matrices or vectors (one matrix or vector for each time step)}
