\docType{methods}
\name{TimeMap,missing,missing,missing,array,numeric-method}
\title{TimeMap,missing,missing,missing,array,numeric-method}
\alias{TimeMap,missing,missing,missing,array,numeric-method}
\usage{\S4method{TimeMap}{missing,missing,missing,array,numeric}(data,
times,
lag=0)}
\arguments{
\item{data}{: of class array}
\item{times}{: of class numeric}
\item{lag}{:  delay}
}
\description{create a  TimeMap object from an array and a numeric vector of times}
