\docType{methods}
\name{TimeMap,function,numeric,numeric-method}
\title{TimeMap,function,numeric,numeric-method 
 constructor}
\alias{TimeMap,function,numeric,numeric-method}
\usage{\S4method{TimeMap}{function,numeric,numeric}(map,
starttime,
endtime,
lag=0)}
\arguments{
\item{map}{: of class function}
\item{starttime}{: of class numeric}
\item{endtime}{: of class numeric}
\item{lag}{:  delay}
}
\description{create a  TimeMap object from the function definition and the time interval  }
