\docType{methods}
\name{TimeMap,function,numeric,numeric,missing,missing-method}
\title{TimeMap,function,numeric,numeric,missing,missing-method 
 constructor}
\alias{TimeMap,function,numeric,numeric,missing,missing-method}
\usage{\S4method{TimeMap}{function,numeric,numeric,missing,missing}(map,
starttime,
endtime)}
\arguments{
\item{map}{: of class function}
\item{starttime}{: of class numeric}
\item{endtime}{: of class numeric}
}
\description{create a  TimeMap object from the function definition and the time interval  }
