% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,missing,missing,missing,numeric,matrix-method}
\alias{TimeMap,missing,missing,missing,numeric,matrix-method}
\title{automatic title}
\usage{
\S4method{TimeMap}{missing,missing,missing,numeric,matrix}(times, data, lag = 0, interpolation = splinefun)
}
\arguments{
\item{times}{object of class:\code{numeric}, no manual documentation}

\item{data}{object of class:\code{matrix}, no manual documentation}

\item{lag}{no manual documentation}

\item{interpolation}{no manual documentation}
}
\description{
automatic title
}
