\name{TimeMap.from.Matrix}
\alias{TimeMap.from.Matrix}
\title{TimeMap from Matrix}
\description{This function is the basic constructor of the class TimeMap.}
\usage{TimeMap.from.Matrix(mat)}
\arguments{
  \item{mat}{A Matrix containing points in time and corresponding values
the time is expected to be represented by the first column}
}

\value{An object of class TimeMap that can be used to describe models.}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





