% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap.new}
\alias{TimeMap.new}
\title{deprecated constructor of the class TimeMap.}
\usage{
TimeMap.new(t_start, t_end, f)
}
\arguments{
\item{t_start}{A number marking the begin of the time domain where the
function is valid}

\item{t_end}{A number the end of the time domain where the function is valid}

\item{f}{The time dependent function definition (a function in R's sense)}
}
\value{
An object of class TimeMap that can be used to describe models.
}
\description{
deprecated functions #################### use the generic TimeMap(...)
instead
}
