\description{no Description}
\name{TransportDecompositionOperator-class}
\alias{TransportDecompositionOperator-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "TransportDecompositionOperator")}: ... } \code{\link{getFunctionDefinition,TransportDecompositionOperator-method}}
	 }}
\title{TransportDecompositionOperator S4 class}
\section{Superclasses}{\code{\link{TimeMap-class}}\cr}
