% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{UnBoundInFlux-class}
\alias{UnBoundInFlux-class}
\title{constant decomposition operator}
\description{
no Description
}
\section{Methods}{
 Exported methods directly defined for class UnBoundInFlux:

\describe{ \item{getFunctionDefinition}{\code{signature(object =
"UnBoundInFlux")}: ... }
\code{\link{getFunctionDefinition,UnBoundInFlux-method}}
\item{getTimeRange}{\code{signature(object = "UnBoundInFlux")}: ... }
\code{\link{getTimeRange,UnBoundInFlux-method}} }

Methods inherited from superclasses:

from class InFlux:

\describe{ \item{GeneralInFlux}{\code{signature(object = "InFlux")}: ... }
\code{\link{GeneralInFlux,InFlux-method}} }
}

