% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{UnBoundLinDecompOp-class}
\alias{UnBoundLinDecompOp-class}
\title{constant decomposition operator}
\description{
no Description
}
\section{Methods}{
 Exported methods directly defined for class
UnBoundLinDecompOp:

\describe{ \item{BoundLinDecompOp}{\code{signature(map =
"UnBoundLinDecompOp")}: ... }
\code{\link{BoundLinDecompOp,UnBoundLinDecompOp-method}}
\item{getFunctionDefinition}{\code{signature(object =
"UnBoundLinDecompOp")}: ... }
\code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}
\item{getTimeRange}{\code{signature(object = "UnBoundLinDecompOp")}: ... }
\code{\link{getTimeRange,UnBoundLinDecompOp-method}} }

Methods inherited from superclasses:

from class DecompOp:

\describe{ \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ...
} \code{\link{GeneralDecompOp,DecompOp-method}} }
}

