% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{UnBoundLinDecompOp,function-method}
\alias{UnBoundLinDecompOp,function-method}
\title{UnBoundLinDecompOp,function-method}
\usage{
\S4method{UnBoundLinDecompOp}{`function`}(matFunc)
}
\arguments{
\item{matFunc}{object of class:\code{function}. }
}
\description{
UnBoundLinDecompOp,function-method
}
