\docType{methods}
\name{UnBoundLinDecompOp,function-method}
\title{UnBoundLinDecompOp,function-method 
 construct from matrix valued function}
\alias{UnBoundLinDecompOp,function-method}
\usage{\S4method{UnBoundLinDecompOp}{function}(matFunc= )}
\arguments{
\item{matFunc}{: of class function}
}
\description{This method creates a UnBoundLinDecompOp from a matrix
The operator is assumed to act on the vector of carbon stocks
by multiplication of the (time dependent) matrix from the left.}
