% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{UnBoundLinDecompOp,function-method}
\alias{UnBoundLinDecompOp,function-method}
\title{automatic title}
\usage{
\S4method{UnBoundLinDecompOp}{`function`}(matFunc)
}
\arguments{
\item{matFunc}{object of class:\code{function}, : no manual documentation}
}
\description{
automatic title
}
