% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{UnBoundLinDecompOp}
\alias{UnBoundLinDecompOp}
\title{Generic constructor for the class with the same name}
\usage{
UnBoundLinDecompOp(matFunc)
}
\arguments{
\item{matFunc}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{UnBoundLinDecompOp,function-method}}
      }
}
