% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp.R
\docType{class}
\name{UnBoundNonLinDecompOp-class}
\alias{UnBoundNonLinDecompOp-class}
\title{An S4 class to represent the operation nonlinear nonautonomous compartmental matrix}
\description{
An S4 class to represent the operation nonlinear nonautonomous compartmental matrix
}
\concept{UnBoundNonLinDecompOp_constructor}
