% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp.R
\name{UnBoundNonLinDecompOp,function,missing,missing,missing-method}
\alias{UnBoundNonLinDecompOp,function,missing,missing,missing-method}
\title{Constructor for the class with the same name}
\usage{
\S4method{UnBoundNonLinDecompOp}{`function`,missing,missing,missing}(matFunc)
}
\arguments{
\item{matFunc}{A matrix valued function of the state vector and time}
}
\description{
Constructor for the class with the same name
}
\seealso{
Other UnBoundNonLinDecompOp_constructor: 
\code{\link{UnBoundNonLinDecompOp,missing,vector,vector,numeric-method}}
}
\concept{UnBoundNonLinDecompOp_constructor}
