% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp.R
\name{UnBoundNonLinDecompOp,function,missing,missing,missing,ANY,ANY,ANY-method}
\alias{UnBoundNonLinDecompOp,function,missing,missing,missing,ANY,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{UnBoundNonLinDecompOp}{`function`,missing,missing,missing,ANY,ANY,ANY}(matFunc)
}
\arguments{
\item{matFunc}{object of class:\code{function}, : no manual documentation}
}
\description{
automatic title
}
