% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp.R
\name{UnBoundNonLinDecompOp,missing,vector,vector,numeric,ANY,ANY,ANY-method}
\alias{UnBoundNonLinDecompOp,missing,vector,vector,numeric,ANY,ANY,ANY-method}
\title{constructor}
\usage{
\S4method{UnBoundNonLinDecompOp}{missing,vector,vector,numeric,ANY,ANY,ANY}(internal_fluxes, out_fluxes, numberOfPools)
}
\arguments{
\item{internal_fluxes}{object of class:\code{vector}, vector of elements of type InternalFlux_by_PoolName}

\item{out_fluxes}{object of class:\code{vector}, vector of elements of type OutFlux_by_PoolName}

\item{numberOfPools}{object of class:\code{numeric}, no manual documentation}
}
\description{
constructor
}
