% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp_by_PoolNames.R
\name{UnBoundNonLinDecompOp,missing,missing,missing,missing,character,character,UnBoundNonLinDecompOp_by_PoolNames-method}
\alias{UnBoundNonLinDecompOp,missing,missing,missing,missing,character,character,UnBoundNonLinDecompOp_by_PoolNames-method}
\title{convert to Indexed version}
\usage{
\S4method{UnBoundNonLinDecompOp}{missing,missing,missing,missing,character,character,UnBoundNonLinDecompOp_by_PoolNames}(state_variable_names, timeSymbol, operator)
}
\description{
convert to Indexed version
}
