% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{UnBoundNonLinDecompOp_by_PoolNames-class}
\alias{UnBoundNonLinDecompOp_by_PoolNames-class}
\title{An S4 class to represent the of  nonlinear nonautonomous compartmental system independently of the order of state variables}
\description{
An S4 class to represent the of  nonlinear nonautonomous compartmental system independently of the order of state variables
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{UnBoundNonLinDecompOp_by_PoolNames} in their signature:}{
      \itemize{
        \item \code{\link{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp_by_PoolNames,character,character-method}}
        \item \code{\link{Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing,missing-method}}
        \item \code{\link{UnBoundNonLinDecompOp,missing,missing,missing,missing,character,character,UnBoundNonLinDecompOp_by_PoolNames-method}}
      }
  }
}
