% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp_by_PoolNames.R
\name{UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method}
\alias{UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method}
\title{automatic title}
\usage{
\S4method{UnBoundNonLinDecompOp_by_PoolNames}{InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character}(internal_fluxes, out_fluxes, timeSymbol)
}
\arguments{
\item{internal_fluxes}{object of class:\code{InternalFluxList_by_PoolName}, : : no manual documentation}

\item{out_fluxes}{object of class:\code{OutFluxList_by_PoolName}, : : no manual documentation}

\item{timeSymbol}{object of class:\code{character}, : : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
