% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp_by_PoolNames.R
\name{UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method}
\alias{UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method}
\title{UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method}
\usage{
\S4method{UnBoundNonLinDecompOp_by_PoolNames}{InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character}(internal_fluxes, out_fluxes, timeSymbol)
}
\arguments{
\item{internal_fluxes}{object of class:\code{InternalFluxList_by_PoolName}. }

\item{out_fluxes}{object of class:\code{OutFluxList_by_PoolName}. }

\item{timeSymbol}{object of class:\code{character}. }
}
\description{
UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method
}
