% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{UnBoundNonLinDecompOp_by_PoolNames}
\alias{UnBoundNonLinDecompOp_by_PoolNames}
\title{Generic constructor for the class with the same name}
\usage{
UnBoundNonLinDecompOp_by_PoolNames(internal_fluxes, out_fluxes, timeSymbol)
}
\description{
Generic constructor for the class with the same name
}
