% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangThreePoolNonAutonomous_sym.R
\name{WangThreePoolNonAutonomous_sym}
\alias{WangThreePoolNonAutonomous_sym}
\title{A non-autonomous version of the original Wang 3 pool model}
\usage{
WangThreePoolNonAutonomous_sym(
  alpha = 0.5,
  epsilon = 0.4,
  mu_b = 0.2,
  F_NPP = 3e+06,
  V_l = 0.5,
  V_s = 0.5,
  K_l = 1e+05,
  K_s = 100
)
}
\arguments{
\item{alpha}{fraction of carbon influx that directly enters the soil organic matter pool}

\item{epsilon}{microbial growth efficiency}

\item{mu_b}{turnover rate of microbial biomass per year, unit: "year^{-1}"}

\item{F_NPP}{carbon influx into soil, unit: "g C*m^{-2}*year^{-1}"}

\item{V_l}{maximum rate of litter carbon assimilation per unit microbial biomass per year}

\item{V_s}{maximum rate of soil carbon assimilation per unit microbial biomass per year}

\item{K_l}{half-saturation constant for litter carbon assimilation by microbial biomass}

\item{K_s}{half-saturation constant for soil carbon assimilation by microbial biomass}
}
\description{
An Example based on the original non-linear autonomous model as described in  
\url{http://doi.org/10.5194/bg-11-1817-2014}
with state_variables:
\enumerate{
 \item C_l
   desc: litter carbon
   unit: "g C m^{-2}"
 \item C_s
   desc: soil organic matter
   unit: "g C m^{-2}"
 \item C_b: 
   desc: microbial biomass
   unit: "g C m^{-2}"
}
Note that this is not a complete model run like most of the models in SoilR but a 
description of the fluxes that can be extended to a model run if initial values
and times are specified.
The default values are completely arbitrary.
So is one time dependency that has been added to demonstrate that this is possible everywhere and every 
part of the model can become non-autonomous. 
At the moment the variable t is mostly ignored like in the original Wang Model except for the first
influx to pool C_l.
}
