\docType{methods}
\name{$,NlModel-method}
\title{$,NlModel-method 
 overload the $ operator}
\alias{$,NlModel-method}
\usage{\S4method{$}{NlModel}(x,
name)}
\arguments{
\item{x}{: of class NlModel, the model}
\item{name}{:  the property to access}
}
\description{This function is still experimental and might change considerably 
It is meant as an alternative to all the get methods to provide 
a more R like interface.}
