\docType{methods}
\name{[,NlModel,character,ANY,ANY-method}
\title{[,NlModel,character,ANY,ANY-method 
 overloads the [] operator}
\alias{[,NlModel,character,ANY,ANY-method}
\usage{\S4method{[}{NlModel,character,ANY,ANY}(x= ,
i= )}
\arguments{
\item{x}{: of class NlModel, the model}
\item{i}{: of class character, the property to access}
}
\description{This function is still experimental and might change considerably 
It is meant as an alternative to all the get methods to provide 
a more R like interface.}
