\docType{methods}
\name{[[,MCSim-method}
\title{[[,MCSim-method}
\alias{[[,MCSim-method}
\usage{\S4method{[[}{MCSim}(x,
i)}
\arguments{
\item{x}{: of class MCSim, the simulator}
\item{i}{:  the index}
}
\description{This function is still experimental and should no be used in production code yet.   }
