\docType{methods}
\name{[[<-,MCSim-method}
\title{[[<-,MCSim-method 
 overloads the [[<- operator }
\alias{[[<-,MCSim-method}
\arguments{
\item{x}{: of class MCSim, The simulator}
\item{i}{:  the index}
\item{j}{:  ignored}
\item{...}{:  ignored}
\item{value}{:  the tasklist}
}
\description{This function is still experimental and should no be used in production code yet.   }
