\name{getF14}
\alias{getF14}
\title{Calculates the 14C fraction of all pools}
\description{This function computes the radiocarbon fraction for all pools as funtion of time.}
\usage{getF14(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





