\name{initialize_method__TimeMap}
\alias{initialize_method__TimeMap}
\title{initialize called by (new)}
\description{This is the standard constructor for objects of this class
It is called by statements of the form 
\code{new("TimeMap",start,end,f,lag)}
but should be avoided by user code since it is much more 
likely to change.
Rather use the generic constructor TimeMap(...)
which implements a lot of methods }
\arguments{
  \item{.Object}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{map}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





