\name{getTimeRange}
\alias{getTimeRange}
\title{getTimeRange}
\description{This function returns the time range of the given object. }
\usage{getTimeRange(object)}\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





