\name{TimeMap_method__function_numeric_numeric}
\alias{TimeMap_method__function_numeric_numeric}
\title{constructor}
\description{create a  TimeMap object from the function definition and the time interval  }
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





