\name{BoundLinDecompOp_method__function_numeric_numeric_numeric}
\alias{BoundLinDecompOp_method__function_numeric_numeric_numeric}
\title{a constructor for a single function without limits  }
\description{This method creates a BoundLinDecompOp from a timedependent function and a lag only}
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





