\name{BoundInFlux_method__function_numeric_numeric_numeric_missing}
\alias{BoundInFlux_method__function_numeric_numeric_numeric_missing}
\title{constructor }
\description{the method constructs an object from its basic ingredients}
\arguments{
  \item{map}{
}
  \item{starttime}{
}
  \item{endtime}{
}
  \item{lag}{
}
}



\author{Carlos A. Sierra, Markus Mueller}





