\name{BoundLinDecompOp_method__function_numeric_numeric}
\alias{BoundLinDecompOp_method__function_numeric_numeric}
\title{a constructor }
\description{This method creates a BoundLinDecompOp from a timedependent function and its domain }
\arguments{
  \item{map}{a function}
  \item{starttime}{the begin of the time domain}
  \item{endtime}{the end of the time domain}
}



\author{Carlos A. Sierra, Markus Mueller}





