\name{Delta14C_from_AbsoluteFractionModern_method__numeric}
\alias{Delta14C_from_AbsoluteFractionModern_method__numeric}
\title{Converts to Delta14C format}
\description{This method produces Delta14C values from Absolute Fraction Modern
Have a look at the methods for details.}
\arguments{
  \item{AbsoluteFractionModern}{A numeric object containing the values in Absolute Fraction Modern format}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





