\name{DecompOp-class}
\Rdversion{1.1}
\docType{class}
\alias{DecompOp-class}
\title{Class \code{"DecompOp"}}
\description{
NA
}
\section{Methods}{
  \describe{
    \item{DecompOp}{\code{signature(object = "DecompOp")}: ... }
	 }
}
\author{
 Carlos A. Sierra, Markus Mueller 
}
\examples{
showClass("DecompOp")
}
\keyword{classes}
