% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundFc.R
\name{AbsoluteFractionModern,BoundFc-method}
\alias{AbsoluteFractionModern,BoundFc-method}
\title{automatic title}
\usage{
\S4method{AbsoluteFractionModern}{BoundFc}(F)
}
\arguments{
\item{F}{no manual documentation}
}
\description{
automatic title
}
