% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AbsoluteFractionModern}
\alias{AbsoluteFractionModern}
\title{automatic title}
\usage{
AbsoluteFractionModern(F)
}
\arguments{
\item{F}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{AbsoluteFractionModern,BoundFc-method}}
        \item \code{\link{AbsoluteFractionModern,ConstFc-method}}
      }
}
