% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AbsoluteFractionModern}
\alias{AbsoluteFractionModern}
\title{AbsoluteFractionModern}
\usage{
AbsoluteFractionModern(F)
}
\arguments{
\item{F}{see method arguments}
}
\description{
AbsoluteFractionModern
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{AbsoluteFractionModern,BoundFc-method}}
  \item \code{\link{AbsoluteFractionModern,ConstFc-method}}
  \item \code{\link{AbsoluteFractionModern,BoundFc-method}}
  \item \code{\link{AbsoluteFractionModern,ConstFc-method}}
}
}

