\name{AbsoluteFractionModern}
\alias{AbsoluteFractionModern}
\title{convert its argument to an Absolute Fraction Modern representation}
\description{The function returns an object of the same type as its imput
this can be a number a matrix or an object of class FcAtm}
\usage{AbsoluteFractionModern(FcAtm)}
\arguments{
  \item{FcAtm}{an object that contains data and a formatdescription.  So it can be converted into the AbsoluteFractionModern format if a conversion is implemented.}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





