% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AbsoluteFractionModern_from_Delta14C,matrix-method}
\alias{AbsoluteFractionModern_from_Delta14C,matrix-method}
\title{automatic title}
\usage{
\S4method{AbsoluteFractionModern_from_Delta14C}{matrix}(delta14C)
}
\arguments{
\item{delta14C}{no manual documentation}
}
\description{
automatic title
}
