% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{AbsoluteFractionModern_from_Delta14C,matrix-method}
\alias{AbsoluteFractionModern_from_Delta14C,matrix-method}
\title{AbsoluteFractionModern_from_Delta14C,matrix-method Converts from Delta14C to
Absolute Fraction Modern}
\arguments{
\item{delta14C}{: of class matrix, An object of class matrix containing the
values in Delta14C format}
}
\description{
This method produces a matrix of Delta14C values from a Matrix or number of
Absolute Fraction Modern
}
