% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{AbsoluteFractionModern_from_Delta14C,numeric-method}
\alias{AbsoluteFractionModern_from_Delta14C,numeric-method}
\title{AbsoluteFractionModern_from_Delta14C,numeric-method Converts from Delta14C
to Absolute Fraction Modern}
\arguments{
\item{delta14C}{: of class numeric, A numeric object containing the values
in Delta14C format}
}
\description{
Converts a number or vector containing Delta14C values to the appropriate
Absolute Fraction Modern values. Have a look at the methods for details.
}
