% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AbsoluteFractionModern_from_Delta14C,numeric-method}
\alias{AbsoluteFractionModern_from_Delta14C,numeric-method}
\title{automatic title}
\usage{
\S4method{AbsoluteFractionModern_from_Delta14C}{numeric}(delta14C)
}
\arguments{
\item{delta14C}{object of class:\code{numeric}, no manual documentation}
}
\description{
automatic title
}
