\name{AbsoluteFractionModern_from_Delta14C}
\alias{AbsoluteFractionModern_from_Delta14C}
\title{convert its argument to an Absolute Fraction Modern representation}
\description{The function returns an object of the same type as its imput
this can be a number a matrix }
\usage{AbsoluteFractionModern_from_Delta14C(delta14C)}
\arguments{
  \item{delta14C}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





