\name{AbsoluteFractionModern_from_Delta14C}
\alias{AbsoluteFractionModern_from_Delta14C}
\usage{AbsoluteFractionModern_from_Delta14C(delta14C)}
\arguments{
\item{delta14C}{see the method arguments for details}
}
\section{Methods}{	\code{\link{AbsoluteFractionModern_from_Delta14C,matrix-method}}\cr
	\code{\link{AbsoluteFractionModern_from_Delta14C,numeric-method}}\cr}
\title{Converts its argument to an Absolute Fraction Modern representation}
\description{The function returns an object of the same type as its input,
which can be a number or a matrix.
Have a look at the methods for details.}
