\description{no Description}
\name{AbsoluteFractionModern_from_Delta14C}
\alias{AbsoluteFractionModern_from_Delta14C}
\usage{AbsoluteFractionModern_from_Delta14C(delta14C= )}
\title{AbsoluteFractionModern_from_Delta14C S4 generic}
\arguments{
\item{delta14C}{see the method arguments for details}
}
\section{Methods}{	\code{\link{AbsoluteFractionModern_from_Delta14C,matrix-method}}\cr
	\code{\link{AbsoluteFractionModern_from_Delta14C,numeric-method}}\cr}
