% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AbsoluteFractionModern_from_Delta14C}
\alias{AbsoluteFractionModern_from_Delta14C}
\title{conversion}
\usage{
AbsoluteFractionModern_from_Delta14C(delta14C)
}
\arguments{
\item{delta14C}{Object to be converted to AbsoluteFractionModern}
}
\description{
conversion
}
