% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{add_plot,TimeMap-method}
\alias{add_plot,TimeMap-method}
\title{Automatic description: add_plot,TimeMap-method}
\usage{
\S4method{add_plot}{TimeMap}(x, ...)
}
\arguments{
\item{x}{object of class:\code{TimeMap}. : object of class:\code{TimeMap}, no manual documentation}

\item{...}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: add_plot,TimeMap-method
}
