% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{add_plot,TimeMap-method}
\alias{add_plot,TimeMap-method}
\title{automatic title}
\usage{
\S4method{add_plot}{TimeMap}(x, ...)
}
\arguments{
\item{x}{no manual documentation}

\item{...}{no manual documentation}
}
\description{
automatic title
}
