% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{add_plot,TimeMap-method}
\alias{add_plot,TimeMap-method}
\title{add_plot,TimeMap-method}
\usage{
\S4method{add_plot}{TimeMap}(x, ...)
}
\arguments{
\item{x}{object of class:\code{TimeMap}. }

\item{...}{found automatically by inspection, not documented yet}
}
\description{
add_plot,TimeMap-method
}
