% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{add_plot}
\alias{add_plot}
\title{automatic title}
\usage{
add_plot(x, ...)
}
\arguments{
\item{x}{see method arguments}

\item{...}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{add_plot,TimeMap-method}}
      }
}
