% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{as.character,TimeMap-method}
\alias{as.character,TimeMap-method}
\title{as.character,TimeMap-method convert TimeMap Objects to something printable.}
\arguments{
\item{x}{: of class TimeMap, An Object of class time map}

\item{...}{: will be ignored}
}
\description{
This method is needed to print a TimeMap object.
}
