% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{as.character,TimeMap-method}
\alias{as.character,TimeMap-method}
\title{as.character,TimeMap-method}
\usage{
\S4method{as.character}{TimeMap}(x, ...)
}
\arguments{
\item{x}{object of class:\code{TimeMap}. }

\item{...}{found automatically by inspection, not documented yet}
}
\description{
as.character,TimeMap-method
}
