% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{availableParticleProperties,MCSim-method}
\alias{availableParticleProperties,MCSim-method}
\title{availableParticleProperties,MCSim-method}
\usage{
\S4method{availableParticleProperties}{MCSim}(object)
}
\arguments{
\item{object}{object of class:\code{MCSim}. }
}
\description{
availableParticleProperties,MCSim-method
}
