% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{availableParticleProperties}
\alias{availableParticleProperties}
\title{availableParticleProperties}
\usage{
availableParticleProperties(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
availableParticleProperties
}
