% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{availableParticleSets,MCSim-method}
\alias{availableParticleSets,MCSim-method}
\title{automatic title}
\usage{
\S4method{availableParticleSets}{MCSim}(object)
}
\arguments{
\item{object}{object of class:\code{MCSim}, no manual documentation}
}
\description{
automatic title
}
