% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{availableParticleSets}
\alias{availableParticleSets}
\title{availableParticleSets}
\usage{
availableParticleSets(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
availableParticleSets
}
