% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{availableResidentSets}
\alias{availableResidentSets}
\title{automatic title}
\usage{
availableResidentSets(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{availableResidentSets,MCSim-method}}
      }
}
