% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{availableResidentSets}
\alias{availableResidentSets}
\title{availableResidentSets}
\usage{
availableResidentSets(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
availableResidentSets
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{availableResidentSets,MCSim-method}}
  \item \code{\link{availableResidentSets,MCSim-method}}
}
}

