% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundFc.R
\name{BoundFc,character-method}
\alias{BoundFc,character-method}
\title{BoundFc,character-method}
\usage{
\S4method{BoundFc}{character}(format, ...)
}
\arguments{
\item{format}{object of class:\code{character}. }

\item{...}{}
}
\description{
BoundFc,character-method
}
