\title{Objects containing the atmospheric 14C fraction and the format it is provided in. }
\description{no Description}
\name{BoundFc-class}
\alias{BoundFc-class}
\docType{class}
\section{Methods}{No methods defined with class BoundFc in the signature.}
\section{Exported superclasses}{\code{\link{TimeMap-class}}\cr}
\section{Constructors}{	\code{\link{BoundFc}}\cr
 Please also look at constructors of non virtual subclasses }
