\description{no Description}
\name{BoundFc-class}
\alias{BoundFc-class}
\docType{class}
\section{Methods}{No exported methods directly defined for class BoundFc:


Methods inherited from superclasses:

from class TimeMap:

  \describe{
    \item{as.character}{\code{signature(x = "TimeMap")}: ... } \code{\link{as.character,TimeMap-method}}
    \item{GeneralDecompOp}{\code{signature(object = "TimeMap")}: ... } \code{\link{GeneralDecompOp,TimeMap-method}}
    \item{GeneralInFlux}{\code{signature(object = "TimeMap")}: ... } \code{\link{GeneralInFlux,TimeMap-method}}
    \item{getFunctionDefinition}{\code{signature(object = "TimeMap")}: ... } \code{\link{getFunctionDefinition,TimeMap-method}}
    \item{getTimeRange}{\code{signature(object = "TimeMap")}: ... } \code{\link{getTimeRange,TimeMap-method}}
    \item{TimeMap}{\code{signature(map = "TimeMap", starttime = "ANY", endtime = "ANY", times = "ANY", data = "ANY")}: ... } \code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
	 }}
\title{Objects containing the atmospheric 14C fraction and the format it is provided in. }
\section{Superclasses}{\code{\link{TimeMap-class}}\cr
\code{\link{Fc-class}}\cr}
\section{Constructors}{	\code{\link{BoundFc}}\cr}
