% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{BoundFc-class}
\alias{BoundFc-class}
\title{Objects containing the atmospheric 14C fraction and the format it is
provided in.}
\description{
no Description
}
\section{Methods}{
 No exported methods directly defined for class BoundFc:

Methods inherited from superclasses:

from class TimeMap:

\describe{ \item{GeneralDecompOp}{\code{signature(object = "TimeMap")}: ...
} \code{\link{GeneralDecompOp,TimeMap-method}}
\item{GeneralInFlux}{\code{signature(object = "TimeMap")}: ... }
\code{\link{GeneralInFlux,TimeMap-method}}
\item{TimeMap}{\code{signature(map = "TimeMap", starttime = "ANY", endtime =
"ANY", times = "ANY", data = "ANY")}: ... }
\code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
\item{add_plot}{\code{signature(x = "TimeMap")}: ... }
\code{\link{add_plot,TimeMap-method}} \item{as.character}{\code{signature(x
= "TimeMap")}: ... } \code{\link{as.character,TimeMap-method}}
\item{getFunctionDefinition}{\code{signature(object = "TimeMap")}: ... }
\code{\link{getFunctionDefinition,TimeMap-method}}
\item{getTimeRange}{\code{signature(object = "TimeMap")}: ... }
\code{\link{getTimeRange,TimeMap-method}} }
}

