% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{BoundFc,missing-method}
\alias{BoundFc,missing-method}
\title{BoundFc,missing-method A constructor}
\arguments{
\item{format}{: of class missing, If this method is called the format
argument was not present.}

\item{...}{: passed on to TimeMap}
}
\description{
The method constructs an object from a list which must contain a
}
