% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundFc.R
\name{BoundFc,missing-method}
\alias{BoundFc,missing-method}
\title{automatic title}
\usage{
\S4method{BoundFc}{missing}(format, ...)
}
\arguments{
\item{format}{no manual documentation}

\item{...}{no manual documentation}
}
\description{
automatic title
}
