% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundFc.R
\name{BoundFc,missing-method}
\alias{BoundFc,missing-method}
\title{BoundFc,missing-method}
\usage{
\S4method{BoundFc}{missing}(format, ...)
}
\arguments{
\item{format}{object of class:\code{missing}. }

\item{...}{found automatically by inspection, not documented yet}
}
\description{
BoundFc,missing-method
}
