\name{BoundFc}
\alias{BoundFc}
\usage{BoundFc(map,
starttime,
endtime,
lag,
format,
interpolation)}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{format}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\title{generic constructor}
\description{create a BoundFc object from different sources}
