% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{BoundFc}
\alias{BoundFc}
\title{automatic title}
\usage{
BoundFc(format, ...)
}
\arguments{
\item{format}{see method arguments}

\item{...}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{BoundFc,character-method}}
        \item \code{\link{BoundFc,missing-method}}
      }
}
