% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{BoundFc}
\alias{BoundFc}
\title{S4 generic: BoundFc}
\usage{
BoundFc(format, ...)
}
\arguments{
\item{format}{see method arguments}

\item{...}{see method arguments}
}
\description{
S4 generic: BoundFc
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{BoundFc,character-method}}
  \item \code{\link{BoundFc,missing-method}}
}
}

