\description{no Description}
\name{BoundFc}
\alias{BoundFc}
\usage{BoundFc(format,
...)}
\title{BoundFc S4 generic}
\arguments{
\item{format}{see the method arguments for details}
\item{...}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundFc,character-method}}\cr
	\code{\link{BoundFc,missing-method}}\cr}
