% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{BoundInFlux-class}
\alias{BoundInFlux-class}
\title{BoundInFlux S4 class}
\description{
defines a time dependent inputrate as function of time and including the
domain where the function is well defined. This can be used to avoid
interpolations out of range when mixing different time dependent data sets
}
\section{Methods}{
 No exported methods directly defined for class
BoundInFlux:

Methods inherited from superclasses:

from class InFlux:

\describe{ \item{GeneralInFlux}{\code{signature(object = "InFlux")}: ... }
\code{\link{GeneralInFlux,InFlux-method}} } from class TimeMap:

\describe{ \item{GeneralDecompOp}{\code{signature(object = "TimeMap")}: ...
} \code{\link{GeneralDecompOp,TimeMap-method}}
\item{GeneralInFlux}{\code{signature(object = "TimeMap")}: ... }
\code{\link{GeneralInFlux,TimeMap-method}}
\item{TimeMap}{\code{signature(map = "TimeMap", starttime = "ANY", endtime =
"ANY", times = "ANY", data = "ANY")}: ... }
\code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
\item{add_plot}{\code{signature(x = "TimeMap")}: ... }
\code{\link{add_plot,TimeMap-method}} \item{as.character}{\code{signature(x
= "TimeMap")}: ... } \code{\link{as.character,TimeMap-method}}
\item{getFunctionDefinition}{\code{signature(object = "TimeMap")}: ... }
\code{\link{getFunctionDefinition,TimeMap-method}}
\item{getTimeRange}{\code{signature(object = "TimeMap")}: ... }
\code{\link{getTimeRange,TimeMap-method}} }
}

