\docType{methods}
\name{BoundInFlux,ANY,ANY,ANY-method}
\title{BoundInFlux,ANY,ANY,ANY-method}
\alias{BoundInFlux,ANY,ANY,ANY-method}
\usage{\S4method{BoundInFlux}{ANY,ANY,ANY}(map,
starttime,
endtime,
lag=0,
interpolation)}
\arguments{
\item{map}{: of class ANY}
\item{starttime}{: of class ANY}
\item{endtime}{: of class ANY}
\item{lag}{}
\item{interpolation}{}
}
\description{The method is used internally to convert TimeMap objects to BoundInFlux objects, since the use of TimeMap objects is now deprecated.}
