\docType{methods}
\name{BoundInFlux,function,numeric,numeric,missing,missing-method}
\title{BoundInFlux,function,numeric,numeric,missing,missing-method}
\alias{BoundInFlux,function,numeric,numeric,missing,missing-method}
\usage{\S4method{BoundInFlux}{function,numeric,numeric,missing,missing}(map,
starttime,
endtime)}
\arguments{
\item{map}{: of class function}
\item{starttime}{: of class numeric}
\item{endtime}{: of class numeric}
}
\description{the method constructs an object from its basic ingredients}
