\docType{methods}
\name{BoundInFlux,ANY,missing,missing-method}
\title{BoundInFlux,ANY,missing,missing-method 
 constructor for BoundInFlux}
\alias{BoundInFlux,ANY,missing,missing-method}
\usage{\S4method{BoundInFlux}{ANY,missing,missing}(map,
lag=0,
interpolation)}
\arguments{
\item{map}{: of class ANY, see the same argument in  \code{\link{TimeMap}}}
\item{lag}{:  see the same argument in  \code{\link{TimeMap}}}
\item{interpolation}{:  see the same argument in  \code{\link{TimeMap}}}
}
\description{The method internally calls \code{\link{TimeMap}} and expects
the same kind of map argument}
