\name{BoundInFlux}
\alias{BoundInFlux}
\usage{BoundInFlux(map,
starttime,
endtime,
lag,
interpolation)}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundInFlux,TimeMap,missing,missing,missing,missing-method}}\cr
	\code{\link{BoundInFlux,data.frame,missing,missing,missing,missing-method}}\cr
	\code{\link{BoundInFlux,data.frame,missing,missing,numeric,function-method}}\cr
	\code{\link{BoundInFlux,function,numeric,numeric,missing,missing-method}}\cr
	\code{\link{BoundInFlux,function,numeric,numeric,numeric,missing-method}}\cr}
\title{generic constructor}
\description{create a BoundInFlux object from different sources}
