% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{BoundLinDecompOp,ANY-method}
\alias{BoundLinDecompOp,ANY-method}
\title{BoundLinDecompOp,ANY-method a constructor}
\arguments{
\item{map}{: of class ANY, passed on to TimeMap}

\item{...}{: passed on to TimeMap}
}
\description{
Creates a BoundLinDecompOp Object.
}
