\description{no Description}
\name{BoundLinDecompOp-class}
\alias{BoundLinDecompOp-class}
\docType{class}
\section{Methods}{No methods defined with class BoundLinDecompOp in the signature.}
\title{a decomposition operator described by a matrix valued function of time}
\section{Superclasses}{\code{\link{DecompOp-class}}\cr
\code{\link{TimeMap-class}}\cr}
\section{Constructors}{	\code{\link{BoundLinDecompOp}}\cr}
