\description{no Description}
\name{BoundLinDecompOp-class}
\alias{BoundLinDecompOp-class}
\docType{class}
\section{Methods}{No exported methods directly defined for class BoundLinDecompOp:


Methods inherited from superclasses:

from class DecompOp:

  \describe{
    \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ... } \code{\link{GeneralDecompOp,DecompOp-method}}
	 }
from class TimeMap:

  \describe{
    \item{add_plot}{\code{signature(x = "TimeMap")}: ... } \code{\link{add_plot,TimeMap-method}}
    \item{as.character}{\code{signature(x = "TimeMap")}: ... } \code{\link{as.character,TimeMap-method}}
    \item{GeneralDecompOp}{\code{signature(object = "TimeMap")}: ... } \code{\link{GeneralDecompOp,TimeMap-method}}
    \item{GeneralInFlux}{\code{signature(object = "TimeMap")}: ... } \code{\link{GeneralInFlux,TimeMap-method}}
    \item{getFunctionDefinition}{\code{signature(object = "TimeMap")}: ... } \code{\link{getFunctionDefinition,TimeMap-method}}
    \item{getTimeRange}{\code{signature(object = "TimeMap")}: ... } \code{\link{getTimeRange,TimeMap-method}}
    \item{TimeMap}{\code{signature(map = "TimeMap", starttime = "ANY", endtime = "ANY", times = "ANY", data = "ANY")}: ... } \code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
	 }}
\title{a decomposition operator described by a matrix valued function of time}
\section{Superclasses}{\code{\link{DecompOp-class}}\cr
\code{\link{TimeMap-class}}\cr}
\section{Constructors}{	\code{\link{BoundLinDecompOp}}\cr}
