\docType{methods}
\name{BoundLinDecompOp,ANY-method}
\title{BoundLinDecompOp,ANY-method 
 a constructor }
\alias{BoundLinDecompOp,ANY-method}
\usage{\S4method{BoundLinDecompOp}{ANY}(map= ,
...= )}
\arguments{
\item{map}{: of class ANY, passed on to TimeMap}
\item{...}{:  passed on to TimeMap}
}
\description{Creates a BoundLinDecompOp Object. }
