% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundLinDecompOp.R
\name{BoundLinDecompOp,UnBoundLinDecompOp-method}
\alias{BoundLinDecompOp,UnBoundLinDecompOp-method}
\title{BoundLinDecompOp,UnBoundLinDecompOp-method}
\usage{
\S4method{BoundLinDecompOp}{UnBoundLinDecompOp}(map, starttime = -Inf, endtime = Inf)
}
\arguments{
\item{map}{object of class:\code{UnBoundLinDecompOp}. }

\item{starttime}{}

\item{endtime}{}
}
\description{
BoundLinDecompOp,UnBoundLinDecompOp-method
}
