\description{no Description}
\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\usage{BoundLinDecompOp(map,
starttime,
endtime,
lag,
interpolation)}
\title{BoundLinDecompOp S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundLinDecompOp,ANY,ANY,ANY-method}}\cr
	\code{\link{BoundLinDecompOp,ANY,missing,missing-method}}\cr
	\code{\link{BoundLinDecompOp,UnBoundLinDecompOp,ANY,ANY-method}}\cr}
