% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\title{Generic constructor for the class with the same name}
\usage{
BoundLinDecompOp(map, ...)
}
\description{
Generic constructor for the class with the same name
}
