\description{no Description}
\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\usage{BoundLinDecompOp(map,
starttime,
endtime)}
\title{BoundLinDecompOp S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundLinDecompOp,ConstLinDecompOp,missing,missing-method}}\cr
	\code{\link{BoundLinDecompOp,data.frame,missing,missing-method}}\cr
	\code{\link{BoundLinDecompOp,function,numeric,numeric-method}}\cr
	\code{\link{BoundLinDecompOp,TimeMap,missing,missing-method}}\cr
	\code{\link{BoundLinDecompOp,UnBoundLinDecompOp,missing,missing-method}}\cr}
