\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\usage{BoundLinDecompOp(map,
starttime,
endtime)}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundLinDecompOp,ConstLinDecompOp,missing,missing-method}}\cr
	\code{\link{BoundLinDecompOp,data.frame,missing,missing-method}}\cr
	\code{\link{BoundLinDecompOp,function,numeric,numeric-method}}\cr
	\code{\link{BoundLinDecompOp,TimeMap,missing,missing-method}}\cr}
\title{Generic constructor}
\description{Creates a LinearDecompositonOperator from different sources.
Please look at the methods to see what kind of input is supported. }
