% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\title{Generic constructor for the class with the same name}
\usage{
BoundLinDecompOp(map, ...)
}
\arguments{
\item{map}{see method arguments}

\item{...}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{BoundLinDecompOp,ANY-method}}
        \item \code{\link{BoundLinDecompOp,UnBoundLinDecompOp-method}}
      }
}
