\description{no Description}
\name{BoundLinDecompOp}
\alias{BoundLinDecompOp}
\usage{BoundLinDecompOp(map= ,
...= )}
\title{BoundLinDecompOp S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{...}{see the method arguments for details}
}
\section{Methods}{	\code{\link{BoundLinDecompOp,ANY-method}}\cr
	\code{\link{BoundLinDecompOp,UnBoundLinDecompOp-method}}\cr}
