% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFluxRate_by_PoolName.R
\name{by_PoolIndex,ConstantInFluxRate_by_PoolName,ANY,ANY-method}
\alias{by_PoolIndex,ConstantInFluxRate_by_PoolName,ANY,ANY-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolIndex}{ConstantInFluxRate_by_PoolName,ANY,ANY}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{ConstantInFluxRate_by_PoolName}. }

\item{poolNames}{object of class:\code{ANY}. }
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
