% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRate_by_PoolName.R
\name{by_PoolIndex,ConstantInternalFluxRate_by_PoolName,ANY,ANY-method}
\alias{by_PoolIndex,ConstantInternalFluxRate_by_PoolName,ANY,ANY-method}
\title{new object with the source pool id converted to a PoolName if necessary}
\usage{
\S4method{by_PoolIndex}{ConstantInternalFluxRate_by_PoolName,ANY,ANY}(obj, poolNames)
}
\arguments{
\item{obj}{an object of class ConstantInternalFluxRate_by_PoolName}

\item{poolNames}{names of the pools}
}
\description{
new object with the source pool id converted to a PoolName if necessary
}
