% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolName.R
\name{by_PoolIndex,ConstantInternalFluxRateList_by_PoolName,ANY,ANY-method}
\alias{by_PoolIndex,ConstantInternalFluxRateList_by_PoolName,ANY,ANY-method}
\title{convert to a list indexed by pool names}
\usage{
\S4method{by_PoolIndex}{ConstantInternalFluxRateList_by_PoolName,ANY,ANY}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{ConstantInternalFluxRateList_by_PoolName}. }

\item{poolNames}{object of class:\code{ANY}. }
}
\description{
convert to a list indexed by pool names
}
