% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolName.R
\name{by_PoolIndex,ConstantInternalFluxRateList_by_PoolName,ANY,ANY-method}
\alias{by_PoolIndex,ConstantInternalFluxRateList_by_PoolName,ANY,ANY-method}
\title{convert to a list indexed by pool names}
\usage{
\S4method{by_PoolIndex}{ConstantInternalFluxRateList_by_PoolName,ANY,ANY}(obj, poolNames)
}
\description{
convert to a list indexed by pool names
}
