% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantOutFluxRateList_by_PoolName.R
\name{by_PoolIndex,ConstantOutFluxRateList_by_PoolName,ANY,ANY-method}
\alias{by_PoolIndex,ConstantOutFluxRateList_by_PoolName,ANY,ANY-method}
\title{convert to a list indexed by pool names}
\usage{
\S4method{by_PoolIndex}{ConstantOutFluxRateList_by_PoolName,ANY,ANY}(obj, poolNames)
}
\arguments{
\item{obj}{an object of class ConstantOutFluxRateList_by_PoolName}

\item{poolNames}{names of the pools}
}
\description{
convert to a list indexed by pool names
}
