% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFlux_by_PoolName.R
\name{by_PoolIndex,InFlux_by_PoolName,character,character-method}
\alias{by_PoolIndex,InFlux_by_PoolName,character,character-method}
\title{Convert the pool names to indices}
\usage{
\S4method{by_PoolIndex}{InFlux_by_PoolName,character,character}(obj, poolNames, timeSymbol)
}
\arguments{
\item{obj}{no manual documentation}

\item{poolNames}{no manual documentation}

\item{timeSymbol}{no manual documentation}
}
\description{
Convert the pool names to indices
}
