% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFlux_by_PoolName.R
\name{by_PoolIndex,InFlux_by_PoolName,character,character-method}
\alias{by_PoolIndex,InFlux_by_PoolName,character,character-method}
\title{automatic title}
\usage{
\S4method{by_PoolIndex}{InFlux_by_PoolName,character,character}(obj, poolNames, timeSymbol)
}
\arguments{
\item{obj}{object of class:\code{InFlux_by_PoolName}, no manual documentation}

\item{poolNames}{object of class:\code{character}, no manual documentation}

\item{timeSymbol}{object of class:\code{character}, no manual documentation}
}
\description{
automatic title
}
