% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolSource.R
\name{by_PoolIndex,PoolSource,ANY,ANY-method}
\alias{by_PoolIndex,PoolSource,ANY,ANY-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolIndex}{PoolSource,ANY,ANY}(obj, poolNames)
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
