% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolTarget.R
\name{by_PoolIndex,PoolTarget,ANY,ANY-method}
\alias{by_PoolIndex,PoolTarget,ANY,ANY-method}
\title{convert the source pool id to a number if necessary}
\usage{
\S4method{by_PoolIndex}{PoolTarget,ANY,ANY}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{PoolTarget}. }

\item{poolNames}{object of class:\code{ANY}. }
}
\description{
convert the source pool id to a number if necessary
}
