% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInFlux_by_PoolIndex.R
\name{by_PoolName,ConstantInFlux_by_PoolIndex-method}
\alias{by_PoolName,ConstantInFlux_by_PoolIndex-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolName}{ConstantInFlux_by_PoolIndex}(obj, poolNames)
}
\arguments{
\item{obj}{an object of class ConstantInFlux_by_PoolIndex}

\item{poolNames}{names of pools}
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
