% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRate_by_PoolIndex.R
\name{by_PoolName,ConstantInternalFluxRate_by_PoolIndex-method}
\alias{by_PoolName,ConstantInternalFluxRate_by_PoolIndex-method}
\title{new object with the source pool id converted to a PoolIndex if necessary}
\usage{
\S4method{by_PoolName}{ConstantInternalFluxRate_by_PoolIndex}(obj, poolNames)
}
\arguments{
\item{obj}{object of class:\code{ConstantInternalFluxRate_by_PoolIndex}. }

\item{poolNames}{}
}
\description{
new object with the source pool id converted to a PoolIndex if necessary
}
