% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolSource.R
\docType{methods}
\name{by_PoolName,PoolSource-method}
\alias{by_PoolName,PoolSource-method}
\title{new object with the source pool id converted to a PoolName if necessary}
\usage{
\S4method{by_PoolName}{PoolSource}(obj, poolNames)
}
\description{
new object with the source pool id converted to a PoolName if necessary
}
