% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{by_PoolName}
\alias{by_PoolName}
\title{by_PoolName}
\usage{
by_PoolName(obj, poolNames)
}
\arguments{
\item{obj}{see method arguments}

\item{poolNames}{see method arguments}
}
\description{
by_PoolName
}
