\name{C14Atm}
\alias{C14Atm}
\docType{data}
\title{
$^{14}C$  fraction in per mille 
}
\description{
    The usually used data set for the atmospheric value of the $^{14}C$  fraction in per mille
}
\usage{data(C14Atm)}
\format{
  A data frame with 108 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
  }
}
\examples{
data(C14Atm)
## maybe str(C14Atm) ; plot(C14Atm) ...
}
\keyword{datasets}
