% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C14Atm.R
\docType{data}
\name{C14Atm}
\alias{C14Atm}
\title{Atmospheric 14C fraction}
\format{
A data frame with 108 observations on the following 2 variables.
	\enumerate{ 
   \item V1 a numeric vector 
}
}
\usage{
data(C14Atm)
}
\description{
Atmospheric 14C fraction in units of Delta14C for the bomb period in the
	northern hemisphere.
	
	@note
	This dataset will be deprecated soon. Please use \link{C14Atm_NH} or
	\link{Hua2013} instead.
}
\examples{
#Notice that C14Atm is a shorter version of C14Atm_NH
require("SoilR")
data("C14Atm_NH")
plot(C14Atm_NH,type="l")
lines(C14Atm,col=2)

}
\keyword{datasets}
