\name{C14Atm_NH}
\alias{C14Atm_NH}
\docType{data}
\title{
$^{14}C$  fraction in per mille 
}
\description{
    The usually used data set for the atmospheric value of the $^{14}C$  fraction in per mille
}
\usage{data(C14Atm_NH)}
\format{
  A data frame with 111 observations on the following 2 variables.
  \describe{
    \item{\code{YEAR}}{a numeric vector}
    \item{\code{Atmosphere}}{a numeric vector}
  }
}
\examples{
data(C14Atm_NH)
## maybe str(C14Atm_NH) ; plot(C14Atm_NH) ...
}
\keyword{datasets}
