% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{C14Atm_NH}
\alias{C14Atm_NH}
\title{Post-bomb atmospheric 14C fraction}
\format{A data frame with 111 observations on the following 2 variables.
\describe{ \item{list("YEAR")}{a numeric vector with year of measurement.}
\item{list("Atmosphere")}{a numeric vector with the Delta 14 value of
atmospheric CO2 in per mil.} }}
\description{
Atmospheric 14C concentrations for the post-bomb period expressed as Delta
14C in per mile. This dataset contains a combination of observations from
locations in Europe and North America. It is representative for the Northern
Hemisphere.
}
\examples{

plot(C14Atm_NH,type="l")

}
\keyword{datasets}
