% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{$,NlModel-method}
\alias{$,NlModel-method}
\title{automatic title}
\usage{
\S4method{$}{NlModel}(x, name)
}
\arguments{
\item{x}{object of class:\code{NlModel}, : no manual documentation}

\item{name}{: no manual documentation}
}
\description{
automatic title
}
