% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection_by_PoolIndex.R
\name{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}
\alias{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}
\title{automatic title}
\usage{
\S4method{check_pool_ids}{PoolConnection_by_PoolIndex,integer}(obj, pools)
}
\arguments{
\item{obj}{object of class:\code{PoolConnection_by_PoolIndex}, : : no manual documentation}

\item{pools}{object of class:\code{integer}, : : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
