% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{check_pool_ids}
\alias{check_pool_ids}
\title{check_pool_ids}
\usage{
check_pool_ids(obj, pools)
}
\arguments{
\item{obj}{see method arguments}

\item{pools}{see method arguments}
}
\description{
check_pool_ids
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}}
  \item \code{\link{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}}
}
}

