% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{computeResults}
\alias{computeResults}
\title{automatic title}
\usage{
computeResults(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{computeResults,MCSim-method}}
        \item \code{\link{computeResults,NlModel-method}}
      }
}
