% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{computeResults}
\alias{computeResults}
\title{S4 generic: computeResults}
\usage{
computeResults(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: computeResults
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{computeResults,MCSim-method}}
  \item \code{\link{computeResults,NlModel-method}}
}
}

