% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstantInFluxList_by_PoolIndex-class}
\alias{ConstantInFluxList_by_PoolIndex-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInFlux_by_PoolIndex}}
\description{
Subclass of list that is guaranteed to contain only elements of type
\linkS4class{ConstantInFlux_by_PoolIndex}
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{ConstantInFluxList_by_PoolIndex} in their signature:}{
      \itemize{
        \item \code{\link{ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method}}
        \item \code{\link{InFluxes,ConstantInFluxList_by_PoolIndex-method}}
      }
  }
}
