% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstantInternalFluxRate_by_PoolIndex}
\alias{ConstantInternalFluxRate_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
ConstantInternalFluxRate_by_PoolIndex(
  sourceIndex,
  destinationIndex,
  src_to_dest,
  rate_constant
)
}
\arguments{
\item{sourceIndex}{see method arguments}

\item{destinationIndex}{see method arguments}

\item{src_to_dest}{see method arguments}

\item{rate_constant}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{ConstantInternalFluxRate_by_PoolIndex,missing,missing,character,numeric-method}}
        \item \code{\link{ConstantInternalFluxRate_by_PoolIndex,numeric,numeric,missing,numeric-method}}
      }
}
