% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstantInternalFluxRate_by_PoolIndex}
\alias{ConstantInternalFluxRate_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
ConstantInternalFluxRate_by_PoolIndex(
  sourceIndex,
  destinationIndex,
  src_to_dest,
  rate_constant
)
}
\description{
Generic constructor for the class with the same name
}
