% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRate_by_PoolName.R
\name{ConstantInternalFluxRate_by_PoolName,character,character,missing,numeric-method}
\alias{ConstantInternalFluxRate_by_PoolName,character,character,missing,numeric-method}
\title{constructor with argument conversion}
\usage{
\S4method{ConstantInternalFluxRate_by_PoolName}{character,character,missing,numeric}(
  sourceName,
  destinationName,
  rate_constant
)
}
\arguments{
\item{sourceName}{object of class:\code{character}. }

\item{destinationName}{object of class:\code{character}. }

\item{rate_constant}{object of class:\code{numeric}. }
}
\description{
constructor with argument conversion
}
