% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolIndex.R
\name{ConstantInternalFluxRateList_by_PoolIndex,numeric-method}
\alias{ConstantInternalFluxRateList_by_PoolIndex,numeric-method}
\title{automatic title}
\usage{
\S4method{ConstantInternalFluxRateList_by_PoolIndex}{numeric}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
