% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantInternalFluxRateList_by_PoolIndex.R
\name{ConstantInternalFluxRateList_by_PoolIndex,numeric-method}
\alias{ConstantInternalFluxRateList_by_PoolIndex,numeric-method}
\title{ConstantInternalFluxRateList_by_PoolIndex,numeric-method}
\usage{
\S4method{ConstantInternalFluxRateList_by_PoolIndex}{numeric}(object)
}
\arguments{
\item{object}{object of class:\code{numeric}. }
}
\description{
ConstantInternalFluxRateList_by_PoolIndex,numeric-method
}
