% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{ConstInFlux-class}
\alias{ConstInFlux-class}
\title{ConstInFlux S4 class}
\description{
defines a constant inputrate
}
\section{Methods}{
 Exported methods directly defined for class ConstInFlux:

\describe{ \item{getFunctionDefinition}{\code{signature(object =
"ConstInFlux")}: ... }
\code{\link{getFunctionDefinition,ConstInFlux-method}}
\item{getTimeRange}{\code{signature(object = "ConstInFlux")}: ... }
\code{\link{getTimeRange,ConstInFlux-method}} }

Methods inherited from superclasses:

from class InFlux:

\describe{ \item{GeneralInFlux}{\code{signature(object = "InFlux")}: ... }
\code{\link{GeneralInFlux,InFlux-method}} }
}

