\title{ConstInFlux S4 class}
\description{defines a constant inputrate }
\name{ConstInFlux-class}
\alias{ConstInFlux-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "ConstInFlux")}: ... } \code{\link{getFunctionDefinition,ConstInFlux-method}}
    \item{getTimeRange}{\code{signature(object = "ConstInFlux")}: ... } \code{\link{getTimeRange,ConstInFlux-method}}
	 }}
\section{Constructors}{	\code{\link{ConstInFlux}}\cr
 Please also look at constructors of non virtual subclasses }
