% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstInFluxes.R
\name{ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method}
\alias{ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method}
\title{ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method}
\usage{
\S4method{ConstInFluxes}{ConstantInFluxList_by_PoolIndex,numeric}(map, numberOfPools)
}
\arguments{
\item{map}{object of class:\code{ConstantInFluxList_by_PoolIndex}. }

\item{numberOfPools}{object of class:\code{numeric}. }
}
\description{
ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method
}
