% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstInFluxes}
\alias{ConstInFluxes}
\title{automatic title}
\usage{
ConstInFluxes(map, numberOfPools)
}
\arguments{
\item{map}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
automatic title
}
