% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstInFluxes}
\alias{ConstInFluxes}
\title{automatic title}
\usage{
ConstInFluxes(map, numberOfPools)
}
\arguments{
\item{map}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{ConstInFluxes,ConstantInFluxList_by_PoolIndex,numeric-method}}
        \item \code{\link{ConstInFluxes,numeric,ANY-method}}
      }
}
