\description{no Description}
\name{ConstLinDecompOp-class}
\alias{ConstLinDecompOp-class}
\docType{class}
\section{Methods}{Exported methods directly defined for class ConstLinDecompOp:

  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
    \item{getMeanTransitTime}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getMeanTransitTime,ConstLinDecompOp-method}}
    \item{getTimeRange}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getTimeRange,ConstLinDecompOp-method}}
    \item{getTransitTimeDistributionDensity}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getTransitTimeDistributionDensity,ConstLinDecompOp-method}}
	 }

Methods inherited from superclasses:

from class DecompOp:

  \describe{
    \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ... } \code{\link{GeneralDecompOp,DecompOp-method}}
	 }}
\title{constant decomposition operator }
\section{Superclasses}{\code{\link{DecompOp-class}}\cr}
\section{Constructors}{	\code{\link{ConstLinDecompOp}}\cr}
