% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstLinDecompOp-class}
\alias{ConstLinDecompOp-class}
\title{A class to represent a constant (=nonautonomous,linear) compartmental matrix 
or equivalently a combination of ordered constant internal flux rates and 
constant out flux rates.}
\description{
A class to represent a constant (=nonautonomous,linear) compartmental matrix 
or equivalently a combination of ordered constant internal flux rates and 
constant out flux rates.
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{ConstLinDecompOp} in their signature:}{
      \itemize{
        \item \code{\link{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}}
        \item \code{\link{getConstantCompartmentalMatrix,ConstLinDecompOp-method}}
        \item \code{\link{getConstantInternalFluxRateList_by_PoolIndex,ConstLinDecompOp-method}}
        \item \code{\link{getConstantOutFluxRateList_by_PoolIndex,ConstLinDecompOp-method}}
        \item \code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
        \item \code{\link{getMeanTransitTime,ConstLinDecompOp-method}}
        \item \code{\link{getTimeRange,ConstLinDecompOp-method}}
        \item \code{\link{getTransitTimeDistributionDensity,ConstLinDecompOp-method}}
        \item \code{\link{initialize,ConstLinDecompOp-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{ConstLinDecompOp} in their signature:}{
    \subsection{superclass DecompOp}{
      \itemize{
        \item \code{\link{GeneralDecompOp,DecompOp-method}}
      }
    }
  }
}
