% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstLinDecompOp-class}
\alias{ConstLinDecompOp-class}
\title{A class to represent a constant (=nonautonomous,linear) compartmental matrix 
or equivalently a combination of ordered constant internal flux rates and 
constant out flux rates.}
\description{
A class to represent a constant (=nonautonomous,linear) compartmental matrix 
or equivalently a combination of ordered constant internal flux rates and 
constant out flux rates.
}
