\description{no Description}
\name{ConstLinDecompOp-class}
\alias{ConstLinDecompOp-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
    \item{getMeanTransitTime}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getMeanTransitTime,ConstLinDecompOp-method}}
    \item{getTimeRange}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getTimeRange,ConstLinDecompOp-method}}
    \item{getTransitTimeDistributionDensity}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getTransitTimeDistributionDensity,ConstLinDecompOp-method}}
	 }}
\title{constant decomposition operator }
\section{Exported superclasses}{\code{\link{DecompOp-class}}\cr}
\section{Constructors}{	\code{\link{ConstLinDecompOp}}\cr
 Please also look at constructors of non virtual subclasses }
