% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{ConstLinDecompOp-class}
\alias{ConstLinDecompOp-class}
\title{constant decomposition operator}
\description{
no Description
}
\section{Methods}{
 Exported methods directly defined for class
ConstLinDecompOp:

\describe{ \item{getFunctionDefinition}{\code{signature(object =
"ConstLinDecompOp")}: ... }
\code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
\item{getMeanTransitTime}{\code{signature(object = "ConstLinDecompOp")}: ...
} \code{\link{getMeanTransitTime,ConstLinDecompOp-method}}
\item{getTimeRange}{\code{signature(object = "ConstLinDecompOp")}: ... }
\code{\link{getTimeRange,ConstLinDecompOp-method}}
\item{getTransitTimeDistributionDensity}{\code{signature(object =
"ConstLinDecompOp")}: ... }
\code{\link{getTransitTimeDistributionDensity,ConstLinDecompOp-method}} }

Methods inherited from superclasses:

from class DecompOp:

\describe{ \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ...
} \code{\link{GeneralDecompOp,DecompOp-method}} }
}

