\title{constant decomposition operator }
\description{no Description}
\name{ConstLinDecompOp-class}
\alias{ConstLinDecompOp-class}
\docType{class}
\section{Methods}{  \describe{
    \item{BoundLinDecompOp}{\code{signature(map = "ConstLinDecompOp", starttime = "missing", endtime = "missing")}: ... } \code{\link{BoundLinDecompOp,ConstLinDecompOp,missing,missing-method}}
    \item{getFunctionDefinition}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
    \item{getMeanTransitTime}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getMeanTransitTime,ConstLinDecompOp-method}}
    \item{getTimeRange}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getTimeRange,ConstLinDecompOp-method}}
    \item{getTransitTimeDistributionDensity}{\code{signature(object = "ConstLinDecompOp")}: ... } \code{\link{getTransitTimeDistributionDensity,ConstLinDecompOp-method}}
	 }}
