% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstLinDecompOp}
\alias{ConstLinDecompOp}
\title{Generic constructor for the class with the same name}
\usage{
ConstLinDecompOp(
  mat,
  internal_flux_rates,
  out_flux_rates,
  numberOfPools,
  poolNames
)
}
\description{
Generic constructor for the class with the same name
}
