\name{ConstLinDecompOp}
\alias{ConstLinDecompOp}
\usage{ConstLinDecompOp(mat)}
\arguments{
\item{mat}{see the method arguments for details}
}
\section{Methods}{	\code{\link{ConstLinDecompOp,matrix-method}}\cr}
\title{Generic constructor}
\description{Creates a ConstantDecompositonOperator object. }
