% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ConstLinDecompOp}
\alias{ConstLinDecompOp}
\title{Generic constructor for the class with the same name}
\usage{
ConstLinDecompOp(
  mat,
  internal_flux_rates,
  out_flux_rates,
  numberOfPools,
  poolNames
)
}
\arguments{
\item{mat}{a square compartmental matrix}

\item{internal_flux_rates}{rates of internal transfers among pools}

\item{out_flux_rates}{rates of transfer out of poolss}

\item{numberOfPools}{total number of pools in the system}

\item{poolNames}{names of all pools}
}
\description{
Generic constructor for the class with the same name
}
