% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstLinDecompOpWithLinearScalarFactor-class}
\alias{ConstLinDecompOpWithLinearScalarFactor-class}
\title{A class to represent a constant (=nonautonomuous,linear) compartmental matrix 
with a time dependent (linear) scalar pre factor 
This is a special case of a linear compartmental operator/matrix}
\description{
A class to represent a constant (=nonautonomuous,linear) compartmental matrix 
with a time dependent (linear) scalar pre factor 
This is a special case of a linear compartmental operator/matrix
}
