% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ConstLinDecompOpWithLinearScalarFactor-class}
\alias{ConstLinDecompOpWithLinearScalarFactor-class}
\title{A class to represent a constant (=nonautonomous,linear) compartmental matrix 
with a time dependent (linear) scalar pre factor 
This is a special case of a linear compartmental operator/matrix}
\description{
A class to represent a constant (=nonautonomous,linear) compartmental matrix 
with a time dependent (linear) scalar pre factor 
This is a special case of a linear compartmental operator/matrix
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{ConstLinDecompOpWithLinearScalarFactor} in their signature:}{
      \itemize{
        \item \code{\link{getConstantCompartmentalMatrix,ConstLinDecompOpWithLinearScalarFactor-method}}
        \item \code{\link{getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method}}
        \item \code{\link{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}}
        \item \code{\link{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}}
        \item \code{\link{getTimeRange,ConstLinDecompOpWithLinearScalarFactor-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{ConstLinDecompOpWithLinearScalarFactor} in their signature:}{
    \subsection{superclass DecompOp}{
      \itemize{
        \item \code{\link{GeneralDecompOp,DecompOp-method}}
      }
    }
  }
}
\section{S4-superclasses (in the package)}{
      \itemize{
        \item \code{\linkS4class{DecompOp}}
      }
}
