% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\name{ConstLinDecompOpWithLinearScalarFactor,matrix,missing,missing,missing,ScalarTimeMap-method}
\alias{ConstLinDecompOpWithLinearScalarFactor,matrix,missing,missing,missing,ScalarTimeMap-method}
\title{convert names of  vectors or lists to class ConstantOutFluxRate 
convert names of  vectors or lists to class ConstantInternalFluxRate}
\usage{
\S4method{ConstLinDecompOpWithLinearScalarFactor}{matrix,missing,missing,missing,ScalarTimeMap}(mat, xi)
}
\arguments{
\item{mat}{object of class:\code{matrix}. }

\item{xi}{object of class:\code{ScalarTimeMap}. }
}
\description{
convert names of  vectors or lists to class ConstantOutFluxRate 
convert names of  vectors or lists to class ConstantInternalFluxRate

ConstLinDecompOpWithLinearScalarFactor,matrix,missing,missing,missing,ScalarTimeMap-method
}
