\title{decomposition operator }
\description{no Description}
\name{DecompOp-class}
\alias{DecompOp-class}
\docType{class}
\section{Methods}{  \describe{
    \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ... } \code{\link{GeneralDecompOp,DecompOp-method}}
	 }}
\section{Subclasses}{\describe{
	\code{\link{BoundLinDecompOp-class}}\cr
	\code{\link{DecompositionOperator-class}}\cr
	\code{\link{ConstLinDecompOp-class}}\cr
}}
