\title{decomposition operator }
\description{no Description}
\name{DecompOp-class}
\alias{DecompOp-class}
\docType{class}
\section{Methods}{  \describe{
    \item{DecompOpSubClassInstance}{\code{signature(object = "DecompOp")}: ... } \code{\link{DecompOpSubClassInstance,DecompOp-method}}
	 }}
\section{Subclasses}{\describe{
	\code{\link{BoundLinDecompOp-class}}\cr
	\code{\link{DecompositionOperator-class}}\cr
	\code{\link{ConstLinDecompOp-class}}\cr
}}
\section{Constructors found by naming convention}{Since the class is virtual it can not be instanciated directly, but a function:
	\code{\link{DecompOpSubClassInstance}}\cr
has been found, that produces instances of subclasses.
             Please also look at constructors of non virtual subclasses }
