% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{DecompOp-class}
\alias{DecompOp-class}
\title{S4-class to represent compartmental operators}
\description{
S4-class to represent compartmental operators
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{DecompOp} in their signature:}{
      \itemize{
        \item \code{\link{GeneralDecompOp,DecompOp-method}}
      }
  }
}
\section{S4-subclasses}{
      \itemize{
        \item \code{\linkS4class{DecompositionOperator}}
        \item \code{\linkS4class{ConstLinDecompOp}}
        \item \code{\linkS4class{ConstLinDecompOpWithLinearScalarFactor}}
        \item \code{\linkS4class{BoundLinDecompOp}}
        \item \code{\linkS4class{UnBoundLinDecompOp}}
        \item \code{\linkS4class{UnBoundNonLinDecompOp}}
      }
}
