\title{decomposition operator }
\description{no Description}
\name{DecompOp-class}
\alias{DecompOp-class}
\docType{class}
\section{Methods}{No methods defined with class DecompOp in the signature.}
\section{Subclasses}{\describe{
	\code{\link{BoundLinDecompOp-class}}\cr
	\code{\link{DecompositionOperator-class}}\cr
	\code{\link{ConstLinDecompOp-class}}\cr
}}
