% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{DecompositionOperator-class}
\alias{DecompositionOperator-class}
\title{deprecated decomposition operator class}
\description{
no Description
}
\section{Methods}{
 Exported methods directly defined for class
DecompositionOperator:

\describe{ \item{getFunctionDefinition}{\code{signature(object =
"DecompositionOperator")}: ... }
\code{\link{getFunctionDefinition,DecompositionOperator-method}}
\item{getTimeRange}{\code{signature(object = "DecompositionOperator")}: ...
} \code{\link{getTimeRange,DecompositionOperator-method}} }

Methods inherited from superclasses:

from class DecompOp:

\describe{ \item{GeneralDecompOp}{\code{signature(object = "DecompOp")}: ...
} \code{\link{GeneralDecompOp,DecompOp-method}} }
}

