\description{no Description}
\name{DecompositionOperator-class}
\alias{DecompositionOperator-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "DecompositionOperator")}: ... } \code{\link{getFunctionDefinition,DecompositionOperator-method}}
    \item{getTimeRange}{\code{signature(object = "DecompositionOperator")}: ... } \code{\link{getTimeRange,DecompositionOperator-method}}
	 }}
\title{deprecated decomposition operator class }
\section{Superclasses}{\code{\link{DecompOp-class}}\cr}
