\name{DecompositionOperator-class}
\Rdversion{1.1}
\docType{class}
\alias{DecompositionOperator-class}
\title{Class \code{"DecompositionOperator"}}
\description{
NA
}
\section{Methods}{
  \describe{
    \item{DecompositionOperator}{\code{signature(object = "DecompositionOperator")}: ... }
	 }
}
\author{
 Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de> 
}
\examples{
showClass("DecompositionOperator")
}
\keyword{classes}
