% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundFc.R
\name{Delta14C,BoundFc-method}
\alias{Delta14C,BoundFc-method}
\title{Delta14C,BoundFc-method}
\usage{
\S4method{Delta14C}{BoundFc}(F)
}
\arguments{
\item{F}{object of class:\code{BoundFc}. }
}
\description{
Delta14C,BoundFc-method
}
