% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstFc.R
\name{Delta14C,ConstFc-method}
\alias{Delta14C,ConstFc-method}
\title{automatic title}
\usage{
\S4method{Delta14C}{ConstFc}(F)
}
\arguments{
\item{F}{object of class:\code{ConstFc}, no manual documentation}
}
\description{
automatic title
}
