\name{Delta14C_from_AbsoluteFractionModern}
\alias{Delta14C_from_AbsoluteFractionModern}
\title{convert its argument from an Absolute Fraction Modern to a Delta14C representation}
\description{The function returns an object of the same type as its imput
this can be a number a matrix }
\usage{Delta14C_from_AbsoluteFractionModern(AbsoluteFractionModern)}
\arguments{
  \item{AbsoluteFractionModern}{an object}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





