% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{Delta14C_from_AbsoluteFractionModern}
\alias{Delta14C_from_AbsoluteFractionModern}
\title{automatic title}
\usage{
Delta14C_from_AbsoluteFractionModern(AbsoluteFractionModern)
}
\arguments{
\item{AbsoluteFractionModern}{see method arguments}
}
\description{
automatic title
}
