\description{no Description}
\name{Delta14C_from_AbsoluteFractionModern}
\alias{Delta14C_from_AbsoluteFractionModern}
\usage{Delta14C_from_AbsoluteFractionModern(AbsoluteFractionModern)}
\title{Delta14C_from_AbsoluteFractionModern S4 generic}
\arguments{
\item{AbsoluteFractionModern}{A numeric object}
}
\section{Methods}{	\code{\link{Delta14C_from_AbsoluteFractionModern,matrix-method}}\cr
	\code{\link{Delta14C_from_AbsoluteFractionModern,numeric-method}}\cr}
