\name{deSolve.lsoda.wrapper}
\alias{deSolve.lsoda.wrapper}
\title{deSolve lsoda wrapper}

\usage{deSolve.lsoda.wrapper(t, ydot, startValues)}
\arguments{
  \item{t}{A row vector containing the points in time where the solution is sought.}
  \item{ydot}{The function of y and t that computes the derivative for a given
point in time and a column vector y.}
  \item{startValues}{A column vector with the starting values.}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





