% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.2DBoundInFluxFromFunction.R
\name{example.2DBoundInFluxFromFunction}
\alias{example.2DBoundInFluxFromFunction}
\title{example.2DBoundInFluxFromFunction}
\usage{
example.2DBoundInFluxFromFunction()
}
\value{
The returned object represents a time dependent Influx into a two
pool model.
}
\description{
Create a 2-dimensionsonal example of a BoundInFlux object
}
