\name{example.2DBoundInFluxFromFunction}
\alias{example.2DBoundInFluxFromFunction}
\usage{example.2DBoundInFluxFromFunction()}
\arguments{
}
\description{Create a 2-dimensionsonal example of a BoundInflux object}
\title{example.2DBoundInFluxFromFunction}
\value{The returned object represents a time dependentinflux into 
a two pool model.}
