\name{example.2DConstInFluxFromVector}
\alias{example.2DConstInFluxFromVector}
\usage{example.2DConstInFluxFromVector()}
\arguments{
}
\description{An example used in tests and other examples.}
\title{2D example  for constant Influx}
\value{The returned object represents a time invariant constant influx into 
a two pool model.}
