\name{example.2DConstInFluxFromVector}
\alias{example.2DConstInFluxFromVector}
\usage{example.2DConstInFluxFromVector()}
\arguments{
}
\description{For testing and documenting purposes only}
\title{2D example  for constant Influx}
\value{The returned object represents a time invariant constant influx into 
a two pool model.}
