% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{Fc-class}
\alias{Fc-class}
\title{Fc S4 class}
\description{
The \eqn{^{14}C}{14C} fraction is a necessary ingredient of any
\code{\linkS4class{Model_14}} object. In the most general case it is a real
valued function of time, accompanied by a string describing the unit or
format (i.e. "Delta14C" or "afn" for Absolute Fraction Modern) . In the most
simple case it is constant real number plus format.
}
\section{Methods}{
 No exported methods directly defined for class Fc:
}

