% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{Fc-class}
\alias{Fc-class}
\title{automatic title}
\description{
automatic title
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{Fc} in their signature:}{
      \itemize{
        \item \code{\link{getFormat,Fc-method}}
      }
  }
}
\section{S4-subclasses}{
      \itemize{
        \item \code{\linkS4class{BoundFc}}
        \item \code{\linkS4class{ConstFc}}
      }
}
