\name{fT.LandT}
\alias{fT.LandT}
\usage{fT.LandT(Temp)}
\arguments{
\item{Temp}{A scalar or vector containing values of soil temperature for which the effects on decomposition rates are calculated}
}
\title{Effects of temperature on decomposition rates according to a function proposed by Lloyd and Taylor (1994)}
\description{Calculates the effects of temperature on decomposition rates according to a function proposed by Lloyd and Taylor (1994).}
\references{Lloyd, J., and J. A. Taylor (1994), On the Temperature Dependence of Soil Respiration, 
Functional Ecology, 8(3), 315-323.}
\value{A scalar or a vector containing the effects of temperature on decomposition rates (unitless).}
