% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralDecompOp,DecompOp-method}
\alias{GeneralDecompOp,DecompOp-method}
\title{GeneralDecompOp,DecompOp-method pass through factory}
\arguments{
\item{object}{: of class DecompOp}
}
\description{
This method handles the case that no actual construction is necessary since
the argument is already of a subclass of DecompOp. See the \code{subclasses}
section of \code{\link{DecompOp-class}}
}
