% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{GeneralDecompOp,function-method}
\alias{GeneralDecompOp,function-method}
\title{Automatic description: GeneralDecompOp,function-method}
\usage{
\S4method{GeneralDecompOp}{`function`}(object)
}
\arguments{
\item{object}{object of class:\code{function}. : object of class:\code{function}, no manual documentation}
}
\description{
Automatic description: GeneralDecompOp,function-method
}
