% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralDecompOp,function-method}
\alias{GeneralDecompOp,function-method}
\title{GeneralDecompOp,function-method creates a UnBoundLinDecompOp from a matrix
valued function}
\arguments{
\item{object}{: of class function}
}
\description{
The resulting operator is created by a call to the constructor of class
UnBoundLinDecompOp
}
