% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{GeneralDecompOp,list-method}
\alias{GeneralDecompOp,list-method}
\title{GeneralDecompOp,list-method}
\usage{
\S4method{GeneralDecompOp}{list}(object)
}
\arguments{
\item{object}{object of class:\code{list}. }
}
\description{
GeneralDecompOp,list-method
}
