% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{GeneralDecompOp,list-method}
\alias{GeneralDecompOp,list-method}
\title{GeneralDecompOp,list-method creates a BoundLinDecompOp from a nested list of
a times vector and a list of matrices (a matrix for each time step)}
\arguments{
\item{object}{: of class list}
}
\description{
The resulting operator is creted by a call to the constructor of class
BoundLinDecompOp
}
