% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompOp.R
\name{GeneralDecompOp,matrix-method}
\alias{GeneralDecompOp,matrix-method}
\title{Automatic description: GeneralDecompOp,matrix-method}
\usage{
\S4method{GeneralDecompOp}{matrix}(object)
}
\arguments{
\item{object}{object of class:\code{matrix}. : object of class:\code{matrix}, no manual documentation}
}
\description{
Automatic description: GeneralDecompOp,matrix-method
}
