% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompOp.R
\name{GeneralDecompOp,matrix-method}
\alias{GeneralDecompOp,matrix-method}
\title{GeneralDecompOp,matrix-method}
\usage{
\S4method{GeneralDecompOp}{matrix}(object)
}
\arguments{
\item{object}{object of class:\code{matrix}. }
}
\description{
GeneralDecompOp,matrix-method
}
