\docType{methods}
\name{GeneralDecompOp,list-method}
\title{GeneralDecompOp,list-method}
\alias{GeneralDecompOp,list-method}
\usage{\S4method{GeneralDecompOp}{list}(object)}
\arguments{
\item{object}{: of class list}
}
\description{an a list of matrices (a matrix for each time step)
The resulting operator is creted by a call to the constructor of class
BoundLinDecompOp}
