\docType{methods}
\name{GeneralDecompOp,TimeMap-method}
\title{GeneralDecompOp,TimeMap-method 
 creates a BoundLinDecompOp from a TimeMap object}
\alias{GeneralDecompOp,TimeMap-method}
\usage{\S4method{GeneralDecompOp}{TimeMap}(object= )}
\arguments{
\item{object}{: of class TimeMap}
}
\description{The resulting operator is creted by a call to the constructor of class
BoundLinDecompOp}
