% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompOp.R
\name{GeneralDecompOp,TimeMap-method}
\alias{GeneralDecompOp,TimeMap-method}
\title{Automatic description: GeneralDecompOp,TimeMap-method}
\usage{
\S4method{GeneralDecompOp}{TimeMap}(object)
}
\arguments{
\item{object}{object of class:\code{TimeMap}. : object of class:\code{TimeMap}, no manual documentation}
}
\description{
Automatic description: GeneralDecompOp,TimeMap-method
}
