% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralDecompOp}
\alias{GeneralDecompOp}
\title{A generic factory for subclasses of GeneralDecompOp}
\usage{
GeneralDecompOp(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
The class of the output depends on the provided arguments
}
