% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralDecompOp}
\alias{GeneralDecompOp}
\title{A generic factory for subclasses of GeneralDecompOp}
\usage{
GeneralDecompOp(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
A generic factory for subclasses of GeneralDecompOp
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{GeneralDecompOp,DecompOp-method}}
        \item \code{\link{GeneralDecompOp,function-method}}
        \item \code{\link{GeneralDecompOp,list-method}}
        \item \code{\link{GeneralDecompOp,matrix-method}}
        \item \code{\link{GeneralDecompOp,TimeMap-method}}
      }
}
