\description{no Description}
\name{GeneralDecompOp}
\alias{GeneralDecompOp}
\usage{GeneralDecompOp(object)}
\title{GeneralDecompOp S4 generic}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{GeneralDecompOp,DecompOp-method}}\cr
	\code{\link{GeneralDecompOp,TimeMap-method}}\cr
	\code{\link{GeneralDecompOp,function-method}}\cr
	\code{\link{GeneralDecompOp,list-method}}\cr
	\code{\link{GeneralDecompOp,matrix-method}}\cr}
