\docType{methods}
\name{GeneralInFlux,list-method}
\title{GeneralInFlux,list-method 
 creates a BoundInFlux from a nested list of a times vector and a list of vectors (a vector for each time step)}
\alias{GeneralInFlux,list-method}
\usage{\S4method{GeneralInFlux}{list}(object)}
\arguments{
\item{object}{: of class list}
}
\description{The resulting object is created by a call to the constructor of class
BoundInFlux}
