\docType{methods}
\name{GeneralInFlux,numeric-method}
\title{GeneralInFlux,numeric-method 
 conversion of a vector to an object of class \code{\link{ConstInFlux}}}
\alias{GeneralInFlux,numeric-method}
\usage{\S4method{GeneralInFlux}{numeric}(object= )}
\arguments{
\item{object}{: of class numeric}
}
\description{This method enables the model creating functions to handle constant input streams }
