\docType{methods}
\name{GeneralInFlux,TimeMap-method}
\title{GeneralInFlux,TimeMap-method 
 create a BoundInFlux from a TimeMap object}
\alias{GeneralInFlux,TimeMap-method}
\usage{\S4method{GeneralInFlux}{TimeMap}(object= )}
\arguments{
\item{object}{: of class TimeMap}
}
\description{The method is used to ensure compatibility 
TimeMap class.
The resulting BoundInFlux is created by a call to
the constructor BoundInFlux(object) of that class.}
