\name{GeneralModel}
\alias{GeneralModel}
\title{a general constructor }
\description{create a Model object from different sources}
\usage{GeneralModel(t, A, ivList, inputFluxes, ...)}\arguments{
  \item{t}{
}
  \item{A}{
}
  \item{ivList}{
}
  \item{inputFluxes}{
}
  \item{\dots}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller
<mamueller@bgc-jena.mpg.de>}





