\description{no Description}
\name{GeneralModel_14}
\alias{GeneralModel_14}
\usage{GeneralModel_14(t,
A,
ivList,
initialValF,
inputFluxes,
inputFc,
Fc,
di=-0.0001209681,
lambda=-0.0001209681,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE,
...)}
\title{GeneralModel_14 S4 generic}
\arguments{
\item{t}{see the method arguments for details}
\item{A}{see the method arguments for details}
\item{ivList}{see the method arguments for details}
\item{initialValF}{see the method arguments for details}
\item{inputFluxes}{see the method arguments for details}
\item{inputFc}{see the method arguments for details}
\item{Fc}{see the method arguments for details}
\item{di}{see the method arguments for details}
\item{lambda}{see the method arguments for details}
\item{solverfunc}{The function used by to actually solve the ODE system. This can be \code{\link{deSolve.lsoda.wrapper}} or any other user provided function with the same interface.}
\item{pass}{if TRUE Forces the constructor to create the model even if it is invalid}
\item{...}{see the method arguments for details}
}
\section{Methods}{	\code{\link{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,missing,missing,missing,logical-method}}\cr
	\code{\link{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,logical-method}}\cr
	\code{\link{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,missing-method}}\cr
	\code{\link{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,missing,logical-method}}\cr
	\code{\link{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,missing,missing-method}}\cr
	\code{\link{GeneralModel_14,numeric,ANY,numeric,ANY,ANY,missing,ANY,numeric,missing,missing,missing-method}}\cr}
