% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolId.R
\name{GeneralPoolId,character-method}
\alias{GeneralPoolId,character-method}
\title{GeneralPoolId,character-method}
\usage{
\S4method{GeneralPoolId}{character}(id)
}
\arguments{
\item{id}{object of class:\code{character}. }
}
\description{
GeneralPoolId,character-method
}
