% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralPoolId}
\alias{GeneralPoolId}
\title{GeneralPoolId}
\usage{
GeneralPoolId(id)

GeneralPoolId(id)
}
\arguments{
\item{id}{see method arguments}
}
\description{
GeneralPoolId

GeneralPoolId
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{GeneralPoolId,character-method}}
  \item \code{\link{GeneralPoolId,numeric-method}}
  \item \code{\link{GeneralPoolId,character-method}}
  \item \code{\link{GeneralPoolId,numeric-method}}
  \item \code{\link{GeneralPoolId,character-method}}
  \item \code{\link{GeneralPoolId,numeric-method}}
  \item \code{\link{GeneralPoolId,character-method}}
  \item \code{\link{GeneralPoolId,numeric-method}}
}
}

