% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralPoolId}
\alias{GeneralPoolId}
\title{GeneralPoolId}
\usage{
GeneralPoolId(id)

GeneralPoolId(id)
}
\arguments{
\item{id}{see method arguments}
}
\description{
GeneralPoolId

GeneralPoolId
}
