% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{GeneralPoolId}
\alias{GeneralPoolId}
\title{automatic title}
\usage{
GeneralPoolId(id)

GeneralPoolId(id)
}
\arguments{
\item{id}{see method arguments}
}
\description{
automatic title

automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{GeneralPoolId,character-method}}
        \item \code{\link{GeneralPoolId,numeric-method}}
      }
}
