% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getAccumulatedRelease,Model-method}
\alias{getAccumulatedRelease,Model-method}
\title{getAccumulatedRelease,Model-method}
\usage{
\S4method{getAccumulatedRelease}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. }
}
\description{
getAccumulatedRelease,Model-method
}
