% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getAccumulatedRelease,Model-method}
\alias{getAccumulatedRelease,Model-method}
\title{getAccumulatedRelease,Model-method time integrals of release fluxes per pool}
\arguments{
\item{object}{: of class Model}
}
\description{
The method integrates the release flux of every pool for all times in the
interval specified by the model definition.
}
