\name{getAccumulatedRelease}
\alias{getAccumulatedRelease}
\title{access to the C content of the pools }
\description{This function computes the overall  carbon release of the given model as funtion of time }
\usage{getAccumulatedRelease(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





