% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getAccumulatedRelease}
\alias{getAccumulatedRelease}
\title{S4 generic: getAccumulatedRelease}
\usage{
getAccumulatedRelease(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getAccumulatedRelease
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getAccumulatedRelease,Model-method}}
}
}

