% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getAccumulatedRelease}
\alias{getAccumulatedRelease}
\title{automatic title}
\usage{
getAccumulatedRelease(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getAccumulatedRelease,Model-method}}
      }
}
