\description{no Description}
\value{A n x m matrix of cummulative release fluxes with m columns representing the number of pools, and n rows representing the times as specified by the argument
\code{t} in \code{\link{GeneralModel}} or other model creating functions.}
\name{getAccumulatedRelease}
\alias{getAccumulatedRelease}
\usage{getAccumulatedRelease(object= )}
\title{getAccumulatedRelease S4 generic}
\arguments{
\item{object}{A Model object (e.g. of class Model or Model14)
Have a look at the methods for details.}
}
\section{Methods}{	\code{\link{getAccumulatedRelease,Model-method}}\cr}
