% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{getC,Model_by_PoolNames-method}
\alias{getC,Model_by_PoolNames-method}
\title{automatic title}
\usage{
\S4method{getC}{Model_by_PoolNames}(object)
}
\arguments{
\item{object}{object of class:\code{Model_by_PoolNames}, : : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
