% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_by_PoolNames.R
\name{getC,Model_by_PoolNames-method}
\alias{getC,Model_by_PoolNames-method}
\title{automatic title}
\usage{
\S4method{getC}{Model_by_PoolNames}(object)
}
\arguments{
\item{object}{object of class:\code{Model_by_PoolNames}, no manual documentation}
}
\description{
automatic title
}
