% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getC,NlModel-method}
\alias{getC,NlModel-method}
\title{getC,NlModel-method}
\usage{
\S4method{getC}{NlModel}(object, as.closures = FALSE)
}
\arguments{
\item{object}{object of class:\code{NlModel}. }

\item{as.closures}{}
}
\description{
getC,NlModel-method
}
