% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getC}
\alias{getC}
\title{Generic Function to obtain the contents of the pools for all time steps}
\usage{
getC(object, as.closures = F)
}
\arguments{
\item{object}{see method arguments}

\item{as.closures}{see method arguments}
}
\description{
Generic Function to obtain the contents of the pools for all time steps
}
