% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_14.R
\name{getC14,Model_14-method}
\alias{getC14,Model_14-method}
\title{Automatic description: getC14,Model_14-method}
\usage{
\S4method{getC14}{Model_14}(object)
}
\arguments{
\item{object}{object of class:\code{Model_14}. : object of class:\code{Model_14}, no manual documentation}
}
\description{
Automatic description: getC14,Model_14-method
}
