\name{getC14}
\alias{getC14}
\title{access to the C content of the pools }
\description{This function computes the value for C (mass or concentration ) as function of time}
\usage{getC14(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





