% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getC14}
\alias{getC14}
\title{Generic that yields the ^{14}C content for all pools and all times}
\usage{
getC14(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Generic that yields the ^{14}C content for all pools and all times
}
