\name{getC14}
\alias{getC14}
\usage{getC14(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\title{Calculates the mass of radiocarbon (14C fraction times C stock) in all pools}
\description{This function computes the mass of 14C (14C fraction times C stock) for all pools as function of time.
Have a look at the methods for details.}
