% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\alias{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{getCompartmentalMatrixFunc}{ConstLinDecompOp,ANY,ANY}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
