% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\alias{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\title{Automatic description:
getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}
\usage{
\S4method{getCompartmentalMatrixFunc}{ConstLinDecompOp,ANY,ANY}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. : object of class:\code{ConstLinDecompOp}, no manual
documentation}
}
\description{
Automatic description:
getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method
}
