% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundNonLinDecompOp.R
\name{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp,ANY,ANY-method}
\alias{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp,ANY,ANY-method}
\title{Extract the matrix valued function of time and state vector for the compartmental matrix}
\usage{
\S4method{getCompartmentalMatrixFunc}{UnBoundNonLinDecompOp,ANY,ANY}(object)

\S4method{getCompartmentalMatrixFunc}{UnBoundNonLinDecompOp,ANY,ANY}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
Extract the matrix valued function of time and state vector for the compartmental matrix

automatic title
}
