% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getCompartmentalMatrixFunc}
\alias{getCompartmentalMatrixFunc}
\title{getCompartmentalMatrixFunc}
\usage{
getCompartmentalMatrixFunc(object, timeSymbol, state_variable_names)
}
\arguments{
\item{object}{see method arguments}

\item{timeSymbol}{see method arguments}

\item{state_variable_names}{see method arguments}
}
\description{
getCompartmentalMatrixFunc
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getCompartmentalMatrixFunc,BoundLinDecompOp,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,TransportDecompositionOperator,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp_by_PoolNames,character,character-method}}
  \item \code{\link{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,BoundLinDecompOp,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,TransportDecompositionOperator,ANY,ANY-method}}
  \item \code{\link{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp_by_PoolNames,character,character-method}}
  \item \code{\link{getCompartmentalMatrixFunc,UnBoundNonLinDecompOp,ANY,ANY-method}}
}
}

