% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getConstantCompartmentalMatrix,ConstLinDecompOp-method}
\alias{getConstantCompartmentalMatrix,ConstLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{getConstantCompartmentalMatrix}{ConstLinDecompOp}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
