% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getConstantCompartmentalMatrix}
\alias{getConstantCompartmentalMatrix}
\title{automatic title}
\usage{
getConstantCompartmentalMatrix(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
