% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getConstantInFluxVector}
\alias{getConstantInFluxVector}
\title{automatic title}
\usage{
getConstantInFluxVector(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getConstantInFluxVector,ConstInFluxes-method}}
      }
}
