% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getConstantInternalFluxRateList_by_PoolIndex,ConstLinDecompOp-method}
\alias{getConstantInternalFluxRateList_by_PoolIndex,ConstLinDecompOp-method}
\title{Constant internal flux rate list by pool index}
\usage{
\S4method{getConstantInternalFluxRateList_by_PoolIndex}{ConstLinDecompOp}(object)
}
\arguments{
\item{object}{a ConstLinDecomOp object}
}
\description{
Constant internal flux rate list by pool index
}
