% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\name{getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method}
\alias{getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method}
\title{getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method}
\usage{
\S4method{getConstLinDecompOp}{ConstLinDecompOpWithLinearScalarFactor}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOpWithLinearScalarFactor}. }
}
\description{
getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method
}
