% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getConstLinDecompOp}
\alias{getConstLinDecompOp}
\title{getConstLinDecompOp}
\usage{
getConstLinDecompOp(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
getConstLinDecompOp
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method}}
  \item \code{\link{getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method}}
}
}

