% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getCumulativeC,NlModel-method}
\alias{getCumulativeC,NlModel-method}
\title{automatic title}
\usage{
\S4method{getCumulativeC}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}, no manual documentation}
}
\description{
automatic title
}
