% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getDecompOp,Model-method}
\alias{getDecompOp,Model-method}
\title{Extract the Compartmental Operator}
\usage{
\S4method{getDecompOp}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}, no manual documentation}
}
\description{
Extract the Compartmental Operator
}
