% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getDecompOp,Model-method}
\alias{getDecompOp,Model-method}
\title{getDecompOp,Model-method}
\arguments{
\item{object}{: of class Model}
}
\description{
Extracts the Operator from a model object
}
