% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getDecompOp,NlModel-method}
\alias{getDecompOp,NlModel-method}
\title{getDecompOp,NlModel-method}
\arguments{
\item{object}{: of class NlModel}
}
\description{
This method returns the Decomposition Operator of the Model
}
