% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getDecompOp}
\alias{getDecompOp}
\title{automatic title}
\usage{
getDecompOp(object)
}
\arguments{
\item{object}{: see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getDecompOp,Model-method}}
  \item \code{\link{getDecompOp,NlModel-method}}
}
}

