% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getDotOut,TransportDecompositionOperator-method}
\alias{getDotOut,TransportDecompositionOperator-method}
\title{getDotOut,TransportDecompositionOperator-method}
\usage{
\S4method{getDotOut}{TransportDecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}. }
}
\description{
getDotOut,TransportDecompositionOperator-method
}
