% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_14.R
\name{getF14,Model_14-method}
\alias{getF14,Model_14-method}
\title{automatic title}
\usage{
\S4method{getF14}{Model_14}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
