% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getF14}
\alias{getF14}
\title{Generic that yields the ^{14}C fraction for the content all pools and all times}
\usage{
getF14(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Generic that yields the ^{14}C fraction for the content all pools and all times
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getF14,Model_14-method}}
      }
}
