\name{getF14}
\alias{getF14}
\usage{getF14(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\title{Calculates the 14C fraction of all pools}
\description{This function computes the radiocarbon fraction for all pools as funtion of time.
Have a look at the methods for details.}
