% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getF14C}
\alias{getF14C}
\title{Generic that yields the 14C fraction for the cumulative content of all pools and all times}
\usage{
getF14C(object)
}
\arguments{
\item{object}{a SoilR object of class Model 14}
}
\description{
Generic that yields the 14C fraction for the cumulative content of all pools and all times
}
