\name{getF14C}
\alias{getF14C}
\usage{getF14C(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getF14C,Model_14-method}}\cr}
\title{Calculates the average radiocarbon fraction weighted by the mass of carbon}
\description{This function calculates the average radiocarbon fraction weighted by the mass of carbon at each time step 
Have a look at the methods for details.}
