% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getF14C}
\alias{getF14C}
\title{Generic that yields the ^{14}C fraction for the cumulative content of all pools and all times}
\usage{
getF14C(object)
}
\description{
Generic that yields the ^{14}C fraction for the cumulative content of all pools and all times
}
