% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getF14R}
\alias{getF14R}
\title{Generic that yields the ^{14}C fraction for the release flux of all pools and all times}
\usage{
getF14R(object)
}
\description{
Generic that yields the ^{14}C fraction for the release flux of all pools and all times
}
