\name{getFormat}
\alias{getFormat}
\title{extract the format from an object that contains one}
\description{The function returns a format string that describes the format the data is given in
The more detailed information is to be found in the methods		 }
\usage{getFormat(object)}
\arguments{
  \item{object}{usually an object of a subclass of TimeMap}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





