% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstInFluxes.R
\name{getFunctionDefinition,ConstInFluxes-method}
\alias{getFunctionDefinition,ConstInFluxes-method}
\title{automatic title}
\usage{
\S4method{getFunctionDefinition}{ConstInFluxes}(object)
}
\arguments{
\item{object}{object of class:\code{ConstInFluxes}, no manual documentation}
}
\description{
automatic title
}
