% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\docType{methods}
\name{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}
\alias{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}
\title{helper function}
\usage{

  \S4method{getFunctionDefinition}{ConstLinDecompOpWithLinearScalarFactor}(object)
}
\description{
helper function
}
