% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompositionOperator_deprecated_in_1.2.R
\name{getFunctionDefinition,DecompositionOperator-method}
\alias{getFunctionDefinition,DecompositionOperator-method}
\title{getFunctionDefinition,DecompositionOperator-method}
\usage{
\S4method{getFunctionDefinition}{DecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{DecompositionOperator}. }
}
\description{
getFunctionDefinition,DecompositionOperator-method
}
