% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFlux_by_PoolIndex.R,
%   R/InFluxList_by_PoolName.R
\name{getFunctionDefinition,InFluxList_by_PoolName-method}
\alias{getFunctionDefinition,InFluxList_by_PoolName-method}
\title{automatic title}
\usage{
\S4method{getFunctionDefinition}{InFluxList_by_PoolName}(object, timeSymbol, poolNames)

\S4method{getFunctionDefinition}{InFluxList_by_PoolName}(object, timeSymbol, poolNames)
}
\arguments{
\item{object}{object of class:\code{InFluxList_by_PoolName}, no manual documentation}

\item{timeSymbol}{no manual documentation}

\item{poolNames}{no manual documentation}
}
\description{
automatic title

automatic title
}
