% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getFunctionDefinition,TransportDecompositionOperator-method}
\alias{getFunctionDefinition,TransportDecompositionOperator-method}
\title{getFunctionDefinition,TransportDecompositionOperator-method}
\usage{
\S4method{getFunctionDefinition}{TransportDecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}. }
}
\description{
getFunctionDefinition,TransportDecompositionOperator-method
}
