% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{getFunctionDefinition,UnBoundInFluxes-method}
\alias{getFunctionDefinition,UnBoundInFluxes-method}
\title{getFunctionDefinition,UnBoundInFluxes-method}
\usage{
\S4method{getFunctionDefinition}{UnBoundInFluxes}(object)
}
\arguments{
\item{object}{object of class:\code{UnBoundInFluxes}. }
}
\description{
getFunctionDefinition,UnBoundInFluxes-method
}
