% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{getFunctionDefinition,UnBoundInFluxes-method}
\alias{getFunctionDefinition,UnBoundInFluxes-method}
\title{automatic title}
\usage{
\S4method{getFunctionDefinition}{UnBoundInFluxes}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
automatic title
}
