% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{getFunctionDefinition,UnBoundLinDecompOp-method}
\alias{getFunctionDefinition,UnBoundLinDecompOp-method}
\title{Extracts the time dependent matrix valued function (compartmental matrix)}
\usage{
\S4method{getFunctionDefinition}{UnBoundLinDecompOp}(object)
}
\arguments{
\item{object}{object of class:\code{UnBoundLinDecompOp}, no manual documentation}
}
\description{
Extracts the time dependent matrix valued function (compartmental matrix)
}
\seealso{
Other UnBoundLinDecompOp_constructor: 
\code{\link{UnBoundLinDecompOp,function-method}}
}
\concept{UnBoundLinDecompOp_constructor}
