\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\title{getFunctionDefinition}
\description{extract the function definition (the R-function) from the argument}
\usage{getFunctionDefinition(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





