\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\usage{getFunctionDefinition(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getFunctionDefinition,ConstInFlux-method}}\cr
	\code{\link{getFunctionDefinition,ConstLinDecompOp-method}}\cr
	\code{\link{getFunctionDefinition,DecompositionOperator-method}}\cr
	\code{\link{getFunctionDefinition,TimeMap-method}}\cr
	\code{\link{getFunctionDefinition,TransportDecompositionOperator-method}}\cr}
\title{getFunctionDefinition}
\description{Extracts the function definition (the R-function) from the argument
Have a look at the methods for details.}
