% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\title{getFunctionDefinition}
\usage{
getFunctionDefinition(object, timeSymbol, poolNames, numberOfPools)
}
\arguments{
\item{object}{see method arguments}

\item{timeSymbol}{see method arguments}

\item{poolNames}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
getFunctionDefinition
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getFunctionDefinition,ConstInFluxes-method}}
  \item \code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
  \item \code{\link{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}}
  \item \code{\link{getFunctionDefinition,DecompositionOperator-method}}
  \item \code{\link{getFunctionDefinition,InFluxList_by_PoolIndex-method}}
  \item \code{\link{getFunctionDefinition,InFluxList_by_PoolName-method}}
  \item \code{\link{getFunctionDefinition,StateDependentInFluxVector-method}}
  \item \code{\link{getFunctionDefinition,TimeMap-method}}
  \item \code{\link{getFunctionDefinition,TransportDecompositionOperator-method}}
  \item \code{\link{getFunctionDefinition,UnBoundInFluxes-method}}
  \item \code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}
  \item \code{\link{getFunctionDefinition,ConstInFluxes-method}}
  \item \code{\link{getFunctionDefinition,ConstLinDecompOp-method}}
  \item \code{\link{getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method}}
  \item \code{\link{getFunctionDefinition,DecompositionOperator-method}}
  \item \code{\link{getFunctionDefinition,InFluxList_by_PoolIndex-method}}
  \item \code{\link{getFunctionDefinition,InFluxList_by_PoolName-method}}
  \item \code{\link{getFunctionDefinition,StateDependentInFluxVector-method}}
  \item \code{\link{getFunctionDefinition,TimeMap-method}}
  \item \code{\link{getFunctionDefinition,TransportDecompositionOperator-method}}
  \item \code{\link{getFunctionDefinition,UnBoundInFluxes-method}}
  \item \code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}
}
}

