\description{no Description}
\name{getFunctionDefinition}
\alias{getFunctionDefinition}
\usage{getFunctionDefinition(object= )}
\title{getFunctionDefinition S4 generic}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getFunctionDefinition,ConstInFlux-method}}\cr
	\code{\link{getFunctionDefinition,ConstLinDecompOp-method}}\cr
	\code{\link{getFunctionDefinition,DecompositionOperator-method}}\cr
	\code{\link{getFunctionDefinition,TimeMap-method}}\cr
	\code{\link{getFunctionDefinition,TransportDecompositionOperator-method}}\cr
	\code{\link{getFunctionDefinition,UnBoundInFlux-method}}\cr
	\code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}\cr}
