% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getInFluxes,Model-method}
\alias{getInFluxes,Model-method}
\title{Extract the InFluxes as provided during creation of the model}
\usage{
\S4method{getInFluxes}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}, A modelrun as produced by the constructors: 
 \code{\link{Model}},
 \code{\link{Model_by_PoolNames}},
 \code{\link{Model_14}}
the function \code{\link{GeneralModel}} or the functions listed in 
\code{\link{predefinedModels}}. \cr
A model represents the initial value problem (IVP)  
for the contents of the pool consisting of 
\itemize{
 \item{ The initial values of the pool content}
 \item{ The system of ordinary differential oqations, as dictated by the fluxes}
 \item{ The times for which the solution of the IVP is evaluated.}
}}
}
\description{
Since the influxes had to be provided to create the model this method
yields no new information that can not be obtained simpler.  
It is usually called internally by other functions.
}
