% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getInFluxes,NlModel-method}
\alias{getInFluxes,NlModel-method}
\title{getInFluxes,NlModel-method}
\usage{
\S4method{getInFluxes}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}. }
}
\description{
getInFluxes,NlModel-method
}
