% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getInFluxes}
\alias{getInFluxes}
\title{S4 generic: getInFluxes}
\usage{
getInFluxes(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getInFluxes
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getInFluxes,Model-method}}
  \item \code{\link{getInFluxes,NlModel-method}}
}
}

