% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getInitialValues,NlModel-method}
\alias{getInitialValues,NlModel-method}
\title{Automatic description: getInitialValues,NlModel-method}
\usage{
\S4method{getInitialValues}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}. : object of class:\code{NlModel}, no manual documentation}
}
\description{
Automatic description: getInitialValues,NlModel-method
}
