% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getLaggingTimeRange}
\alias{getLaggingTimeRange}
\title{S4 generic: getLaggingTimeRange}
\usage{
getLaggingTimeRange(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getLaggingTimeRange
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{}}
}
}

