% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOpWithLinearScaleFactor.R
\name{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}
\alias{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}
\title{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}
\usage{
\S4method{getLinearScaleFactor}{ConstLinDecompOpWithLinearScalarFactor}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOpWithLinearScalarFactor}. }
}
\description{
getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method
}
