% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getLinearScaleFactor}
\alias{getLinearScaleFactor}
\title{getLinearScaleFactor}
\usage{
getLinearScaleFactor(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
getLinearScaleFactor
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}}
  \item \code{\link{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}}
}
}

