% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getLinearScaleFactor}
\alias{getLinearScaleFactor}
\title{automatic title}
\usage{
getLinearScaleFactor(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method}}
      }
}
