% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getMeanTransitTime,ConstLinDecompOp-method}
\alias{getMeanTransitTime,ConstLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{getMeanTransitTime}{ConstLinDecompOp}(object, inputDistribution)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}, no manual documentation}

\item{inputDistribution}{no manual documentation}
}
\description{
automatic title
}
