\name{getMeanTransitTime}
\alias{getMeanTransitTime}
\title{access to the mean transit time }
\description{This function computes the Expected Value of the transit time 
by integrating the product of t and the TransitTimeDistributionDensity}
\usage{getMeanTransitTime(object, inputDistribution)}
\arguments{
  \item{object}{
}
  \item{inputDistribution}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





