% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getMeanTransitTime}
\alias{getMeanTransitTime}
\title{automatic title}
\usage{
getMeanTransitTime(object, inputDistribution)
}
\arguments{
\item{object}{see method arguments}

\item{inputDistribution}{see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getMeanTransitTime,ConstLinDecompOp-method}}
}
}

