\name{getMeanTransitTime}
\alias{getMeanTransitTime}
\title{access to the mean transit time }

\usage{getMeanTransitTime(object, inputDistribution)}
\arguments{
  \item{object}{
}
  \item{inputDistribution}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





