% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getNumberOfPools,TransportDecompositionOperator-method}
\alias{getNumberOfPools,TransportDecompositionOperator-method}
\title{getNumberOfPools,TransportDecompositionOperator-method}
\usage{
\S4method{getNumberOfPools}{TransportDecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}. }
}
\description{
getNumberOfPools,TransportDecompositionOperator-method
}
