% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getOutputFluxes}
\alias{getOutputFluxes}
\title{Generic Function to obtain the fluxes out of of the pools}
\usage{
getOutputFluxes(object, as.closures = F)
}
\arguments{
\item{object}{see method arguments}

\item{as.closures}{see method arguments}
}
\description{
Generic Function to obtain the fluxes out of of the pools
}
