% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\alias{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\title{automatic title}
\usage{
\S4method{getOutputReceivers}{TransportDecompositionOperator,numeric}(object, i)
}
\arguments{
\item{object}{no manual documentation}

\item{i}{no manual documentation}
}
\description{
automatic title
}
