% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\alias{getOutputReceivers,TransportDecompositionOperator,numeric-method}
\title{Automatic description:
getOutputReceivers,TransportDecompositionOperator,numeric-method}
\usage{
\S4method{getOutputReceivers}{TransportDecompositionOperator,numeric}(object, i)
}
\arguments{
\item{object}{object of class:\code{TransportDecompositionOperator}. : object of class:\code{TransportDecompositionOperator}, no
manual documentation}

\item{i}{object of class:\code{numeric}. : object of class:\code{numeric}, no manual documentation}
}
\description{
Automatic description:
getOutputReceivers,TransportDecompositionOperator,numeric-method
}
