% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getOutputReceivers}
\alias{getOutputReceivers}
\title{automatic title}
\usage{
getOutputReceivers(object, i)
}
\arguments{
\item{object}{see method arguments}

\item{i}{see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getOutputReceivers,TransportDecompositionOperator,numeric-method}}
}
}

