% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getParticleMonteCarloSimulator}
\alias{getParticleMonteCarloSimulator}
\title{automatic title}
\usage{
getParticleMonteCarloSimulator(object)
}
\arguments{
\item{object}{: see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getParticleMonteCarloSimulator,NlModel-method}}
}
}

