% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getReleaseFlux}
\alias{getReleaseFlux}
\title{Generic Function to obtain the vector of release fluxes out of the pools for all times.}
\usage{
getReleaseFlux(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Generic Function to obtain the vector of release fluxes out of the pools for all times.
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getReleaseFlux,Model-method}}
        \item \code{\link{getReleaseFlux,Model_by_PoolNames-method}}
        \item \code{\link{getReleaseFlux,NlModel-method}}
      }
}
