\description{no Description}
\value{A matrix. Every column represents a pool and every row a point in time}
\name{getReleaseFlux}
\alias{getReleaseFlux}
\usage{getReleaseFlux(object)}
\title{This function }
\arguments{
\item{object}{An model object (the actual class depends on the method e.g. Model or  Model14}
}
\section{Methods}{	\code{\link{getReleaseFlux,Model-method}}\cr
	\code{\link{getReleaseFlux,NlModel-method}}\cr}
