\name{getReleaseFlux14}
\alias{getReleaseFlux14}
\title{access to the C content of the pools }
\description{This function computes the \eqn{^{14}C}{14C} release of the given model as funtion of time }
\usage{getReleaseFlux14(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





