% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getReleaseFlux14}
\alias{getReleaseFlux14}
\title{Generic that yields the 14C fraction in the release flux}
\usage{
getReleaseFlux14(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
Generic that yields the 14C fraction in the release flux
}
