% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getReleaseFlux14}
\alias{getReleaseFlux14}
\title{automatic title}
\usage{
getReleaseFlux14(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{getReleaseFlux14,Model_14-method}}
      }
}
