\name{getReleaseFlux14}
\alias{getReleaseFlux14}
\usage{getReleaseFlux14(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getReleaseFlux14,Model_14-method}}\cr}
\title{Calculates the mass of radiocarbon in the release flux (14C fraction times release flux)}
\description{This function computes the mass of radiocarbon in the release flux (14C fraction times release flux) as a function of time.
Have a look at the methods for details.}
