% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getRightHandSideOfODE}
\alias{getRightHandSideOfODE}
\title{S4 generic: getRightHandSideOfODE}
\usage{
getRightHandSideOfODE(object, timeSymbol, poolNames, numberOfPools)
}
\arguments{
\item{object}{see method arguments}

\item{timeSymbol}{see method arguments}

\item{poolNames}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
S4 generic: getRightHandSideOfODE
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getRightHandSideOfODE,Model-method}}
  \item \code{\link{getRightHandSideOfODE,Model_by_PoolNames-method}}
}
}

