% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getTimeRange,ConstLinDecompOp-method}
\alias{getTimeRange,ConstLinDecompOp-method}
\title{Automatic description: getTimeRange,ConstLinDecompOp-method}
\usage{
\S4method{getTimeRange}{ConstLinDecompOp}(object)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. : object of class:\code{ConstLinDecompOp}, no manual
documentation}
}
\description{
Automatic description: getTimeRange,ConstLinDecompOp-method
}
