% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{getTimeRange,DecompositionOperator-method}
\alias{getTimeRange,DecompositionOperator-method}
\title{getTimeRange,DecompositionOperator-method ask for the boundaries of the
underlying time interval}
\arguments{
\item{object}{: of class DecompositionOperator}
}
\description{
The method returns the time range of the given object It is ( probably
mostly ) used internally to make sure that time dependent functions
retrieved from data are not used outside the interval where they are valid.
}
