% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompositionOperator_deprecated_in_1.2.R
\name{getTimeRange,DecompositionOperator-method}
\alias{getTimeRange,DecompositionOperator-method}
\title{getTimeRange,DecompositionOperator-method}
\usage{
\S4method{getTimeRange}{DecompositionOperator}(object)
}
\arguments{
\item{object}{object of class:\code{DecompositionOperator}. }
}
\description{
getTimeRange,DecompositionOperator-method
}
