% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{getTimeRange,UnBoundInFluxes-method}
\alias{getTimeRange,UnBoundInFluxes-method}
\title{getTimeRange,UnBoundInFluxes-method}
\usage{
\S4method{getTimeRange}{UnBoundInFluxes}(object)
}
\arguments{
\item{object}{object of class:\code{UnBoundInFluxes}. }
}
\description{
getTimeRange,UnBoundInFluxes-method
}
