% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{getTimeRange,UnBoundLinDecompOp-method}
\alias{getTimeRange,UnBoundLinDecompOp-method}
\title{Extracts the time interval for which the function is valid.}
\usage{
\S4method{getTimeRange}{UnBoundLinDecompOp}(object)
}
\arguments{
\item{object}{no manual documentation}
}
\description{
Extracts the time interval for which the function is valid.
}
