\description{no Description}
\name{getTimeRange}
\alias{getTimeRange}
\usage{getTimeRange(object= )}
\title{getTimeRange S4 generic}
\arguments{
\item{object}{see the method arguments for details}
}
\section{Methods}{	\code{\link{getTimeRange,ConstInFlux-method}}\cr
	\code{\link{getTimeRange,ConstLinDecompOp-method}}\cr
	\code{\link{getTimeRange,DecompositionOperator-method}}\cr
	\code{\link{getTimeRange,TimeMap-method}}\cr
	\code{\link{getTimeRange,UnBoundInFlux-method}}\cr
	\code{\link{getTimeRange,UnBoundLinDecompOp-method}}\cr}
