% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTimeRange}
\alias{getTimeRange}
\title{S4 generic: getTimeRange}
\usage{
getTimeRange(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getTimeRange
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getTimeRange,ConstInFluxes-method}}
  \item \code{\link{getTimeRange,ConstLinDecompOp-method}}
  \item \code{\link{getTimeRange,ConstLinDecompOpWithLinearScalarFactor-method}}
  \item \code{\link{getTimeRange,DecompositionOperator-method}}
  \item \code{\link{getTimeRange,TimeMap-method}}
  \item \code{\link{getTimeRange,UnBoundInFluxes-method}}
  \item \code{\link{getTimeRange,UnBoundLinDecompOp-method}}
}
}

