% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getTimes,Model-method}
\alias{getTimes,Model-method}
\title{Extract the times vector}
\usage{
\S4method{getTimes}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}, no manual documentation}
}
\description{
Extract the times vector
}
