% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getTimes,Model-method}
\alias{getTimes,Model-method}
\title{Automatic description: getTimes,Model-method}
\usage{
\S4method{getTimes}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. : object of class:\code{Model}, no manual documentation}
}
\description{
Automatic description: getTimes,Model-method
}
