\name{getTimes}
\alias{getTimes}
\title{access to the time values the model solution is sougth for }
\description{This functions extracts the times argument}
\usage{getTimes(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





