\name{getTimes}
\alias{getTimes}
\title{extract the times argumen}
\description{This functions extracts the times argument from an argument of class NlModel}
\usage{getTimes(object)}
\arguments{
  \item{object}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





