% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTimes}
\alias{getTimes}
\title{S4 generic: getTimes}
\usage{
getTimes(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
S4 generic: getTimes
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getTimes,Model-method}}
  \item \code{\link{getTimes,NlModel-method}}
}
}

