% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{getTransferCoefficients,NlModel-method}
\alias{getTransferCoefficients,NlModel-method}
\title{getTransferCoefficients,NlModel-method}
\usage{
\S4method{getTransferCoefficients}{NlModel}(object, as.closures = F)
}
\arguments{
\item{object}{object of class:\code{NlModel}. }

\item{as.closures}{found automatically by inspection, not documented yet}
}
\description{
getTransferCoefficients,NlModel-method
}
