% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTransferCoefficients}
\alias{getTransferCoefficients}
\title{S4 generic: getTransferCoefficients}
\usage{
getTransferCoefficients(object, as.closures = F)

getTransferCoefficients(object, as.closures = F)
}
\arguments{
\item{object}{see method arguments}

\item{as.closures}{see method arguments}
}
\description{
S4 generic: getTransferCoefficients

S4 generic: getTransferCoefficients
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getTransferCoefficients,NlModel-method}}
  \item \code{\link{getTransferCoefficients,TransportDecompositionOperator-method}}
  \item \code{\link{getTransferCoefficients,NlModel-method}}
  \item \code{\link{getTransferCoefficients,TransportDecompositionOperator-method}}
}
}

