% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTransferCoefficients}
\alias{getTransferCoefficients}
\title{getTransferCoefficients}
\usage{
getTransferCoefficients(object, as.closures = F)

getTransferCoefficients(object, as.closures = F)
}
\arguments{
\item{object}{see method arguments}

\item{as.closures}{see method arguments}
}
\description{
getTransferCoefficients

getTransferCoefficients
}
