% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTransferMatrix}
\alias{getTransferMatrix}
\title{automatic title}
\usage{
getTransferMatrix(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getTransferMatrix,ANY-method}}
}
}

