% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getTransferMatrixFunc}
\alias{getTransferMatrixFunc}
\title{getTransferMatrixFunc}
\usage{
getTransferMatrixFunc(object)
}
\arguments{
\item{object}{see method arguments}
}
\description{
getTransferMatrixFunc
}
