% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\alias{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\title{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\usage{
\S4method{getTransitTimeDistributionDensity}{ConstLinDecompOp}(object, inputDistribution, times)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. }

\item{inputDistribution}{}

\item{times}{}
}
\description{
getTransitTimeDistributionDensity,ConstLinDecompOp-method
}
