% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\alias{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{getTransitTimeDistributionDensity}{ConstLinDecompOp}(object, inputDistribution, times)
}
\arguments{
\item{object}{no manual documentation}

\item{inputDistribution}{no manual documentation}

\item{times}{no manual documentation}
}
\description{
automatic title
}
