% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\alias{getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\title{Automatic description:
getTransitTimeDistributionDensity,ConstLinDecompOp-method}
\usage{
\S4method{getTransitTimeDistributionDensity}{ConstLinDecompOp}(object, inputDistribution, times)
}
\arguments{
\item{object}{object of class:\code{ConstLinDecompOp}. : object of class:\code{ConstLinDecompOp}, no manual
documentation}

\item{inputDistribution}{found automatically by inspection, not documented yet}

\item{times}{found automatically by inspection, not documented yet}
}
\description{
Automatic description:
getTransitTimeDistributionDensity,ConstLinDecompOp-method
}
