% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{getValues}
\alias{getValues}
\title{automatic title}
\usage{
getValues(object)
}
\arguments{
\item{object}{: see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{getValues,ConstFc-method}}
}
}

