\name{getValues}
\alias{getValues}
\title{extract numeric values from an object that contains additional informations like format or even units}
\description{The function returns the actaul  number(s) 
The more detailed information is to be found in the methods		 }
\usage{getValues(object)}
\arguments{
  \item{object}{usually an object of a class implemented by soilR e.g. a startvalue for the 14C fraction.}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





