\title{InFlux S4 class}
\description{no Description}
\name{InFlux-class}
\alias{InFlux-class}
\docType{class}
\section{Methods}{  \describe{
    \item{InFluxSubClassInstance}{\code{signature(object = "InFlux")}: ... } \code{\link{InFluxSubClassInstance,InFlux-method}}
	 }}
\section{Subclasses}{\describe{
	\code{\link{BoundInFlux-class}}\cr
	\code{\link{ConstInFlux-class}}\cr
}}
\section{Constructors found by naming convention}{Since the class is virtual it can not be instanciated directly, but a function:
	\code{\link{InFluxSubClassInstance}}\cr
has been found, that produces instances of subclasses.
             Please also look at constructors of non virtual subclasses }
