\name{InFlux}
\alias{InFlux}
\usage{InFlux(object)}
\arguments{
\item{object}{see the method arguments for details}
}
\title{Generic constructor and converter }
\description{If the argument is already of a subclass of class InFlux
the function returns the unchanged object.
Otherwise, it creates an object of a subclass of InFlux.
Note that the actual class of the output depends on the argument.
For examples, please look at the methods of this function.}
