% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,ConstantInFluxList_by_PoolIndex-method}
\alias{InFluxes,ConstantInFluxList_by_PoolIndex-method}
\title{automatic title}
\usage{
\S4method{InFluxes}{ConstantInFluxList_by_PoolIndex}(object, numberOfPools)
}
\arguments{
\item{object}{no manual documentation}

\item{numberOfPools}{no manual documentation}
}
\description{
automatic title
}
