% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,numeric-method}
\alias{InFluxes,numeric-method}
\title{Automatic description: InFluxes,numeric-method}
\usage{
\S4method{InFluxes}{numeric}(object)
}
\arguments{
\item{object}{object of class:\code{numeric}. : object of class:\code{numeric}, no manual documentation}
}
\description{
Automatic description: InFluxes,numeric-method
}
