% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,numeric-method}
\alias{InFluxes,numeric-method}
\title{InFluxes,numeric-method}
\usage{
\S4method{InFluxes}{numeric}(object)
}
\arguments{
\item{object}{object of class:\code{numeric}. }
}
\description{
InFluxes,numeric-method
}
