% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxes.R
\name{InFluxes,TimeMap-method}
\alias{InFluxes,TimeMap-method}
\title{Automatic description: InFluxes,TimeMap-method}
\usage{
\S4method{InFluxes}{TimeMap}(object)
}
\arguments{
\item{object}{object of class:\code{TimeMap}. : object of class:\code{TimeMap}, no manual documentation}
}
\description{
Automatic description: InFluxes,TimeMap-method
}
