% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InFluxes}
\alias{InFluxes}
\title{A generic factory for subclasses of \linkS4class{InFluxes}}
\usage{
InFluxes(object, numberOfPools)
}
\arguments{
\item{object}{see method arguments}

\item{numberOfPools}{see method arguments}
}
\description{
A generic factory for subclasses of \linkS4class{InFluxes}
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{InFluxes,ConstantInFluxList_by_PoolIndex-method}}
        \item \code{\link{InFluxes,function-method}}
        \item \code{\link{InFluxes,InFluxes-method}}
        \item \code{\link{InFluxes,list-method}}
        \item \code{\link{InFluxes,numeric-method}}
        \item \code{\link{InFluxes,StateIndependentInFluxList_by_PoolIndex-method}}
        \item \code{\link{InFluxes,TimeMap-method}}
      }
}
