% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InFluxList_by_PoolIndex.R
\name{InFluxList_by_PoolIndex,list-method}
\alias{InFluxList_by_PoolIndex,list-method}
\title{constructor from a normal list}
\usage{
\S4method{InFluxList_by_PoolIndex}{list}(object)
}
\arguments{
\item{object}{object of class:\code{list}, no manual documentation}
}
\description{
after checking the elements
}
