% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{InFluxList_by_PoolName-class}
\alias{InFluxList_by_PoolName-class}
\title{Class for a list of influxes indexed by the names of the target pools}
\description{
Class for a list of influxes indexed by the names of the target pools
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{InFluxList_by_PoolName} in their signature:}{
      \itemize{
        \item \code{\link{by_PoolIndex,InFluxList_by_PoolName,character,character-method}}
        \item \code{\link{getFunctionDefinition,InFluxList_by_PoolName-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{InFluxList_by_PoolName} in their signature:}{
    \subsection{superclass list}{
      \itemize{
        \item \code{\link{ConstantInFluxList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolName,list-method}}
        \item \code{\link{ConstantOutFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantOutFluxRateList_by_PoolName,list-method}}
        \item \code{\link{GeneralDecompOp,list-method}}
        \item \code{\link{InFluxes,list-method}}
        \item \code{\link{InFluxList_by_PoolIndex,list-method}}
        \item \code{\link{InFluxList_by_PoolName,list-method}}
        \item \code{\link{InternalFluxList_by_PoolIndex,list-method}}
        \item \code{\link{InternalFluxList_by_PoolName,list-method}}
        \item \code{\link{OutFluxList_by_PoolIndex,list-method}}
        \item \code{\link{OutFluxList_by_PoolName,list-method}}
        \item \code{\link{StateIndependentInFluxList_by_PoolIndex,list-method}}
        \item \code{\link{TimeMap,list,missing,missing,missing,missing-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,list-method}}
      }
    }
  }
}
