% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{initialize,ConstLinDecompOp-method}
\alias{initialize,ConstLinDecompOp-method}
\title{Automatic description: initialize,ConstLinDecompOp-method}
\usage{
\S4method{initialize}{ConstLinDecompOp}(.Object, mat = matrix())
}
\arguments{
\item{.Object}{object of class:\code{ConstLinDecompOp}. : object of class:\code{ConstLinDecompOp}, no manual
documentation}

\item{mat}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,ConstLinDecompOp-method
}
