% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstLinDecompOp.R
\name{initialize,ConstLinDecompOp-method}
\alias{initialize,ConstLinDecompOp-method}
\title{initialize,ConstLinDecompOp-method}
\usage{
\S4method{initialize}{ConstLinDecompOp}(.Object, mat = matrix())
}
\arguments{
\item{.Object}{object of class:\code{ConstLinDecompOp}. }

\item{mat}{}
}
\description{
initialize,ConstLinDecompOp-method
}
