% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompositionOperator_deprecated_in_1.2.R
\name{initialize,DecompositionOperator-method}
\alias{initialize,DecompositionOperator-method}
\title{Automatic description: initialize,DecompositionOperator-method}
\usage{
\S4method{initialize}{DecompositionOperator}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  map = function(t) {     t },
  lag = 0
)
}
\arguments{
\item{.Object}{object of class:\code{DecompositionOperator}. : object of class:\code{DecompositionOperator}, no manual
documentation}

\item{starttime}{found automatically by inspection, not documented yet}

\item{endtime}{found automatically by inspection, not documented yet}

\item{map}{found automatically by inspection, not documented yet}

\item{lag}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,DecompositionOperator-method
}
