% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecompositionOperator_deprecated_in_1.2.R
\name{initialize,DecompositionOperator-method}
\alias{initialize,DecompositionOperator-method}
\title{initialize,DecompositionOperator-method}
\usage{
\S4method{initialize}{DecompositionOperator}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  map = function(t) {     t },
  lag = 0
)
}
\arguments{
\item{.Object}{object of class:\code{DecompositionOperator}. }

\item{starttime}{}

\item{endtime}{}

\item{map}{}

\item{lag}{}
}
\description{
initialize,DecompositionOperator-method
}
