% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{initialize,Model-method}
\alias{initialize,Model-method}
\title{Automatic description: initialize,Model-method}
\usage{
\S4method{initialize}{Model}(
  .Object,
  times = c(0, 1),
  mat = ConstLinDecompOp(matrix(nrow = 1, ncol = 1, 0)),
  initialValues = numeric(),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{Model}. : object of class:\code{Model}, no manual documentation}

\item{times}{found automatically by inspection, not documented yet}

\item{mat}{found automatically by inspection, not documented yet}

\item{initialValues}{found automatically by inspection, not documented yet}

\item{inputFluxes}{found automatically by inspection, not documented yet}

\item{solverfunc}{found automatically by inspection, not documented yet}

\item{pass}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,Model-method
}
