% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_14.R
\name{initialize,Model_14-method}
\alias{initialize,Model_14-method}
\title{initialize,Model_14-method}
\usage{
\S4method{initialize}{Model_14}(
  .Object,
  times = c(0, 1),
  mat = ConstLinDecompOp(matrix(nrow = 1, ncol = 1, 0)),
  initialValues = numeric(),
  initialValF = ConstFc(values = c(0), format = "Delta14C"),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  c14Fraction = BoundFc(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) }, 0, 1),
  c14DecayRate = 0,
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{Model_14}. }

\item{times}{}

\item{mat}{}

\item{initialValues}{}

\item{initialValF}{}

\item{inputFluxes}{}

\item{c14Fraction}{}

\item{c14DecayRate}{}

\item{solverfunc}{}

\item{pass}{}
}
\description{
initialize,Model_14-method
}
