% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\name{initialize,Model_by_PoolNames-method}
\alias{initialize,Model_by_PoolNames-method}
\title{automatic title}
\usage{
\S4method{initialize}{Model_by_PoolNames}(
  .Object,
  times,
  mat,
  initialValues,
  inputFluxes,
  timeSymbol,
  pass = FALSE,
  solverfunc = deSolve.lsoda.wrapper
)
}
\arguments{
\item{.Object}{object of class:\code{Model_by_PoolNames}, no manual documentation}

\item{times}{no manual documentation}

\item{mat}{no manual documentation}

\item{initialValues}{no manual documentation}

\item{inputFluxes}{no manual documentation}

\item{timeSymbol}{no manual documentation}

\item{pass}{no manual documentation}

\item{solverfunc}{no manual documentation}
}
\description{
automatic title
}
