% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{initialize,NlModel-method}
\alias{initialize,NlModel-method}
\title{automatic title}
\usage{
\S4method{initialize}{NlModel}(
  .Object,
  times = c(0, 1),
  DepComp = new(Class = "TransportDecompositionOperator", 0, 1, function(t) {    
    return(matrix(nrow = 1, ncol = 1, 0)) }, function(t) {     return(matrix(nrow = 1,
    ncol = 1, 0)) }),
  initialValues = numeric(),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{object of class:\code{NlModel}, : : no manual documentation}

\item{times}{: : no manual documentation}

\item{DepComp}{: : no manual documentation}

\item{initialValues}{: : no manual documentation}

\item{inputFluxes}{: : no manual documentation}

\item{solverfunc}{: : no manual documentation}

\item{pass}{: : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
