% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{initialize,NlModel-method}
\alias{initialize,NlModel-method}
\title{automatic title}
\usage{
\S4method{initialize}{NlModel}(
  .Object,
  times = c(0, 1),
  DepComp = new(Class = "TransportDecompositionOperator", 0, 1, function(t) {
    
    return(matrix(nrow = 1, ncol = 1, 0))
 }, function(t) {
     return(matrix(nrow = 1,
    ncol = 1, 0))
 }),
  initialValues = numeric(),
  inputFluxes = BoundInFluxes(function(t) {
     return(matrix(nrow = 1, ncol = 1, 1))

    }, 0, 1),
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)
}
\arguments{
\item{.Object}{no manual documentation}

\item{times}{no manual documentation}

\item{DepComp}{no manual documentation}

\item{initialValues}{no manual documentation}

\item{inputFluxes}{no manual documentation}

\item{solverfunc}{no manual documentation}

\item{pass}{no manual documentation}
}
\description{
automatic title
}
