% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{initialize,TimeMap-method}
\alias{initialize,TimeMap-method}
\title{initialize,TimeMap-method}
\usage{
\S4method{initialize}{TimeMap}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  map = function(t) {     t }
)
}
\arguments{
\item{.Object}{object of class:\code{TimeMap}. }

\item{starttime}{found automatically by inspection, not documented yet}

\item{endtime}{found automatically by inspection, not documented yet}

\item{map}{found automatically by inspection, not documented yet}
}
\description{
initialize,TimeMap-method
}
