% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{initialize,TransportDecompositionOperator-method}
\alias{initialize,TransportDecompositionOperator-method}
\title{initialize,TransportDecompositionOperator-method}
\usage{
\S4method{initialize}{TransportDecompositionOperator}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  numberOfPools = 1,
  alpha = list(),
  f = function(t, O) {     t },
  lag = 0
)
}
\arguments{
\item{.Object}{object of class:\code{TransportDecompositionOperator}. }

\item{starttime}{}

\item{endtime}{}

\item{numberOfPools}{}

\item{alpha}{}

\item{f}{}

\item{lag}{}
}
\description{
initialize,TransportDecompositionOperator-method
}
