% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransportDecompositionOperator.R
\name{initialize,TransportDecompositionOperator-method}
\alias{initialize,TransportDecompositionOperator-method}
\title{automatic title}
\usage{
\S4method{initialize}{TransportDecompositionOperator}(
  .Object,
  starttime = numeric(),
  endtime = numeric(),
  numberOfPools = 1,
  alpha = list(),
  f = function(t, O) {     t },
  lag = 0
)
}
\arguments{
\item{.Object}{object of class:\code{TransportDecompositionOperator}, no manual documentation}

\item{starttime}{no manual documentation}

\item{endtime}{no manual documentation}

\item{numberOfPools}{no manual documentation}

\item{alpha}{no manual documentation}

\item{f}{no manual documentation}

\item{lag}{no manual documentation}
}
\description{
automatic title
}
