% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{initialize,UnBoundInFluxes-method}
\alias{initialize,UnBoundInFluxes-method}
\title{Automatic description: initialize,UnBoundInFluxes-method}
\usage{
\S4method{initialize}{UnBoundInFluxes}(.Object, map = function() {
})
}
\arguments{
\item{.Object}{object of class:\code{UnBoundInFluxes}. : object of class:\code{UnBoundInFluxes}, no manual
documentation}

\item{map}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: initialize,UnBoundInFluxes-method
}
