% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{initialize,UnBoundInFluxes-method}
\alias{initialize,UnBoundInFluxes-method}
\title{initialize,UnBoundInFluxes-method}
\usage{
\S4method{initialize}{UnBoundInFluxes}(.Object, map = function() {
})
}
\arguments{
\item{.Object}{object of class:\code{UnBoundInFluxes}. }

\item{map}{}
}
\description{
initialize,UnBoundInFluxes-method
}
