% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{initialize,UnBoundInFluxes-method}
\alias{initialize,UnBoundInFluxes-method}
\title{automatic title}
\usage{
\S4method{initialize}{UnBoundInFluxes}(.Object, map = function() {
})
}
\arguments{
\item{.Object}{object of class:\code{UnBoundInFluxes}, no manual documentation}

\item{map}{no manual documentation}
}
\description{
automatic title
}
