% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{initialize,UnBoundLinDecompOp-method}
\alias{initialize,UnBoundLinDecompOp-method}
\title{automatic title}
\usage{
\S4method{initialize}{UnBoundLinDecompOp}(.Object, matFunc = function() {
})
}
\arguments{
\item{.Object}{object of class:\code{UnBoundLinDecompOp}, : : no manual documentation}

\item{matFunc}{: : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
