% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFlux_by_PoolIndex.R
\name{InternalFlux_by_PoolIndex,function,numeric,numeric,missing-method}
\alias{InternalFlux_by_PoolIndex,function,numeric,numeric,missing-method}
\title{constructor from an ordered pair of PoolIndex (integer like) objects and a function with vector argument}
\usage{
\S4method{InternalFlux_by_PoolIndex}{function,numeric,numeric,missing}(func, sourceIndex, destinationIndex)
}
\arguments{
\item{func}{function}

\item{sourceIndex}{index of source pool}

\item{destinationIndex}{index of destination pool}
}
\description{
constructor from an ordered pair of PoolIndex (integer like) objects and a function with vector argument
}
