% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{InternalFlux_by_PoolIndex}
\alias{InternalFlux_by_PoolIndex}
\title{Generic constructor for the class with the same name}
\usage{
InternalFlux_by_PoolIndex(func, sourceIndex, destinationIndex, src_to_dest)
}
\arguments{
\item{func}{function with flux rate}

\item{sourceIndex}{index of the source pool}

\item{destinationIndex}{index of the destination pool}

\item{src_to_dest}{source to destination}
}
\description{
Generic constructor for the class with the same name
}
