% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFlux_by_PoolName.R
\name{InternalFlux_by_PoolName,function,missing,missing,character-method}
\alias{InternalFlux_by_PoolName,function,missing,missing,character-method}
\title{automatic title}
\usage{
\S4method{InternalFlux_by_PoolName}{`function`,missing,missing,character}(func, src_to_dest)
}
\arguments{
\item{func}{object of class:\code{function}, : : no manual documentation}

\item{src_to_dest}{object of class:\code{character}, : : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
