% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFlux_by_PoolName.R
\name{InternalFlux_by_PoolName,function,missing,missing,character-method}
\alias{InternalFlux_by_PoolName,function,missing,missing,character-method}
\title{Automatic description:
InternalFlux_by_PoolName,function,missing,missing,character-method}
\usage{
\S4method{InternalFlux_by_PoolName}{`function`,missing,missing,character}(func, src_to_dest)
}
\arguments{
\item{func}{object of class:\code{function}. : object of class:\code{function}, no manual documentation}

\item{src_to_dest}{object of class:\code{character}. : object of class:\code{character}, no manual
documentation}
}
\description{
Automatic description:
InternalFlux_by_PoolName,function,missing,missing,character-method
}
