\name{linMaker}
\alias{linMaker}
\title{linMaker}

\usage{linMaker(nonLinearPoolOutput, globalSolution, poolSolution)}
\arguments{
  \item{nonLinearPoolOutput}{
}
  \item{globalSolution}{
}
  \item{poolSolution}{
}
}



\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}





