\name{listProduct}
\alias{listProduct}
\usage{listProduct(...)}
\examples{listproduct(list('a','b'),list(1,2))}
\arguments{
\item{...}{lists}
}
\description{Creates a list of all combinations of the elements of the inputlists
(like a "tensor product list " The list elements can be of any class.
The function is used in examples and tests to produce all possible combinations 
of arguments to a function.
look at the tests for example usage}
\title{tensor product of lists}
\value{a list of lists each containing one combinations of the elements of the input lists}
