% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{S4 class representing a model run}
\description{
S4 class representing a model run
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{Model} in their signature:}{
      \itemize{
        \item \code{\link{[,Model,character,missing,missing-method}}
        \item \code{\link{getAccumulatedRelease,Model-method}}
        \item \code{\link{getC,Model-method}}
        \item \code{\link{getDecompOp,Model-method}}
        \item \code{\link{getInFluxes,Model-method}}
        \item \code{\link{getReleaseFlux,Model-method}}
        \item \code{\link{getRightHandSideOfODE,Model-method}}
        \item \code{\link{getTimes,Model-method}}
        \item \code{\link{initialize,Model-method}}
        \item \code{\link{plot,Model-method}}
      }
  }
}
