\title{Model_14}
\description{defines a representation of a 14C model
defines a representation of a 14C model}
\name{Model_14-class}
\alias{Model_14-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getC14}{\code{signature(object = "Model_14")}: ... } \code{\link{getC14,Model_14-method}}
    \item{getF14C}{\code{signature(object = "Model_14")}: ... } \code{\link{getF14C,Model_14-method}}
    \item{getF14R}{\code{signature(object = "Model_14")}: ... } \code{\link{getF14R,Model_14-method}}
    \item{getF14}{\code{signature(object = "Model_14")}: ... } \code{\link{getF14,Model_14-method}}
    \item{getReleaseFlux14}{\code{signature(object = "Model_14")}: ... } \code{\link{getReleaseFlux14,Model_14-method}}
	 }}
\section{Constructors}{	\code{\link{Model_14}}\cr}
