\description{This class  extends \code{\linkS4class{Model}, the focal point of SoilR. 
by functionality required to represent \eqn{$^{14}C}[14C} contents.
\itemize{
\item It combines all the components that are needed to solve the
initial value problem for the pool contents.
\eqn{\vec{C}}{C=(C_1,...C_n)^t, }. 
\item It provides the single argument for the different functions 
that are available to compute various results from the solution of the initial value problem. 
See subsection \code{Methods} and the examples.)
}}
\name{Model_14-class}
\alias{Model_14-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getC14}{\code{signature(object = "Model_14")}: ... } \code{\link{getC14,Model_14-method}}
    \item{getF14C}{\code{signature(object = "Model_14")}: ... } \code{\link{getF14C,Model_14-method}}
    \item{getF14R}{\code{signature(object = "Model_14")}: ... } \code{\link{getF14R,Model_14-method}}
    \item{getF14}{\code{signature(object = "Model_14")}: ... } \code{\link{getF14,Model_14-method}}
    \item{getReleaseFlux14}{\code{signature(object = "Model_14")}: ... } \code{\link{getReleaseFlux14,Model_14-method}}
	 }}
\title{Model_14}
\section{Exported superclasses}{\code{\link{Model-class}}\cr}
\section{Constructors}{	\code{\link{Model_14}}\cr}
