\name{Model_14}
\alias{Model_14}
\usage{Model_14(t,
A,
ivList,
initialValF,
inputFluxes,
inputFc,
c14DecayRate,
solverfunc,
pass)}
\arguments{
\item{t}{see the method arguments for details}
\item{A}{see the method arguments for details}
\item{ivList}{see the method arguments for details}
\item{initialValF}{see the method arguments for details}
\item{inputFluxes}{see the method arguments for details}
\item{inputFc}{see the method arguments for details}
\item{c14DecayRate}{see the method arguments for details}
\item{solverfunc}{see the method arguments for details}
\item{pass}{see the method arguments for details}
}
\section{Methods}{	\code{\link{Model_14,numeric,ANY,numeric,ConstFc,ANY,ANY,numeric,function,logical-method}}\cr
	\code{\link{Model_14,numeric,ANY,numeric,ConstFc,ANY,ANY,numeric,missing,logical-method}}\cr}
\title{A general constructor }
\description{Creates a Model_14 object from different sources
Have a look at the methods for details.}
