% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{Model_by_PoolNames-class}
\alias{Model_by_PoolNames-class}
\alias{Model_by_PoolNames}
\title{automatic title}
\description{
automatic title
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{Model_by_PoolNames} in their signature:}{
      \itemize{
        \item \code{\link{getC,Model_by_PoolNames-method}}
        \item \code{\link{getReleaseFlux,Model_by_PoolNames-method}}
        \item \code{\link{getRightHandSideOfODE,Model_by_PoolNames-method}}
        \item \code{\link{getSolution,Model_by_PoolNames-method}}
        \item \code{\link{initialize,Model_by_PoolNames-method}}
        \item \code{\link{plotPoolGraph,Model_by_PoolNames-method}}
      }
  }
}
