% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{Model_by_PoolNames}
\alias{Model_by_PoolNames}
\title{Constructor for \code{\linkS4class{Model_by_PoolNames}}}
\usage{
Model_by_PoolNames(
  smod,
  times,
  mat,
  initialValues,
  inputFluxes,
  internal_fluxes,
  out_fluxes,
  timeSymbol,
  solverfunc
)
}
\value{
A possibly nonlinear Model(run) that contains information about the
pool names and connectivity of the pools and is therefore the preferred
representation for new code.
}
\description{
Constructor for \code{\linkS4class{Model_by_PoolNames}}
}
