\title{NlModel}
\description{serves as a fence to the interface of SoilR functions. So that later implementations can differ	 }
\name{NlModel-class}
\alias{NlModel-class}
\docType{class}
\section{Methods}{  \describe{
    \item{$}{\code{signature(x = "NlModel")}: ... } \code{\link{$,NlModel-method}}
    \item{[}{\code{signature(x = "NlModel", i = "character", j = "ANY", drop = "ANY")}: ... } \code{\link{[,NlModel,character,ANY,ANY-method}}
    \item{getC}{\code{signature(object = "NlModel")}: ... } \code{\link{getC,NlModel-method}}
    \item{getReleaseFlux}{\code{signature(object = "NlModel")}: ... } \code{\link{getReleaseFlux,NlModel-method}}
    \item{getTimes}{\code{signature(object = "NlModel")}: ... } \code{\link{getTimes,NlModel-method}}
	 }}
