% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{plot,TimeMap-method}
\alias{plot,TimeMap-method}
\title{plot,TimeMap-method}
\usage{
\S4method{plot}{TimeMap}(x, y, ...)
}
\arguments{
\item{x}{object of class:\code{TimeMap}. }

\item{y}{}

\item{...}{}
}
\description{
plot,TimeMap-method
}
