% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{plot,TimeMap-method}
\alias{plot,TimeMap-method}
\title{automatic title}
\usage{
\S4method{plot}{TimeMap}(x, y, ...)
}
\arguments{
\item{x}{object of class:\code{TimeMap}, no manual documentation}

\item{y}{no manual documentation}

\item{...}{no manual documentation}
}
\description{
automatic title
}
