% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{plotPoolGraph,Model-method}
\alias{plotPoolGraph,Model-method}
\title{plotPoolGraph,Model-method}
\usage{
\S4method{plotPoolGraph}{Model}(x)
}
\arguments{
\item{x}{object of class:\code{Model}. }
}
\description{
plotPoolGraph,Model-method
}
