% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolConnection.R
\name{PoolConnection,ANY,ANY,missing-method}
\alias{PoolConnection,ANY,ANY,missing-method}
\title{constructor from an ordered pair of PoolId objects}
\usage{
\S4method{PoolConnection}{ANY,ANY,missing}(source, destination)
}
\description{
constructor from an ordered pair of PoolId objects
}
