% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{PoolConnection}
\alias{PoolConnection}
\title{PoolConnection}
\usage{
PoolConnection(source, destination, src_to_dest)
}
\arguments{
\item{source}{see method arguments}

\item{destination}{see method arguments}

\item{src_to_dest}{see method arguments}
}
\description{
PoolConnection
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{}}
}
}

