% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{PoolIndex-class}
\alias{PoolIndex-class}
\title{S4 class for pool indices}
\description{
used to dispatch pool index specific methods like conversion to names.
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{PoolIndex} in their signature:}{
      \itemize{
        \item \code{\link{PoolIndex,PoolIndex-method}}
        \item \code{\link{PoolName,PoolIndex-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{PoolIndex} in their signature:}{
    \subsection{superclass integer}{
      \itemize{
        \item \code{\link{check_pool_ids,PoolConnection_by_PoolIndex,integer-method}}
      }
    }
  }
}
\section{S4-superclasses (in the package)}{
      \itemize{
        \item \code{\linkS4class{PoolId}}
        \item \code{\linkS4class{integer}}
      }
}
