% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolIndex.R
\name{PoolIndex,PoolIndex-method}
\alias{PoolIndex,PoolIndex-method}
\title{pass through constructor fron an object of the same class}
\usage{
\S4method{PoolIndex}{PoolIndex}(id, poolNames)
}
\arguments{
\item{id}{object of class:\code{PoolIndex}, no manual documentation}

\item{poolNames}{no manual documentation}
}
\description{
This is here to be able to call PoolIndex on a PoolIndex object without
having to check before if it is necessary.
the unnecessary poolNames argument will be ignored.
}
