% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolName.R
\docType{methods}
\name{PoolIndex,PoolName-method}
\alias{PoolIndex,PoolName-method}
\title{convert to number like object}
\usage{
\S4method{PoolIndex}{PoolName}(id, poolNames)
}
\description{
convert to number like object
}
