% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{PoolIndex}
\alias{PoolIndex}
\title{automatic title}
\usage{
PoolIndex(id, ...)
}
\arguments{
\item{id}{see method arguments}

\item{...}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{PoolIndex,character-method}}
        \item \code{\link{PoolIndex,numeric-method}}
        \item \code{\link{PoolIndex,PoolIndex-method}}
        \item \code{\link{PoolIndex,PoolName-method}}
      }
}
