% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{PoolName-class}
\alias{PoolName-class}
\title{class for pool-name-strings}
\description{
used to control the creation of PoolName objects which have to be valid R identifiers 
and to dispatch pool name specific methods like conversion to pool indices
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{PoolName} in their signature:}{
      \itemize{
        \item \code{\link{PoolIndex,PoolName-method}}
        \item \code{\link{PoolName,PoolName-method}}
      }
  }
}
