% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolName.R
\docType{methods}
\name{PoolName,PoolName-method}
\alias{PoolName,PoolName-method}
\title{pass throug constructor fron an object of the same class}
\usage{
\S4method{PoolName}{PoolName}(id, poolNames)
}
\description{
This is here to be able to call PoolName on a PoolName ojbect without
having to test before if we have to.
}
