% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{PoolName}
\alias{PoolName}
\title{automatic title}
\usage{
PoolName(id, ...)
}
\arguments{
\item{id}{see method arguments}

\item{...}{see method arguments}
}
\description{
automatic title
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{PoolName,character-method}}
        \item \code{\link{PoolName,PoolIndex-method}}
        \item \code{\link{PoolName,PoolName-method}}
      }
}
