% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{print,NlModel-method}
\alias{print,NlModel-method}
\title{print,NlModel-method}
\usage{
\S4method{print}{NlModel}(x)
}
\arguments{
\item{x}{object of class:\code{NlModel}. }
}
\description{
print,NlModel-method
}
