% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{ScalarTimeMap-class}
\alias{ScalarTimeMap-class}
\title{S4 class for a scalar time dependent function on a finite time interval}
\description{
S4 class for a scalar time dependent function on a finite time interval
}
