% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarTimeMap.R
\name{ScalarTimeMap,missing,missing,missing,missing,numeric-method}
\alias{ScalarTimeMap,missing,missing,missing,missing,numeric-method}
\title{special case for a time map from a constant}
\usage{
\S4method{ScalarTimeMap}{missing,missing,missing,missing,numeric}(starttime = -Inf, endtime = +Inf, data, lag = 0)
}
\arguments{
\item{starttime}{object of class:\code{missing}, no manual documentation}

\item{endtime}{object of class:\code{missing}, no manual documentation}

\item{data}{object of class:\code{numeric}, no manual documentation}

\item{lag}{no manual documentation}
}
\description{
special case for a time map from a constant
}
