% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ScalarTimeMap}
\alias{ScalarTimeMap}
\title{automatic title}
\usage{
ScalarTimeMap(
  map,
  starttime,
  endtime,
  times,
  data,
  lag = 0,
  interpolation = splinefun,
  ...
)
}
\arguments{
\item{map}{see method arguments}

\item{starttime}{see method arguments}

\item{endtime}{see method arguments}

\item{times}{see method arguments}

\item{data}{see method arguments}

\item{lag}{see method arguments}

\item{interpolation}{see method arguments}

\item{...}{see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{ScalarTimeMap,data.frame,missing,missing,missing,missing-method}}
  \item \code{\link{ScalarTimeMap,function,missing,missing,missing,missing-method}}
  \item \code{\link{ScalarTimeMap,function,numeric,numeric,missing,missing-method}}
  \item \code{\link{ScalarTimeMap,missing,missing,missing,missing,numeric-method}}
  \item \code{\link{ScalarTimeMap,missing,missing,missing,numeric,numeric-method}}
}
}

