% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalarTimeMap.R
\name{ScalarTimeMap}
\alias{ScalarTimeMap}
\title{constructor}
\usage{
ScalarTimeMap(...)
}
\arguments{
\item{...}{The arguments are passed through to create a function of time (an object of class \code{\link{TimeMap}}) which is subsequently checked to yield a scalar numeric value}
}
\description{
constructor
}
