% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{ScalarTimeMap}
\alias{ScalarTimeMap}
\title{Constructor for \code{\link{ScalarTimeMap-class}}}
\usage{
ScalarTimeMap(
  map,
  starttime,
  endtime,
  times,
  data,
  lag = 0,
  interpolation = splinefun,
  ...
)
}
\arguments{
\item{map}{see method arguments}

\item{starttime}{see method arguments}

\item{endtime}{see method arguments}

\item{times}{see method arguments}

\item{data}{see method arguments}

\item{lag}{see method arguments}

\item{interpolation}{see method arguments}

\item{...}{see method arguments}
}
\description{
Constructor for \code{\link{ScalarTimeMap-class}}
}
