% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{show,NlModel-method}
\alias{show,NlModel-method}
\title{show,NlModel-method}
\usage{
\S4method{show}{NlModel}(object)
}
\arguments{
\item{object}{object of class:\code{NlModel}. }
}
\description{
show,NlModel-method
}
