\name{SoilR-package}
\alias{SoilR-package}
\alias{SoilR}
\docType{package}
\title{Models of Soil Organic Matter Decomposition}
\description{This package contains functions for modeling Soil Organic Matter decomposition in terrestrial ecosystems.}
\details{
\tabular{ll}{Package: \tab SoilR\cr
Title: \tab Models of Soil Organic Matter Decomposition\cr
Version: \tab 1.0-1\cr
Date: \tab 2012-03-28\cr
Author: \tab Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>\cr
Maintainer: \tab Markus Mueller <mamueller@bgc-jena.mpg.de>\cr
License: \tab GPL-3\cr
Depends: \tab deSolve,methods,hash,FME,MASS,lattice,ggplot2\cr
encodign: \tab utf8\cr}
}
\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>}

\keyword{ package }


