\name{SoilR-package}
\alias{SoilR-package}
\docType{package}
\title{Models of Soil Organic Matter Decomposition}
\description{This package contains functions for modeling Soil Organic Matter decomposition in terrestrial ecosystems.}
\details{
\tabular{ll}{Package: \tab SoilR\cr
Type: \tab Package\cr
Title: \tab Models of Soil Organic Matter Decomposition\cr
Version: \tab 0.1-2\cr
Date: \tab 2010-09-30\cr
Author: \tab Carlos A. Sierra <csierra@bgc-jena.mpg.de>\cr
Maintainer: \tab Carlos A. Sierra <csierra@bgc-jena.mpg.de>\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Packaged: \tab 2010-09-30 09:34:02 UTC; _csierra\cr
Depends: \tab deSolve\cr}
}
\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>}

\keyword{ package }


