% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilR.F0__deprecated_in_1.2.R
\name{SoilR.F0.new}
\alias{SoilR.F0.new}
\title{deprecated function that used to create an object of class SoilR.F0}
\usage{
SoilR.F0.new(values = c(0), format = "Delta14C")
}
\arguments{
\item{values}{a numeric vector}

\item{format}{a character string describing the format e.g. "Delta14C"}
}
\value{
An object of class \code{\link{ConstFc-class}} that contains data
and a format description that can later be used to convert the data into
other formats if the conversion is implemented.
}
\description{
The function internally calls the constructor of the replacement class
\code{\link{ConstFc-class}}.
}
