% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{StateIndependentInFluxList_by_PoolIndex-class}
\alias{StateIndependentInFluxList_by_PoolIndex-class}
\title{Subclass of list that is guaranteed to contain only elements of type
\linkS4class{StateIndependentInFlux_by_PoolIndex}}
\description{
Subclass of list that is guaranteed to contain only elements of type
\linkS4class{StateIndependentInFlux_by_PoolIndex}
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{StateIndependentInFluxList_by_PoolIndex} in their signature:}{
      \itemize{
        \item \code{\link{InFluxes,StateIndependentInFluxList_by_PoolIndex-method}}
      }
  }
  \subsection{S4-methods with superclasses (in the package) of class \code{StateIndependentInFluxList_by_PoolIndex} in their signature:}{
    \subsection{superclass list}{
      \itemize{
        \item \code{\link{ConstantInFluxList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantInternalFluxRateList_by_PoolName,list-method}}
        \item \code{\link{ConstantOutFluxRateList_by_PoolIndex,list-method}}
        \item \code{\link{ConstantOutFluxRateList_by_PoolName,list-method}}
        \item \code{\link{GeneralDecompOp,list-method}}
        \item \code{\link{InFluxes,list-method}}
        \item \code{\link{InFluxList_by_PoolIndex,list-method}}
        \item \code{\link{InFluxList_by_PoolName,list-method}}
        \item \code{\link{InternalFluxList_by_PoolIndex,list-method}}
        \item \code{\link{InternalFluxList_by_PoolName,list-method}}
        \item \code{\link{OutFluxList_by_PoolIndex,list-method}}
        \item \code{\link{OutFluxList_by_PoolName,list-method}}
        \item \code{\link{StateIndependentInFluxList_by_PoolIndex,list-method}}
        \item \code{\link{TimeMap,list,missing,missing,missing,missing-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,list-method}}
      }
    }
  }
}
