% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{[,Model,character,missing,missing-method}
\alias{[,Model,character,missing,missing-method}
\title{[,Model,character,missing,missing-method}
\usage{
\S4method{[}{Model,character,missing,missing}(x, i)
}
\arguments{
\item{x}{object of class:\code{Model}. }

\item{i}{object of class:\code{character}. }
}
\description{
[,Model,character,missing,missing-method
}
