% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{[,Model,character,missing,missing-method}
\alias{[,Model,character,missing,missing-method}
\title{Experimentally overloaded index operator}
\usage{
\S4method{[}{Model,character,missing,missing}(x, i)
}
\arguments{
\item{x}{object of class:\code{Model}, no manual documentation}

\item{i}{object of class:\code{character}, no manual documentation}
}
\description{
The method provides shortcuts and a unified interface to some of the
methods that can be applied to a model.
For a given model `M` the code `M['C'] is equivalent to `getC(M)` and
`M['ReleaseFlux']` is equivalent to `getReleaseFlux(M)`
`M['AccumulatedRelease']` is equivalent to `getAccumulatedRelease(M)`
}
