% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{[,NlModel,character,ANY,ANY-method}
\alias{[,NlModel,character,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{[}{NlModel,character,ANY,ANY}(x, i)
}
\arguments{
\item{x}{no manual documentation}

\item{i}{no manual documentation}
}
\description{
automatic title
}
