% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlModel.R
\name{[,NlModel,character,ANY,ANY-method}
\alias{[,NlModel,character,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{[}{NlModel,character,ANY,ANY}(x, i)
}
\arguments{
\item{x}{object of class:\code{NlModel}, : : no manual documentation}

\item{i}{object of class:\code{character}, : : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
