% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{[[,MCSim-method}
\alias{[[,MCSim-method}
\title{automatic title}
\usage{
\S4method{[[}{MCSim}(x, i)
}
\arguments{
\item{x}{object of class:\code{MCSim}, no manual documentation}

\item{i}{no manual documentation}
}
\description{
automatic title
}
