% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{[[<-,MCSim-method}
\alias{[[<-,MCSim-method}
\title{[[<-,MCSim-method}
\usage{
\S4method{[[}{MCSim}(x, i, j, ...) <- value
}
\arguments{
\item{x}{object of class:\code{MCSim}. }

\item{i}{}

\item{j}{}

\item{...}{}

\item{value}{}
}
\description{
[[<-,MCSim-method
}
