% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSim.R
\name{[[<-,MCSim-method}
\alias{[[<-,MCSim-method}
\title{[[<-,MCSim-method}
\usage{
\S4method{[[}{MCSim}(x, i, j, ...) <- value
}
\arguments{
\item{x}{object of class:\code{MCSim}. }

\item{i}{found automatically by inspection, not documented yet}

\item{j}{found automatically by inspection, not documented yet}

\item{...}{found automatically by inspection, not documented yet}

\item{value}{found automatically by inspection, not documented yet}
}
\description{
[[<-,MCSim-method
}
