% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{summary,Model-method}
\alias{summary,Model-method}
\title{Automatic description: summary,Model-method}
\usage{
\S4method{summary}{Model}(object)
}
\arguments{
\item{object}{object of class:\code{Model}. : object of class:\code{Model}, no manual documentation}
}
\description{
Automatic description: summary,Model-method
}
