% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{SymbolicModel_by_PoolNames-class}
\alias{SymbolicModel_by_PoolNames-class}
\alias{SymbolicModel_by_PoolNames}
\title{A symbolic model description based on flux functions}
\description{
The set of flux functions along with the timesymbol
is complete description of the structure
}
\section{S4-methods}{
  \subsection{S4-methods with class \code{SymbolicModel_by_PoolNames} in their signature:}{
      \itemize{
        \item \code{\link{Model_by_PoolNames,SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing,missing-method}}
        \item \code{\link{plotPoolGraph,SymbolicModel_by_PoolNames-method}}
      }
  }
}
