\name{TimeMap-class}
\Rdversion{1.1}
\docType{class}
\alias{TimeMap-class}
\alias{TimeMap}
\alias{initialize,TimeMap-method}

\title{TimeMap}
\description{NA}
\section{Objects from the Class}{Objects can be created by calls of the form \code{new(TimeMap ...)}}
\section{Slots}{
  \describe{
    \item{\code{starttime}:}{Object of class \code{"numeric"}}
    \item{\code{endtime}:}{Object of class \code{"numeric"}}
    \item{\code{map}:}{Object of class \code{"function"}}
  }
}
\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "TimeMap")}: ... }
	 }
}

\author{Carlos A. Sierra <csierra@bgc-jena.mpg.de>}





