% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{TimeMap,data.frame,missing,missing,missing,missing-method}
\alias{TimeMap,data.frame,missing,missing,missing,missing-method}
\title{TimeMap,data.frame,missing,missing,missing,missing-method constructor}
\arguments{
\item{map}{: of class data.frame, a data frame containing two columns}

\item{lag}{: a time delay}

\item{interpolation}{: the interpolating function}
}
\description{
create a TimeMap object by interpolating the data.frame
}
