% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,data.frame,missing,missing,missing,missing-method}
\alias{TimeMap,data.frame,missing,missing,missing,missing-method}
\title{automatic title}
\usage{
\S4method{TimeMap}{data.frame,missing,missing,missing,missing}(map, lag = 0, interpolation = splinefun)
}
\arguments{
\item{map}{no manual documentation}

\item{lag}{no manual documentation}

\item{interpolation}{no manual documentation}
}
\description{
automatic title
}
