% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{TimeMap,function,numeric,numeric,missing,missing-method}
\alias{TimeMap,function,numeric,numeric,missing,missing-method}
\title{TimeMap,function,numeric,numeric,missing,missing-method constructor}
\arguments{
\item{map}{: of class function}

\item{starttime}{: of class numeric}

\item{endtime}{: of class numeric}
}
\description{
create a TimeMap object from the function definition and the time interval
}
