% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,function,numeric,numeric,missing,missing-method}
\alias{TimeMap,function,numeric,numeric,missing,missing-method}
\title{manual constructor for a function and an interval}
\usage{
\S4method{TimeMap}{`function`,numeric,numeric,missing,missing}(map, starttime, endtime, lag = 0)
}
\arguments{
\item{map}{object of class:\code{function}. }

\item{starttime}{object of class:\code{numeric}. }

\item{endtime}{object of class:\code{numeric}. }

\item{lag}{found automatically by inspection, not documented yet}
}
\description{
manual constructor for a function and an interval
}
