% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,list,missing,missing,missing,missing-method}
\alias{TimeMap,list,missing,missing,missing,missing-method}
\title{automatic title}
\usage{
\S4method{TimeMap}{list,missing,missing,missing,missing}(map, lag = 0, interpolation = splinefun)
}
\arguments{
\item{map}{object of class:\code{list}, A nested list of the form list(times=l1,data=l2)
where l1 is a vector or list of the time values
and l2 is a list of numbers, vectors, matrices or arrays.}

\item{lag}{Time delay for the created function of time}

\item{interpolation}{The function used to compute the interpolation e.g splinefun

Interprets the received list as value table of a time dependent function}
}
\description{
automatic title
}
