% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,missing,missing,missing,numeric,array-method}
\alias{TimeMap,missing,missing,missing,numeric,array-method}
\title{Automatic description: TimeMap,missing,missing,missing,numeric,array-method}
\usage{
\S4method{TimeMap}{missing,missing,missing,numeric,array}(times, data, lag = 0, interpolation = splinefun)
}
\arguments{
\item{times}{object of class:\code{numeric}. : object of class:\code{numeric}, no manual documentation}

\item{data}{object of class:\code{array}. : object of class:\code{array}, no manual documentation}

\item{lag}{found automatically by inspection, not documented yet}

\item{interpolation}{found automatically by inspection, not documented yet}
}
\description{
Automatic description: TimeMap,missing,missing,missing,numeric,array-method
}
