% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,missing,missing,missing,numeric,matrix-method}
\alias{TimeMap,missing,missing,missing,numeric,matrix-method}
\title{automatic title}
\usage{
\S4method{TimeMap}{missing,missing,missing,numeric,matrix}(times, data, lag = 0, interpolation = splinefun)
}
\arguments{
\item{times}{object of class:\code{numeric}, : : no manual documentation}

\item{data}{object of class:\code{matrix}, : : no manual documentation}

\item{lag}{: : no manual documentation}

\item{interpolation}{: : no manual documentation
inspection of the code.  You can use the "update_auto_comment_roclet" to
automatically adapt them to changes in the source code. This will remove
`@param` tags for parameters that are no longer present in the source code
and add `@param` tags with a default description for yet undocumented
parameters.  If you remove this `@autocomment` tag your comments will no
longer be touched by the "update_autocomment_roclet".}
}
\description{
automatic title
}
