% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,missing,missing,missing,numeric,numeric-method}
\alias{TimeMap,missing,missing,missing,numeric,numeric-method}
\title{TimeMap,missing,missing,missing,numeric,numeric-method}
\usage{
\S4method{TimeMap}{missing,missing,missing,numeric,numeric}(times, data, lag = 0, interpolation = splinefun)
}
\arguments{
\item{times}{object of class:\code{numeric}. }

\item{data}{object of class:\code{numeric}. }

\item{lag}{}

\item{interpolation}{}
}
\description{
TimeMap,missing,missing,missing,numeric,numeric-method
}
