% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeMap.R
\name{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}
\alias{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}
\title{automatic title}
\usage{
\S4method{TimeMap}{TimeMap,ANY,ANY,ANY,ANY}(map)
}
\arguments{
\item{map}{object of class:\code{TimeMap}, no manual documentation}
}
\description{
automatic title
}
