% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}
\alias{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}
\title{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method pass through constructor}
\arguments{
\item{map}{: of class TimeMap, the object that will be returned unchanged}
}
\description{
The function just returns its argument. So any function that has to convert
one of its argument can just call TimeMap on it even if the argument is
allready one.
}
