\name{TimeMap}
\alias{TimeMap}
\usage{TimeMap(map,
starttime,
endtime,
lag,
interpolation)}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\section{Methods}{	\code{\link{TimeMap,data.frame,missing,missing-method}}\cr
	\code{\link{TimeMap,function,numeric,numeric-method}}\cr}
\title{generic constructor}
\description{create a TimeMap object from different sources}
