\description{no Description}
\name{TimeMap}
\alias{TimeMap}
\usage{TimeMap(map,
starttime,
endtime,
data,
times,
lag,
interpolation)}
\title{TimeMap S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{data}{see the method arguments for details}
\item{times}{see the method arguments for details}
\item{lag}{see the method arguments for details}
\item{interpolation}{see the method arguments for details}
}
\section{Methods}{	\code{\link{TimeMap,data.frame,missing,missing,ANY,ANY-method}}\cr
	\code{\link{TimeMap,function,numeric,numeric,ANY,ANY-method}}\cr
	\code{\link{TimeMap,list,ANY,ANY,ANY,ANY-method}}\cr
	\code{\link{TimeMap,missing,missing,missing,array,numeric-method}}\cr}
