\description{no Description}
\name{TimeMap}
\alias{TimeMap}
\usage{TimeMap(map,
starttime,
endtime,
times,
data,
lag=0,
interpolation=splinefun)}
\title{TimeMap S4 generic}
\arguments{
\item{map}{see the method arguments for details}
\item{starttime}{see the method arguments for details}
\item{endtime}{see the method arguments for details}
\item{times}{see the method arguments for details}
\item{data}{see the method arguments for details}
\item{lag}{a time delay}
\item{interpolation}{the interpolating function}
}
\section{Methods}{	\code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}\cr
	\code{\link{TimeMap,data.frame,missing,missing,missing,missing-method}}\cr
	\code{\link{TimeMap,function,numeric,numeric,missing,missing-method}}\cr
	\code{\link{TimeMap,list,missing,missing,missing,missing-method}}\cr
	\code{\link{TimeMap,missing,missing,missing,numeric,array-method}}\cr
	\code{\link{TimeMap,missing,missing,missing,numeric,list-method}}\cr
	\code{\link{TimeMap,missing,missing,missing,numeric,matrix-method}}\cr
	\code{\link{TimeMap,missing,missing,missing,numeric,numeric-method}}\cr}
