% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{TimeMap}
\alias{TimeMap}
\title{Constructor for \code{\link{TimeMap-class}}}
\usage{
TimeMap(
  map,
  starttime,
  endtime,
  times,
  data,
  lag = 0,
  interpolation = splinefun,
  ...
)
}
\arguments{
\item{map}{see method arguments}

\item{starttime}{see method arguments}

\item{endtime}{see method arguments}

\item{times}{see method arguments}

\item{data}{see method arguments}

\item{lag}{see method arguments}

\item{interpolation}{see method arguments}

\item{...}{see method arguments}
}
\description{
Constructor for \code{\link{TimeMap-class}}
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{TimeMap,data.frame,missing,missing,missing,missing-method}}
        \item \code{\link{TimeMap,function,missing,missing,missing,missing-method}}
        \item \code{\link{TimeMap,function,numeric,numeric,missing,missing-method}}
        \item \code{\link{TimeMap,list,missing,missing,missing,missing-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,array-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,list-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,matrix-method}}
        \item \code{\link{TimeMap,missing,missing,missing,numeric,numeric-method}}
        \item \code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
      }
}
