\description{no Description}
\name{TransportDecompositionOperator-class}
\alias{TransportDecompositionOperator-class}
\docType{class}
\section{Methods}{Exported methods directly defined for class TransportDecompositionOperator:

  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "TransportDecompositionOperator")}: ... } \code{\link{getFunctionDefinition,TransportDecompositionOperator-method}}
	 }

Methods inherited from superclasses:

from class TimeMap:

  \describe{
    \item{GeneralDecompOp}{\code{signature(object = "TimeMap")}: ... } \code{\link{GeneralDecompOp,TimeMap-method}}
    \item{GeneralInFlux}{\code{signature(object = "TimeMap")}: ... } \code{\link{GeneralInFlux,TimeMap-method}}
    \item{TimeMap}{\code{signature(map = "TimeMap", starttime = "ANY", endtime = "ANY", times = "ANY", data = "ANY")}: ... } \code{\link{TimeMap,TimeMap,ANY,ANY,ANY,ANY-method}}
    \item{add_plot}{\code{signature(x = "TimeMap")}: ... } \code{\link{add_plot,TimeMap-method}}
    \item{as.character}{\code{signature(x = "TimeMap")}: ... } \code{\link{as.character,TimeMap-method}}
    \item{getFunctionDefinition}{\code{signature(object = "TimeMap")}: ... } \code{\link{getFunctionDefinition,TimeMap-method}}
    \item{getTimeRange}{\code{signature(object = "TimeMap")}: ... } \code{\link{getTimeRange,TimeMap-method}}
	 }}
\title{TransportDecompositionOperator S4 class}
\section{Superclasses}{\code{\link{TimeMap-class}}\cr}
