\description{no Description}
\name{UnBoundInFlux-class}
\alias{UnBoundInFlux-class}
\docType{class}
\section{Methods}{  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "UnBoundInFlux")}: ... } \code{\link{getFunctionDefinition,UnBoundInFlux-method}}
    \item{getTimeRange}{\code{signature(object = "UnBoundInFlux")}: ... } \code{\link{getTimeRange,UnBoundInFlux-method}}
	 }}
\title{constant decomposition operator }
\section{Superclasses}{\code{\link{InFlux-class}}\cr}
