\description{no Description}
\name{UnBoundInFlux-class}
\alias{UnBoundInFlux-class}
\docType{class}
\section{Methods}{Exported methods directly defined for class UnBoundInFlux:

  \describe{
    \item{getFunctionDefinition}{\code{signature(object = "UnBoundInFlux")}: ... } \code{\link{getFunctionDefinition,UnBoundInFlux-method}}
    \item{getTimeRange}{\code{signature(object = "UnBoundInFlux")}: ... } \code{\link{getTimeRange,UnBoundInFlux-method}}
	 }

Methods inherited from superclasses:

from class InFlux:

  \describe{
    \item{GeneralInFlux}{\code{signature(object = "InFlux")}: ... } \code{\link{GeneralInFlux,InFlux-method}}
	 }}
\title{constant decomposition operator }
\section{Superclasses}{\code{\link{InFlux-class}}\cr}
