% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{UnBoundInFluxes,function-method}
\alias{UnBoundInFluxes,function-method}
\title{automatic title}
\usage{
\S4method{UnBoundInFluxes}{`function`}(map)
}
\arguments{
\item{map}{no manual documentation}
}
\description{
automatic title
}
