% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundInflux.R
\name{UnBoundInFluxes,function-method}
\alias{UnBoundInFluxes,function-method}
\title{UnBoundInFluxes,function-method}
\usage{
\S4method{UnBoundInFluxes}{`function`}(map)
}
\arguments{
\item{map}{object of class:\code{function}. }
}
\description{
UnBoundInFluxes,function-method
}
