% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{UnBoundInFluxes}
\alias{UnBoundInFluxes}
\title{automatic title}
\usage{
UnBoundInFluxes(map)
}
\arguments{
\item{map}{: see method arguments}
}
\description{
automatic title
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{UnBoundInFluxes,function-method}}
}
}

