% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{UnBoundInFluxes}
\alias{UnBoundInFluxes}
\title{UnBoundInFluxes}
\usage{
UnBoundInFluxes(map)
}
\arguments{
\item{map}{see method arguments}
}
\description{
UnBoundInFluxes
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{UnBoundInFluxes,function-method}}
  \item \code{\link{UnBoundInFluxes,function-method}}
}
}

