\description{no Description}
\name{UnBoundLinDecompOp-class}
\alias{UnBoundLinDecompOp-class}
\docType{class}
\section{Methods}{  \describe{
    \item{BoundLinDecompOp}{\code{signature(map = "UnBoundLinDecompOp", starttime = "ANY", endtime = "ANY")}: ... } \code{\link{BoundLinDecompOp,UnBoundLinDecompOp,ANY,ANY-method}}
    \item{getFunctionDefinition}{\code{signature(object = "UnBoundLinDecompOp")}: ... } \code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}
    \item{getTimeRange}{\code{signature(object = "UnBoundLinDecompOp")}: ... } \code{\link{getTimeRange,UnBoundLinDecompOp-method}}
	 }}
\title{constant decomposition operator }
\section{Exported superclasses}{\code{\link{DecompOp-class}}\cr}
\section{Constructors}{	\code{\link{UnBoundLinDecompOp}}\cr
 Please also look at constructors of non virtual subclasses }
