\description{no Description}
\name{UnBoundLinDecompOp-class}
\alias{UnBoundLinDecompOp-class}
\docType{class}
\section{Methods}{  \describe{
    \item{BoundLinDecompOp}{\code{signature(map = "UnBoundLinDecompOp")}: ... } \code{\link{BoundLinDecompOp,UnBoundLinDecompOp-method}}
    \item{getFunctionDefinition}{\code{signature(object = "UnBoundLinDecompOp")}: ... } \code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}
    \item{getTimeRange}{\code{signature(object = "UnBoundLinDecompOp")}: ... } \code{\link{getTimeRange,UnBoundLinDecompOp-method}}
	 }}
\title{constant decomposition operator }
\section{Superclasses}{\code{\link{DecompOp-class}}\cr}
\section{Constructors}{	\code{\link{UnBoundLinDecompOp}}\cr}
