% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnBoundLinDecompOp.R
\name{UnBoundLinDecompOp,function-method}
\alias{UnBoundLinDecompOp,function-method}
\title{Generic constructor for the class with the same name}
\usage{
\S4method{UnBoundLinDecompOp}{function}(matFunc)
}
\arguments{
\item{matFunc}{no manual documentation}
}
\description{
Generic constructor for the class with the same name
}
\seealso{
Other UnBoundLinDecompOp_constructor: 
\code{\link{getFunctionDefinition,UnBoundLinDecompOp-method}}
}
\concept{UnBoundLinDecompOp_constructor}
