% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{methods}
\name{UnBoundLinDecompOp,function-method}
\alias{UnBoundLinDecompOp,function-method}
\title{UnBoundLinDecompOp,function-method construct from matrix valued function}
\arguments{
\item{matFunc}{: of class function}
}
\description{
This method creates a UnBoundLinDecompOp from a matrix The operator is
assumed to act on the vector of carbon stocks by multiplication of the (time
dependent) matrix from the left.
}
