% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{UnBoundNonLinDecompOp}
\alias{UnBoundNonLinDecompOp}
\title{Generic constructor for the class with the same name}
\usage{
UnBoundNonLinDecompOp(
  matFunc,
  internal_fluxes,
  out_fluxes,
  numberOfPools,
  state_variable_names,
  timeSymbol,
  operator
)
}
\description{
Generic constructor for the class with the same name
}
