% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4ClassDefinitions.R
\docType{class}
\name{UnBoundNonLinDecompOp_by_PoolNames-class}
\alias{UnBoundNonLinDecompOp_by_PoolNames-class}
\title{An S4 class to represent the of  nonlinear nonautonomuous compartmental system independently of the order of state variables}
\description{
An S4 class to represent the of  nonlinear nonautonomuous compartmental system independently of the order of state variables
}
