% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{UnBoundNonLinDecompOp_by_PoolNames}
\alias{UnBoundNonLinDecompOp_by_PoolNames}
\title{Generic constructor for the class with the same name}
\usage{
UnBoundNonLinDecompOp_by_PoolNames(internal_fluxes, out_fluxes, timeSymbol)
}
\arguments{
\item{internal_fluxes}{see method arguments}

\item{out_fluxes}{see method arguments}

\item{timeSymbol}{see method arguments}
}
\description{
Generic constructor for the class with the same name
}
\section{S4-methods}{
      \itemize{
        \item \code{\link{UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method}}
      }
}
